### 2025-11-20 #9483 @crvernon
@alegresor - Thank you for your submission to JOSS.  We are managing a large backlog of submissions currently but I'll get you a topic editor assigned as soon as one becomes available.
