### 2025-11-22 efba502
Address code review feedback: preserve retry logs and extract config values
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
