### 2025-11-22 e656e05
Final improvements: better comments and config safety checks
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
