### 2025-11-22 d18c60b
Add clarifying comments to retry logic
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
