### 2025-11-23 9777935
Uncomment torch test installation in workflow
