### 2025-11-23 97e1357
Update torch dependency to allow versions >= 2.0.0
Trying to fix a test failure.
