+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -3,15 +3,15 @@
 \label{tab2}
 \begin{tabular}{ll@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r}
 \toprule
-Method & Sampler & Mean & Std Dev & Mean  & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
+Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
  &  &  &   &  Error &  Error &   & &  \\
 \midrule
-QuantLib & IIDStdUniform & 105.3313 & 21.4671 & 0.204195 & 0.2296 & 0.919130 & 0.000000 & - & 0.892649 & 0.000000 & - \\
-QMCPy & IIDStdUniform & 105.1443 & 21.3848 & 0.017226 & 0.1474 & 0.624667 & 0.000000 & 1.471393 & 0.631089 & 0.000000 & 1.414459 \\
-QuantLib & Sobol & 105.0929 & 21.0518 & 0.034202 & 0.1857 & 1.000159 & 0.000000 & - & 0.976046 & 0.000000 & - \\
-QMCPy & Sobol & 105.1274 & 21.2337 & 0.000266 & 0.0037 & 0.626359 & 0.000000 & 1.467418 & 0.629713 & 0.000000 & 1.417549 \\
-QMCPy & Lattice & 105.1284 & 21.2460 & 0.001278 & 0.0085 & 0.760563 & 0.000000 & 1.208487 & 0.750671 & 0.000000 & 1.189135 \\
-QMCPy & Halton & 105.1271 & 21.2337 & 0.000033 & 0.0037 & 2.871192 & 0.000000 & 0.320121 & 3.454727 & 0.000000 & 0.258385 \\
+QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.90853098 & 0.00000000 & - \\
+QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.63580689 & 0.00000000 & 1.42894171 \\
+QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.99214080 & 0.00000000 & - \\
+QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.67312322 & 0.00000000 & 1.34972461 \\
+QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.76385953 & 0.00000000 & 1.18939535 \\
+QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 2.98949570 & 0.00000000 & 0.30390777 \\
 \bottomrule
 \end{tabular}
 \end{table}
