### 2025-11-27 3e6a8da
Reduce printout
