### 2025-11-28 dacab22
Update booktests workflow
