### 2025-11-28 f87c3a8
Change test command from booktests_parallel_no_docker to booktests_no_docker
