+++ test/booktests/tb_umbridge.py
@@ -7,8 +7,7 @@ class NotebookTests(BaseNotebookTest):
     def setUp(self):
         super().setUp()
 
-    @unittest.skip("Requires umbridge and docker")
-    @testbook('../../demos/umbridge.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/umbridge.ipynb', execute=False, timeout=TB_TIMEOUT)
     def test_umbridge_notebook(self, tb):
         pass
 
