+++ test/booktests/tb_dakota_genz.py
@@ -28,7 +28,7 @@ class NotebookTests(BaseNotebookTest):
         else:
             self._created_dummy_file = None
 
-    @testbook('../../demos/DAKOTA_Genz/dakota_genz.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/DAKOTA_Genz/dakota_genz.ipynb', execute=False, timeout=TB_TIMEOUT)
     def test_dakota_genz_notebook(self, tb):
         pass
 
