### 2025-11-28 865b61d
Update booktests workflow to install LaTeX
Added installation of LaTeX dependencies for testing.
