### 2025-11-28 5b06f43
Optimize workflow by cleaning up space and updating LaTeX install
Added a step to free up space before installing dependencies and modified LaTeX installation to use specific packages.
