+++ .github/workflows/booktests.yml
@@ -24,12 +24,19 @@ jobs:
         with:
           miniconda-version: "latest"
           auto-activate-base: true
+      - name: Free up space before install
+        run: |
+          sudo apt-get clean
+          sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
+          df -h
+          conda clean --all --yes
+          pip cache purge || true
       - name: Install base dependencies
         run: pip install -e .[test]
       - name: Install LaTeX
         run: |    
           sudo apt update
-          sudo apt install texlive-full latexmk
+          sudo apt install texlive texlive-science texlive-latex-extra latexmk
       - name: Check for missing booktest files
         run: make check_booktests
       - name: Generate missing booktest files
