### 2025-11-28 3b03e30
Refactor LaTeX installation in booktests workflow
Combined installation commands for LaTeX into a single line and added installation of recommended fonts.
