### 2025-11-28 3b7c205
Refactor LaTeX installation steps in workflow
Split LaTeX installation into separate steps for clarity.
