### 2025-11-28 f8334fb
Consolidate LaTeX installation commands
