+++ .github/workflows/booktests.yml
@@ -36,9 +36,7 @@ jobs:
       - name: Install LaTeX
         run: |    
           sudo apt update
-          sudo apt install texlive texlive-science texlive-latex-extra latexmk
-      - name: Install LaTeX dependencies
-          run: sudo apt-get update && sudo apt-get install -y texlive-fonts-extra
+          sudo apt install texlive texlive-science texlive-latex-extra texlive-fonts-extra latexmk
       - name: Check for missing booktest files
         run: make check_booktests
       - name: Generate missing booktest files
