### 2025-11-28 39492ed
Update OS matrix and optimize test command
