+++ .github/workflows/booktests.yml
@@ -49,4 +49,4 @@ jobs:
           pip install -e .[test_umbridge]
       - name: Run notebook tests
         shell: bash -l {0}
-        run: make -j1 booktests_no_docker
+        run: make -j2 booktests_no_docker TESTS="tb_quickstart"
