### 2025-11-28 bbb1189
Reduce memory usage
