+++ .github/workflows/booktests.yml
@@ -1,4 +1,3 @@
-# cSpell:words Booktests ipynb booktests booktest gpytorch botorch umbridge
 name: Booktests
 
 on:
@@ -13,19 +12,24 @@ on:
 
 jobs:
   booktests:
-    name: Notebook Tests on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: ["ubuntu-latest"]
+    runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v4
 
       # -----------------------------------------------------------
-      # Add swap space (prevents out-of-memory errors)
+      # 1. Install Miniconda
       # -----------------------------------------------------------
-      - name: Add 6GB swap
+      - uses: actions/checkout@v4
+      - uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniconda-version: "latest"
+          auto-activate-base: true
+
+      # -----------------------------------------------------------
+      # 2. Add swap space
+      # -----------------------------------------------------------
+      - name: Add 6GB swap (safe, after conda)
         run: |
           sudo fallocate -l 6G /swapfile
           sudo chmod 600 /swapfile
@@ -34,34 +38,29 @@ jobs:
           free -h
 
       # -----------------------------------------------------------
-      # Install Miniconda with Mamba (lower memory footprint)
+      # 3. Free disk + conda cache early
       # -----------------------------------------------------------
-      - uses: conda-incubator/setup-miniconda@v3
-        with:
-          miniconda-version: "latest"
-          auto-activate-base: true
-          mamba-version: "latest"
-          use-mamba: true
-
-      - name: Free space before environment setup
+      - name: Free space
         run: |
           sudo apt-get clean
           sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
-          df -h
           conda clean --all --yes
           pip cache purge || true
-
-      - name: Install base dependencies
-        run: mamba install pip && pip install -e .[test]
+          df -h
 
       # -----------------------------------------------------------
-      # Install minimal TeX (avoid full texlive: too big)
+      # 4. Install dependencies
+      # -----------------------------------------------------------
+      - name: Install base dependencies
+        run: pip install -e .[test]
+
+      # -----------------------------------------------------------
+      # 5. Install minimal LaTeX (not texlive-full)
       # -----------------------------------------------------------
       - name: Install minimal LaTeX
         run: |
           sudo apt update
-          sudo apt install -y texlive-latex-base texlive-latex-recommended texlive-latex-extra texlive-fonts-extra latexmk
-          df -h
+          sudo apt install -y texlive-latex-base texlive-latex-recommended latexmk
 
       - name: Check for missing booktest files
         run: make check_booktests
@@ -70,8 +69,8 @@ jobs:
         run: make generate_booktests
 
       # -----------------------------------------------------------
-      # Install additional dependencies (Torch, GP, etc.)
-      # -----------------------------------------------------------
+      # 6. Install additional dependencies (Torch, GP, etc.)
+      # -----------------------------------------------------------     
       - name: Install additional test dependencies
         run: |
           pip install -e .[test_torch]
@@ -81,7 +80,7 @@ jobs:
           conda clean --all --yes
 
       # -----------------------------------------------------------
-      # Run notebook tests sequentially (-j1)
+      # 7. Run notebook tests sequentially (-j1)
       # -----------------------------------------------------------
       - name: Run notebook tests
         shell: bash -l {0}
