+++ .github/workflows/booktests.yml
@@ -86,4 +86,4 @@ jobs:
         shell: bash -l {0}
         env:
           MPLBACKEND: Agg   # ensures headless, low-memory matplotlib
-        run: make -j1 booktests_no_docker TESTS="tb_quickstart"
+        run: make -j1 booktests_no_docker  # TESTS="tb_quickstart"
