### 2025-11-28 a686462
Add latex components
