+++ test/booktests/tb_Argonne_2023_Talk_Figures.py
@@ -6,8 +6,6 @@ import shutil
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
-# Skipping this test class
-@unittest.skip("Skipping NotebookTests class")
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
