+++ .github/workflows/booktests.yml
@@ -60,7 +60,7 @@ jobs:
       - name: Install minimal LaTeX
         run: |
           sudo apt update
-          sudo apt install -y texlive-latex-base texlive-latex-recommended latexmk
+          sudo apt install -y texlive-latex-base texlive-latex-extra  texlive-latex-recommended latexmk dvipng cm-super ghostscript
 
       - name: Check for missing booktest files
         run: make check_booktests
