+++ test/booktests/tb_dakota_genz.py
@@ -5,6 +5,7 @@ import numpy as np
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@unittest.skip("Skipping NotebookTests class")
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
