+++ test/booktests/tb_prob_failure_gp_ci.py
@@ -4,6 +4,7 @@ import shutil
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@unittest.skip("Skipping NotebookTests class")
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
@@ -11,7 +12,7 @@ class NotebookTests(BaseNotebookTest):
         # Create outputs directory
         os.makedirs('outputs', exist_ok=True)
 
-    @testbook('../../demos/talk_paper_demos/ProbFailureSorokinRao/prob_failure_gp_ci.ipynb', execute=False, timeout=TB_TIMEOUT)
+    @testbook('../../demos/talk_paper_demos/ProbFailureSorokinRao/prob_failure_gp_ci.ipynb', execute=True, timeout=TB_TIMEOUT)
     def test_prob_failure_gp_ci_notebook(self, tb):
         # Execute cells manually, skipping problematic ones
         for i, cell in enumerate(self.cells):
