+++ test/booktests/tb_MCQMC2022_Article_Figures.py
@@ -5,6 +5,7 @@ import shutil
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@unittest.skip("Skipping NotebookTests class")
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
@@ -21,7 +22,7 @@ class NotebookTests(BaseNotebookTest):
                 shutil.copy2(src, pkl_file)
                 print(f"\n\tCopied {src} to {pkl_file}")
 
-    @testbook('../../demos/talk_paper_demos/MCQMC2022_Article_Figures/MCQMC2022_Article_Figures.ipynb', execute=False, timeout=TB_TIMEOUT)
+    @testbook('../../demos/talk_paper_demos/MCQMC2022_Article_Figures/MCQMC2022_Article_Figures.ipynb', execute=True, timeout=TB_TIMEOUT)
     def test_mcqmc2022_article_figures_notebook(self, tb):
         # Execute cells up to but not including the stop_notebook cell
         for i in range(len(self.cells)):  
