### 2025-11-28 15b817a
Update booktests.yml
