### 2025-11-28 9b95d00
Add scikit-optimize installation to booktests workflow
