### 2025-11-28 833f946
Fix parsl_fest_2025.ipynb test failures and verbose output (#421)
* Initial plan

* Fix notebook test issues: improve retry logic and suppress verbose output

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

* Address code review feedback: preserve retry logs and extract config values

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

* Final improvements: better comments and config safety checks

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

* Add clarifying comments to retry logic

Co-auth
[…]
