+++ .github/workflows/booktests.yml
@@ -90,7 +90,7 @@ jobs:
         shell: bash -l {0}
         env:
           MPLBACKEND: Agg   # ensures headless, low-memory matplotlib
-        run: make -j1 booktests_no_docker  # TESTS="tb_quickstart"
+        run: make -j2 booktests_no_docker  # TESTS="tb_quickstart"
 
 
 # Filter for GitHub Actions runs: 
