+++ demos/talk_paper_demos/parsel_fest_2025/output/sequential_time.txt
@@ -1,0 +1,1 @@
+22.43929433822632
