+++ demos/parsel_fest_2025/output/sequential_time.txt
@@ -1,0 +1,1 @@
+13.603798151016235
