### 2025-12-01 5bfa102
Split into 3 notebooks
