+++ demos/talk_paper_demos/parsel_fest_2025/output/parallel_time.txt
@@ -1,1 +1,0 @@
-16.70528507232666
