+++ test/booktests/tb_prob_failure_gp_ci.py
@@ -12,7 +12,7 @@ class NotebookTests(BaseNotebookTest):
         # Create outputs directory
         os.makedirs('outputs', exist_ok=True)
 
-    @testbook('../../demos/talk_paper_demos/ProbFailureSorokinRao/prob_failure_gp_ci.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/talk_paper_demos/ProbFailureSorokinRao/prob_failure_gp_ci.ipynb', execute=False, timeout=TB_TIMEOUT)
     def test_prob_failure_gp_ci_notebook(self, tb):
         # Execute cells manually, skipping problematic ones
         for i, cell in enumerate(self.cells):
@@ -21,12 +21,12 @@ class NotebookTests(BaseNotebookTest):
                 
                 # Skip cells that would cause issues in test environment
                 if ('qp.util.stop_notebook()' in source or 
-                    'import umbridge' ins source or
+                    'import umbridge' in source or
                     'docker run' in source):
                     print(f"Skipping cell {i}: {source[:50]}...")
                     print(f"Stopping execution at cell {i} (docker dependency)")
                     break
-                    s
+
                 try:
                     self.execute_cell(i)
                 except Exception as e:
