+++ test/booktests/tb_Purdue_Talk_Figures.py
@@ -24,7 +24,7 @@ class NotebookTests(BaseNotebookTest):
                 shutil.copy2(src, pkl_file)
                 print(f"\n\tCopied {src} to {pkl_file}")
 
-    @testbook('../../demos/talk_paper_demos/Purdue_Talk_2023_March/Purdue_Talk_Figures.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/talk_paper_demos/Purdue_Talk_2023_March/Purdue_Talk_Figures.ipynb', execute=False, timeout=TB_TIMEOUT)
     def test_purdue_talk_figures_notebook(self, tb):
         # Execute cells up to but not including the stop_notebook cell
         for i in range(len(self.cells)):  
