+++ test/booktests/tb_Argonne_2023_Talk_Figures.py
@@ -28,7 +28,7 @@ class NotebookTests(BaseNotebookTest):
                 shutil.copy2(src, dst)
                 print(f"\n\tCopied {src} to {dst}")
 
-    @testbook('../../demos/talk_paper_demos/Argonne_Talk_2023_May/Argonne_2023_Talk_Figures.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/talk_paper_demos/Argonne_Talk_2023_May/Argonne_2023_Talk_Figures.ipynb', execute=False, timeout=TB_TIMEOUT)
     def test_argonne_talk_2023_figures_notebook(self, tb):
         # Execute cells up to but not including the stop_notebook cell
         for i in range(len(self.cells)):
