### 2025-12-01 3953321
Use LJF strategy (no better speedup)
