+++ demos/talk_paper_demos/parsel_fest_2025/README.md
@@ -32,4 +32,45 @@ Run `02_parallel.ipynb` to execute tests in parallel with Parsl.
 Run `03_visualize_speedup.ipynb` to:
 
 
+## Load Balancing with LPT Scheduling
 
+The parallel test runner uses the **Longest Processing Time (LPT)** algorithm (also known as Longest Job First) to optimize load balancing across Parsl workers.
+
+### How It Works
+
+1. **Runtime Estimates**: `test_runtimes.py` stores measured execution times for each notebook test from sequential runs.
+
+2. **Optimal Bin Packing**: Before submitting jobs, tests are sorted by estimated runtime (longest first) and assigned to the least-loaded worker using a min-heap.
+
+3. **Balanced Workloads**: This ensures all workers finish at approximately the same time, maximizing parallelism.
+
+### Scalability Analysis
+
+Based on measured runtimes (total sequential time: ~759s):
+
+| Workers | Estimated Makespan | Speedup | Efficiency |
+|---------|-------------------|---------|------------|
+| 1 | 759s | 1.0x | 100% |
+| 2 | 379s | 2.0x | 100% |
+| 4 | 190s | 4.0x | 100% |
+| 6 | 138s | 5.5x | 92% |
+| 8 | 138s | 5.5x | 69% |
+
+**Maximum theoretical speedup: 5.5x** (limited by the longest single test: `tb_iris` at 138s).
+
+### Assumptions
+
+1. **Reproducible runtimes**: Test execution times are consistent across runs.
+2. **Independent tests**: Each notebook test can run independently without shared state.
+3. **Single-threaded tests**: Individual tests do not benefit from multiple cores internally.
+4. **No I/O contention**: Parallel tests do not compete for disk or network resources.
+5. **Accurate estimates**: Runtime estimates in `test_runtimes.py` reflect actual execution times.
+
+### Updating Runtime Estimates
+
+After running sequential tests, update `test_runtimes.py` with new timing data:
+
+```bash
+make booktests_no_docker  # Run sequential tests, note the timing output
+# Then update TEST_RUNTIMES dict in test_runtimes.py
+```
