+++ demos/parsel_fest_2025/output/sequential_time.txt
@@ -1,1 +1,0 @@
-13.603798151016235
