+++ demos/parsel_fest_2025/output/parallel_time.txt
@@ -1,1 +1,0 @@
-17.16316270828247
