### 2025-12-01 167602b
Clean up booktests.yml a bit
