+++ pyproject.toml
@@ -75,6 +75,7 @@ test = [
     "parsl == 2025.7.28",
     "seaborn == 0.8",
     "scikit-learn >= 1.0.0",
+    "scikit-optimize",
     "ipykernel",
     "matplotlib >= 3.9.0",
     ]
