+++ .github/workflows/booktests.yml
@@ -88,7 +88,6 @@ jobs:
           pip install -e .[test_gpytorch]
           pip install -e .[test_botorch]
           pip install -e .[test_umbridge]
-          pip install scikit-optimize  # skopt
           conda clean --all --yes
 
       # -----------------------------------------------------------
