### 2025-12-02 2c23d28
updated gitignore, updated makefile, updated base notebook test class to display memory usage/ including jupyter notebook runs
