+++ makefile
@@ -41,56 +41,79 @@ unittests:
 	python -m pytest -x --cov qmcpy/ --cov-report term --cov-report json --no-header --cov-append test/
 
 check_booktests:
-	find demos -name '*.ipynb' | while read nb; do \
+	@echo "Checking notebook ↔ booktest coverage..."
+	@total_notebooks=0; \
+	skipped_notebooks=0; \
+	run_notebooks=0; \
+	missing_notebooks=0; \
+	for nb in $$(find demos -name '*.ipynb' | sort); do \
+		total_notebooks=$$((total_notebooks+1)); \
 		base=$$(basename "$$nb" .ipynb); \
 		test_base=$$(echo "$$base" | sed 's/[-.]/_/g'); \
-		if [ "$$base" != "parsl_fest_2025" ] && ! ls test/booktests/tb_"$$test_base".py > /dev/null 2>&1; then \
-			echo "    Missing test for: $$nb -> Expected: test/booktests/tb_$$test_base.py"; \
+		# ---------- explicit skip list ---------- \
+		if [ "$$base" = "parsl_fest_2025" ]; then \
+			echo "    Skipping $$nb (demo orchestrates booktests, not itself a booktest)"; \
+			skipped_notebooks=$$((skipped_notebooks+1)); \
+			continue; \
 		fi; \
-	done
-	@echo "Total notebooks:  $$(find demos -name '*.ipynb' | wc -l)"
-	@echo "Total test files: $$(find test/booktests -name 'tb_*.py' | wc -l)"
+		# ---------- check for matching Testbook file ---------- \
+		if ! ls test/booktests/tb_"$$test_base".py > /dev/null 2>&1; then \
+			echo "    Missing test for: $$nb -> Expected: test/booktests/tb_$$test_base.py"; \
+			missing_notebooks=$$((missing_notebooks+1)); \
+		else \
+			run_notebooks=$$((run_notebooks+1)); \
+		fi; \
+	done; \
+	echo "Total notebooks found:        $$total_notebooks"; \
+	echo "Total skipped explicitly (from .py creation):     $$skipped_notebooks"; \
+	echo "Total notebooks WITH tests:   $$run_notebooks"; \
+	echo "Total notebooks MISSING tests:$$missing_notebooks"; \
+	echo "Booktests will be run for $$run_notebooks notebooks."
 
 
 # This helps locate generated or local-only folders like build, .pytest_cache, etc.
 find_local_only_files:
-	chmod +x scripts/find_local_only_folders.sh
-	./scripts/find_local_only_folders.sh
+	@chmod +x scripts/find_local_only_folders.sh
+	@./scripts/find_local_only_folders.sh
 
 clean_local_only_files:
-	chmod +x scripts/find_local_only_folders.sh
-	for f in $(shell ./scripts/find_local_only_folders.sh); do \
-		rm -f "$$f"; \
+	@chmod +x scripts/find_local_only_folders.sh
+	@./scripts/find_local_only_folders.sh | while read f; do \
+		if [ -n "$$f" ]; then \
+			rm -f "$$f"; \
+		fi; \
 	done
 
 generate_booktests:
-	@echo "\nGenerating missing booktest files..."
-	cd test/booktests/ && python generate_test.py --check-missing
+	@echo ""
+	@echo "Generating missing booktest files..."
+	@cd test/booktests/ && python generate_test.py --check-missing
 
 
 booktests_no_docker: check_booktests generate_booktests clean_local_only_files
[…]
