### 2025-12-02 52c8f9c
updated contributing.md to include system dependencies for booktests
