+++ test/booktests/parsl_test_runner.py
@@ -1,49 +1,241 @@
-import parsl
-from parsl import bash_app
 import glob
 import os
 import re
+import subprocess
+import signal
 import time
 from pathlib import Path
-import time
+
+import parsl
 import parsl as pl
+from parsl import bash_app
 from parsl.configs.htex_local import config
 
-# Import runtime estimates for load balancing
+# Import runtime estimates for load balancing (LPT scheduling)
 try:
     from test_runtimes import sort_by_runtime, get_runtime, print_schedule, optimal_schedule
     HAS_RUNTIME_ESTIMATES = True
 except ImportError:
     HAS_RUNTIME_ESTIMATES = False
 
+
+# ---------------------------------------------------------------------------
+# EXPLICIT MODULE SKIP LIST
+#
+# These are tb_*.py modules that we DO NOT want Parsl to execute in parallel.
+# Typically they correspond to notebooks that are:
+#   - heavy LaTeX / many figures,
+#   - orchestration-only demos,
+#   - or otherwise excluded from CI booktests.
+#
+# To skip a new tb_*.py module from Parsl runs, ADD ITS MODULE NAME HERE.
+# Example: to skip test/booktests/tb_new_heavy_demo.py, add "tb_new_heavy_demo".
+# ---------------------------------------------------------------------------
+SKIP_MODULES = {
+    "tb_dakota_genz",
+    "tb_Argonne_2023_Talk_Figures",
+    "tb_MCQMC2022_Article_Figures",
+    "tb_prob_failure_gp_ci",
+    "tb_Purdue_Talk_Figures",
+    "tb_pydata_chi_2023",
+}
+# ---------------------------------------------------------------------------
+
+
 @bash_app
-def run_single_test(test_file, stdout='test_output.txt', stderr='test_error.txt'):
-    """Run a single test file using bash"""
+def run_single_test(test_module, stdout='test_output.txt', stderr='test_error.txt'):
+    """Run a single tb_*.py unittest module via bash."""
+    # We run `python -m unittest <module>` where <module> is e.g. tb_qmcpy_intro
+    # NOTE: no .py extension here.
     return f"""
-    PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" python -m unittest {test_file}
+    PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" \
+    python -m unittest {test_module}
     """
 
+
+def _kill_interchange_processes(retries=3, delay=0.5):
+    """Best-effort cleanup of any lingering Parsl 'interchange.py' processes."""
+    for attempt in range(retries):
+        # Try pkill first
+        try:
+            subprocess.run(['pkill', '-f', 'interchange.py'], check=False)
+        except Exception:
+            pass
+
+        # Try pgrep -> kill
+        try:
+            p = subprocess.run(
+                ['pgrep', '-f', 'interchange.py'],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.DEVNULL,
+                text=True,
+            )
+            if p.stdout:
+                for line in p.stdout.splitlines():
+                    try:
+                        pid = int(line.strip())
+                        try:
+                            os.kill(pid, signal.SIGTERM)
+                        except ProcessLookupError:
+                            pass
+                    except Exception:
+                        pass
+        except Exception:
+            pass
+
+        time.sleep(delay)
+
+        # Force kill remaining
+        try:
+            p = subprocess.run(
+                ['pgrep', '-f', 'interchange.py'],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.DEVNULL,
+                text=True,
+            )
+            if p.stdout:
+                for line in p.stdout.splitlines():
[…]
