### 2025-12-03 f9298cb
troubleshooting parallelization speedup
