+++ test/booktests/parsl_test_runner.py
@@ -1,8 +1,6 @@
 import glob
 import os
 import re
-import subprocess
-import signal
 import time
 from pathlib import Path
 
@@ -13,214 +11,127 @@ from parsl.configs.htex_local import config
 
 # Import runtime estimates for load balancing (LPT scheduling)
 try:
-    from test_runtimes import sort_by_runtime, get_runtime, print_schedule, optimal_schedule
+    from test_runtimes import optimal_schedule, get_runtime
     HAS_RUNTIME_ESTIMATES = True
 except ImportError:
     HAS_RUNTIME_ESTIMATES = False
 
-
 # ---------------------------------------------------------------------------
-# EXPLICIT MODULE SKIP LIST
+# SINGLE SOURCE OF TRUTH FOR SKIPS
 #
-# These are tb_*.py modules that we DO NOT want Parsl to execute in parallel.
-# Typically they correspond to notebooks that are:
-#   - heavy LaTeX / many figures,
-#   - orchestration-only demos,
-#   - or otherwise excluded from CI booktests.
-#
-# To skip a new tb_*.py module from Parsl runs, ADD ITS MODULE NAME HERE.
-# Example: to skip test/booktests/tb_new_heavy_demo.py, add "tb_new_heavy_demo".
+# These are the NOTEBOOKS we explicitly skip (with reasons).
+# tb_*.py modules are derived from these paths automatically, so there is
+# NO separate SKIP_MODULES list anymore.
 # ---------------------------------------------------------------------------
-SKIP_MODULES = {
-    "tb_dakota_genz",
-    "tb_Argonne_2023_Talk_Figures",
-    "tb_MCQMC2022_Article_Figures",
-    "tb_prob_failure_gp_ci",
-    "tb_Purdue_Talk_Figures",
-    "tb_pydata_chi_2023",
+SKIPPED_NOTEBOOKS = {
+    "demos/DAKOTA_Genz/dakota_genz.ipynb":
+        "requires large manual file / heavy memory use",
+    "demos/talk_paper_demos/Argonne_Talk_2023_May/Argonne_2023_Talk_Figures.ipynb":
+        "heavy LaTeX + many figures; skipped in booktests_no_docker",
+    "demos/talk_paper_demos/MCQMC2022_Article_Figures/MCQMC2022_Article_Figures.ipynb":
+        "MCQMC 2022 article figures; not run in CI",
+    "demos/talk_paper_demos/ProbFailureSorokinRao/prob_failure_gp_ci.ipynb":
+        "prob_failure_gp_ci talk demo; heavy GP / prob. failure example",
+    "demos/talk_paper_demos/Purdue_Talk_2023_March/Purdue_Talk_Figures.ipynb":
+        "Purdue 2023 talk figures; not a CI booktest target",
+    "demos/talk_paper_demos/parsel_fest_2025/01_sequential.ipynb":
+        "helper notebook for parsl_fest_2025; not a standalone booktest",
+    "demos/talk_paper_demos/parsel_fest_2025/02_parallel.ipynb":
+        "helper notebook for parsl_fest_2025; not a standalone booktest",
+    "demos/talk_paper_demos/parsel_fest_2025/03_visualize_speedup.ipynb":
+        "helper notebook for parsl_fest_2025; not a standalone booktest",
+    "demos/talk_paper_demos/pydata_chi_2023.ipynb":
+        "PyData Chicago 2023 talk notebook; demo-only",
 }
 # ---------------------------------------------------------------------------
 
 
+def _derive_skipped_modules_from_notebooks():
+    """
+    Derive tb_*.py module names from SKIPPED_NOTEBOOKS keys.
+
[…]
