### 2025-12-03 583e176
improved MPMC module
