+++ qmcpy/discrete_distribution/__init__.py
@@ -13,8 +13,7 @@ Net = DigitalNetB2
 # Import MPMC if PyTorch dependencies are available
 try:
     from .mpmc import MPMC
-except ImportError:
-    # PyTorch dependencies not available
-    pass
+except Exception:  # pragma: no cover - optional dependency
+    MPMC = None
 
 
