+++ demos/talk_paper_demos/parsl_fest_2025/output/parallel_output_4.txt
@@ -1,0 +1,125 @@
+rm -fr demos/.ipynb_checkpoints/*checkpoint.ipynb && \
+	find demos -name '*.ipynb' | while read nb; do \
+		base=$(basename "$nb" .ipynb); \
+		test_base=$(echo "$base" | sed 's/[-.]/_/g'); \
+		if echo "$nb" | grep -q "parsl_fest_2025"; then \
+			continue; \
+		fi; \
+		if ! ls test/booktests/tb_"$test_base".py > /dev/null 2>&1; then \
+			echo "    Missing test for: $nb -> Expected: test/booktests/tb_$test_base.py"; \
+		fi; \
+	done
+Total notebooks:        39
+Total test files:       32
+
+Generating missing booktest files...
+cd test/booktests/ && python generate_test.py --check-missing
+No missing test files found.
+rm -fr test/booktests/.ipynb_checkpoints/
+chmod +x scripts/find_local_only_folders.sh > /dev/null 2>&1
+for f in ; do \
+		rm -f "$f"; > /dev/null 2>&1; \
+	done
+
+Notebook tests with Parsl
+pip install -q -e ".[test]"  && \
+	cd test/booktests/ && \
+	rm -fr *.eps *.jpg *.pdf *.png *.part *.txt *.log && rm -fr logs && rm -fr runinfo prob_failure_gp_ci_plots && \
+	PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" \
+	python parsl_test_runner.py  -v --failfast && \
+	cd ../.. 
+Parsl configuration loaded successfully.
+Applied optimal LPT scheduling for 4 workers
+Estimated makespan: 190.1s (speedup: 3.99x)
+Found 32 test modules to execute in parallel...
+All tests submitted to Parsl executor...
+[1/32] tb_iris: PASSED
+tb_iris ...     Memory used: 0.10 GB.  Test time: 64.69 s
+ok
+[2/32] tb_elliptic_pde: PASSED
+tb_elliptic_pde ...     Memory used: 0.10 GB.  Test time: 26.29 s
+ok
+[3/32] tb_qei_demo_for_blog: PASSED
+tb_qei_demo_for_blog ...     Memory used: 0.10 GB.  Test time: 40.98 s
+ok
+[4/32] tb_asian_option_mlqmc: PASSED
+tb_asian_option_mlqmc ...     Memory used: 0.10 GB.  Test time: 45.13 s
+ok
+[5/32] tb_linear_scrambled_halton: PASSED
+tb_linear_scrambled_halton ...     Memory used: 0.11 GB.  Test time: 7.15 s
+ok
+[6/32] tb_ray_tracing: PASSED
+tb_ray_tracing ...     Memory used: 0.10 GB.  Test time: 16.41 s
+ok
+[7/32] tb_digital_net_b2: PASSED
+tb_digital_net_b2 ...     Memory used: 0.10 GB.  Test time: 17.17 s
+ok
+[8/32] tb_vectorized_qmc_bayes: PASSED
+tb_vectorized_qmc_bayes ...     Memory used: 0.09 GB.  Test time: 29.33 s
+ok
+[9/32] tb_plot_proj_function: PASSED
+tb_plot_proj_function ...     Memory used: 0.11 GB.  Test time: 6.03 s
+ok
+[10/32] tb_lattice_random_generator: PASSED
+tb_lattice_random_generator ...     Memory used: 0.10 GB.  Test time: 7.79 s
+ok
[…]
