### 2025-12-04 be0c48f
add class switch for qmcpy installation
