### 2025-12-04 f608e9f
adjust torch dependency so tests pass
