### 2025-12-04 0f65622
changed torch to 2.9.0 required
