+++ pyproject.toml
@@ -75,7 +75,7 @@ test = [
     "phmutest == 1.0.1",
     ]
 test_torch = [
-    "torch == 2.7.0",
+    "torch == 2.9.0",
     ]
 test_gpytorch = [
     "gpytorch == 1.13",
