### 2025-12-04 e995cd8
Fix MPMC doctest
