### 2025-12-04 9b4a2bd
pin class package versions
