+++ pyproject.toml
@@ -116,11 +116,11 @@ class = [
     "pandas >= 2.0.0",
     "scikit-learn >= 1.0.0",
     "scikit-optimize >= 0.10.2",
-    "ipynbname",
-    "jupyterlab_execute_time",
-    "ipywidgets",
-    "statsmodels",
-    "tqdm",
+    "ipynbname >= 2023.2.0.0",
+    "jupyterlab_execute_time >= 3.1.2",
+    "ipywidgets >= 8.1.0",
+    "statsmodels >= 0.14.3",
+    "tqdm >= 4.66.0",
     "sympy >= 1.12",
     "arviz >= 0.17",
     "plotly >= 5.0",
