### 2025-12-04 afbb8b1
Simplify outputs and code
