+++ test/booktests/parsl_test_runner.py
@@ -18,12 +18,7 @@ if platform.system() == 'Darwin':
 else:
     from parsl.configs.htex_local import config
 
-# Import runtime estimates for load balancing
-try:
-    from test_runtimes import sort_by_runtime, get_runtime, print_schedule, optimal_schedule
-    HAS_RUNTIME_ESTIMATES = True
-except ImportError:
-    HAS_RUNTIME_ESTIMATES = False
+
 
 @bash_app
 def run_single_test(test_file, stdout='test_output.txt', stderr='test_error.txt'):
@@ -63,28 +58,6 @@ def execute_parallel_tests():
         print(f"Note: Removed {len(test_modules) - len(unique_test_modules)} duplicate test module(s)")
     test_modules = unique_test_modules
 
-    # Apply optimal scheduling for better load balancing
-    if HAS_RUNTIME_ESTIMATES:
-        # Try to get the number of workers from environment or Parsl config
-        try:
-            num_workers = int(os.environ.get('PARSL_MAX_WORKERS'))
-        except Exception:
-            try:
-                exec_obj = config.executors[0] if config.executors else None
-                if exec_obj:
-                    # ThreadPoolExecutor uses max_threads, HTEX uses max_workers
-                    num_workers = getattr(exec_obj, 'max_threads', None) or getattr(exec_obj, 'max_workers', None) or 8
-                else:
-                    num_workers = 8
-            except:
-                num_workers = 8
-        
-        # Use optimal bin-packing schedule
-        test_modules, assignments, est_makespan = optimal_schedule(test_modules, num_workers)
-        total_seq = sum(get_runtime(m) for m in test_modules)
-        print(f"Applied optimal LPT scheduling for {num_workers} workers")
-        print(f"Estimated makespan: {est_makespan:.1f}s (speedup: {total_seq/est_makespan:.2f}x)")
-    
     print(f"Found {len(test_modules)} test modules to execute in parallel...")
     
     # Submit all jobs
@@ -255,196 +228,6 @@ def generate_summary_report(results, execution_time=0.0, total_test_time=0.0):
                 print(f"Error: {status}")
                 print()
 
-def reload_parsl_config(max_workers=None, wait=1):
-    """Safely reload Parsl with an updated config.
-
-    - If a Parsl DFK exists, try to cleanup/shutdown it first.
-    - Call `pl.clear()` to reset Parsl state.
-    - Optionally set `config.max_workers` before loading.
-    - Load the (possibly modified) config and wait briefly.
-    """
-    import os
-    import signal
-    import subprocess
-    import time as _time
-
-    def kill_interchange_processes(retries=3, delay=0.5):
-        """Kill any running 'interchange.py' processes using multiple strategies."""
-        for attempt in range(retries):
-            killed_any = False
-            # Try pkill first (simpler)
-            try:
-                subprocess.run(['pkill', '-f', 'interchange.py'], check=False)
-            except Exception:
-                pass
-
-            # Try pgrep -> kill
-            try:
-                p = subprocess.run(['pgrep', '-f', 'interchange.py'], stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, text=True)
-                if p.stdout:
-                    for line in p.stdout.splitlines():
-                        try:
-                            pid = int(line.strip())
-                            try:
-                                os.kill(pid, signal.SIGTERM)
-                                killed_any = True
-                            except ProcessLookupError:
-                                pass
-                        except Exception:
-                            pass
-            except Exception:
-                pass
-
-            # Wait a bit for processes to die
-            _time.sleep(delay)
-
-            # Force kill remaining
-            try:
-                p = subprocess.run(['pgrep', '-f', 'interchange.py'], stdout=subprocess.PIPE, stderr=subprocess.DEVNULL, text=True)
-                if p.stdout:
[…]
