+++ test/booktests/test_runtimes.py
@@ -9,10 +9,13 @@ To update these values, run the sequential tests and extract timing info:
 Tests are sorted by runtime (longest first) for optimal load balancing.
 """
 import heapq
+import os
+import pandas as pd
 
 # Estimated runtimes in seconds (from sequential run on 2025-12-01)
 # Sorted from longest to shortest for easy reference
-TEST_RUNTIMES = {
+# These are fallback values if CSV file is not available
+TEST_RUNTIMES_FALLBACK = {
     # Long-running tests (> 60s) - these are the bottleneck
     'tb_iris': 138.10,
     'tb_elliptic_pde': 96.12,
@@ -48,37 +51,36 @@ TEST_RUNTIMES = {
     # Skipped tests (assigned small default time)
     'tb_Argonne_2023_Talk_Figures': 1.0,
     'tb_MCQMC2022_Article_Figures': 1.0,
-    'tb_Purdue_Talk_Figures': 1.0,
-    'tb_dakota_genz': 1.0,
-    'tb_prob_failure_gp_ci': 1.0,
     'tb_pydata_chi_2023': 1.0,
 }
 
 # Default runtime for unknown tests (assume medium-length)
 DEFAULT_RUNTIME = 30.0
 
+# Try to load runtimes from CSV file
+TEST_RUNTIMES = {}
+csv_path = os.path.join(os.path.dirname(__file__), '../../demos/talk_paper_demos/parsl_fest_2025/output/sequential_output_time.csv')
+if os.path.exists(csv_path):
+    try:
+        df = pd.read_csv(csv_path)
+        # Convert from test_name to tb_name format
+        for _, row in df.iterrows():
+            notebook = row['Notebook']
+            time_s = float(row['Time_s'])
+            # Convert test_xxx to tb_xxx format
+            tb_name = notebook.replace('test_', 'tb_')
+            TEST_RUNTIMES[tb_name] = time_s
+        print(f"Loaded {len(TEST_RUNTIMES)} test runtimes from {csv_path}")
+    except Exception as e:
+        print(f"Warning: Failed to load runtimes from CSV: {e}")
+        TEST_RUNTIMES = TEST_RUNTIMES_FALLBACK.copy()
+else:
+    # Fall back to hardcoded values
+    TEST_RUNTIMES = TEST_RUNTIMES_FALLBACK.copy()
 
 def get_runtime(test_name: str) -> float:
     """Get estimated runtime for a test module."""
     return TEST_RUNTIMES.get(test_name, DEFAULT_RUNTIME)
-
-
-def sort_by_runtime(test_modules: list, longest_first: bool = True) -> list:
-    """
-    Sort test modules by estimated runtime (simple LJF).
-    
-    Args:
-        test_modules: List of test module names
-        longest_first: If True, sort longest jobs first (LJF scheduling)
-        
-    Returns:
-        Sorted list of test module names
-    """
-    return sorted(
-        test_modules,
-        key=lambda m: get_runtime(m),
-        reverse=longest_first
-    )
 
 
 def optimal_schedule(test_modules: list, num_workers: int) -> tuple:
@@ -100,7 +102,7 @@ def optimal_schedule(test_modules: list, num_workers: int) -> tuple:
         raise ValueError("num_workers must be positive")
     
     # Sort by runtime (longest first) - LPT heuristic
-    sorted_modules = sort_by_runtime(test_modules, longest_first=True)
+    sorted_modules = sorted(test_modules, key=lambda m: get_runtime(m), reverse=True)
     
     # Min-heap of (current_load, worker_id, [assigned_tests])
     workers = [(0.0, i, []) for i in range(num_workers)]
@@ -132,90 +134,3 @@ def optimal_schedule(test_modules: list, num_workers: int) -> tuple:
[…]
