### 2025-12-04 26b7c99
Fix workflow syntax error
