+++ .github/workflows/booktests.yml
@@ -2,11 +2,11 @@ name: Booktests
 
 on:
   push:
-  branches:
-    - master
-    - develop
-    - booktests_choi
-    - brandon-fix-dakota-parallel
+    branches:
+      - master
+      - develop
+      - booktests_choi
+      - brandon-fix-dakota-parallel
     paths:
       - 'demos/**/*.ipynb'
       - '.github/workflows/booktests.yml'
