### 2025-12-04 a995761
Added to the README for the class installation option
