### 2025-12-04 af91b38
Add test OS platforms
