+++ .github/workflows/booktests.yml
@@ -22,7 +22,11 @@ on:
 
 jobs:
   booktests:
-    runs-on: ubuntu-latest
+    name: Booktests on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: ["macos-latest", "ubuntu-latest", "windows-latest"]
     env:
       ACTIONS_STEP_DEBUG: false   # true to enable debug logging
     steps:
@@ -38,9 +42,10 @@ jobs:
           auto-activate-base: true
 
       # -----------------------------------------------------------
-      # 2. Add swap space
+      # 2. Add swap space (Linux only)
       # -----------------------------------------------------------
       - name: Add 12GB swap
+        if: runner.os == 'Linux'
         run: |
           sudo fallocate -l 12G /swapfile
           sudo chmod 600 /swapfile
@@ -51,7 +56,8 @@ jobs:
       # -----------------------------------------------------------
       # 3. Free disk + conda cache early
       # -----------------------------------------------------------
-      - name: Free space
+      - name: Free space (Linux)
+        if: runner.os == 'Linux'
         run: |
           sudo apt-get clean
           sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
@@ -59,11 +65,26 @@ jobs:
           pip cache purge || true
           df -h
 
-      - name: Show runner resources
+      - name: Free space (macOS/Windows)
+        if: runner.os != 'Linux'
         run: |
-          echo "CPUs: $(nproc)"
-          free -h
+          conda clean --all --yes
+          pip cache purge || true
+
+      - name: Show runner resources (Linux/macOS)
+        if: runner.os != 'Windows'
+        run: |
+          echo "CPUs: $(nproc || sysctl -n hw.ncpu)"
+          free -h || vm_stat
           df -h
+
+      - name: Show runner resources (Windows)
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: |
+          Write-Host "CPUs: $env:NUMBER_OF_PROCESSORS"
+          Get-CimInstance Win32_OperatingSystem | Select-Object TotalVisibleMemorySize,FreePhysicalMemory
+          Get-PSDrive -PSProvider FileSystem
 
       # -----------------------------------------------------------
       # 4. Install dependencies
@@ -72,12 +93,29 @@ jobs:
         run: pip install -e .[test]
 
       # -----------------------------------------------------------
-      # 5. Install minimal LaTeX (not texlive-full)
+      # 5. Install minimal LaTeX (OS-specific)
       # -----------------------------------------------------------
-      - name: Install minimal LaTeX
+      - name: Install minimal LaTeX (Linux)
+        if: runner.os == 'Linux'
         run: |
           sudo apt update
-          sudo apt install -y texlive-latex-base texlive-latex-extra  texlive-latex-recommended latexmk dvipng cm-super ghostscript
+          sudo apt install -y texlive-latex-base texlive-latex-extra texlive-latex-recommended latexmk dvipng cm-super ghostscript
+
+      - name: Install minimal LaTeX (macOS)
+        if: runner.os == 'macOS'
+        run: |
+          brew install --cask basictex
+          eval "$(/usr/libexec/path_helper)"
+          sudo tlmgr update --self
+          sudo tlmgr install latexmk dvipng collection-fontsrecommended
+
+      - name: Install minimal LaTeX (Windows)
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: |
+          choco install miktex -y
+          $env:Path = [System.Environment]::GetEnvironmentVariable("Path","Machine") + ";" + [System.Environment]::GetEnvironmentVariable("Path","User")
+          miktex --admin packages install latexmk dvipng
 
       - name: Check for missing booktest files
         run: make check_booktests
