### 2025-12-04 a7a4a6c
Take out windows as Parsl does not support it yet
