### 2025-12-05 21da3b6
move md sec from README to CONTRIBUTING
