### 2025-12-05 e48c0de
clean up from some old .ipynb checkpoints
