### 2025-12-05 f47ff31
Booktests: Unit tests for Jupiter Notebooks (#424)
* Update booktests

* Add failing fast option for all tests

* Add notebook generating functions

* Fix unit tests failure

* Fix unit tests failures

* Install testbook and parsl

* Fix unit tests

* Fix booktests

* Fix booktests

* Delete tb_importance_sampling.py

* Delete tb_integration_examples.py

* Update tb_gbm_demo.py

* Fix booktests

* Fix path

* Create booktest workflow

* Update booktests

* Fix booktests

* Fix booktests

* Update .gitignore

* Update generate_test.py

* Test booktest workflow

* Fix unit tests for notebooks

*
[…]
