+++ test/booktests/tb_linear_scrambled_halton.py
@@ -1,0 +1,12 @@
+import unittest
+from testbook import testbook
+from __init__ import TB_TIMEOUT, BaseNotebookTest
+
+class NotebookTests(BaseNotebookTest):
+
+    @testbook('../../demos/linear-scrambled-halton.ipynb', execute=True, timeout=TB_TIMEOUT)
+    def test_linear_scrambled_halton_notebook(self, tb):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
