+++ pyproject.toml
@@ -54,7 +54,7 @@ dependencies = [
 
 [project.optional-dependencies]
 torch = [
-    "torch >= 2.0.0",
+    "torch >= 2.7.0",
     ]
 gpytorch = [
     "gpytorch >= 1.7.0",
@@ -70,6 +70,14 @@ test = [
     "pytest-cov == 6.1.1",
     "phmutest == 1.0.1",
     "pytest-accept == 0.1.10",
+    "testbook == 0.4.2",
+    "psutil == 5.9.5",
+    "parsl == 2025.7.28",
+    "seaborn == 0.8",
+    "scikit-learn >= 1.0.0",
+    "scikit-optimize",
+    "ipykernel",
+    "matplotlib >= 3.9.0",
     ]
 test_torch = [
     "torch >= 2.0.0",

[…]
