### 2025-12-05 2301bc0
add thesis to demos + docs patch
