### 2025-12-05 0f20e4c
Extend SciPyWrapper with dependent and custom distributions for issue #406
