+++ qmcpy/true_measure/scipy_wrapper.py
@@ -4,95 +4,513 @@ from ..discrete_distribution.abstract_discrete_distribution import AbstractDiscr
 from ..discrete_distribution import DigitalNetB2
 import numpy as np
 import scipy.stats
-from typing import Union
+import warnings
+
+
+class _MVNAdapter:
+    """
+    Small adapter that turns a SciPy multivariate normal like object into
+    something with a simple "transform(u)" interface.
+
+    Idea:
+      1. Start from u in (0,1)^d.
+      2. Map to standard normals z via norm.ppf.
+      3. Apply a Cholesky factor to inject the correlation structure.
+    """
+
+    def __init__(self, mvn_like):
+        # Keep the original object so we can still call logpdf later.
+        self._mvn = mvn_like
+
+        # SciPy's multivariate_normal has attributes "mean" and "cov".
+        mean = np.asarray(mvn_like.mean)
+        cov = np.asarray(mvn_like.cov)
+
+        if mean.ndim != 1:
+            raise ParameterError(
+                "SciPyWrapper currently expects a 1D mean vector for multivariate_normal."
+            )
+        if cov.ndim != 2 or cov.shape[0] != cov.shape[1]:
+            raise ParameterError(
+                "SciPyWrapper expects a square covariance matrix for multivariate_normal."
+            )
+        if cov.shape[0] != mean.size:
+            raise DimensionError(
+                f"Mean dimension {mean.size} and cov shape {cov.shape} do not match."
+            )
+
+        self.dim = mean.size
+        self._mean = mean
+        # Cholesky factor gives L such that cov = L L^T.
+        self._chol = np.linalg.cholesky(cov)
+
+    def transform(self, u):
+        """
+        Take u in (0,1)^d and turn it into correlated normal samples.
+        """
+        u = np.asarray(u, dtype=float)
+        if u.shape[-1] != self.dim:
+            raise DimensionError(
+                f"MVNAdapter expected last axis {self.dim}, got {u.shape[-1]}"
+            )
+
+        # Clip so we never hit exactly 0 or 1 inside norm.ppf.
+        eps = np.finfo(float).eps
+        u_clip = np.clip(u, eps, 1.0 - eps)
+
+        # Map to i.i.d. standard normals.
+        z = scipy.stats.norm.ppf(u_clip)
+
+        # Flatten, apply Cholesky, then reshape back.
+        z_flat = z.reshape(-1, self.dim)
+        x_flat = z_flat @ self._chol.T + self._mean
+        return x_flat.reshape(z.shape)
+
+    def logpdf(self, x):
+        """
+        Forward to the SciPy logpdf, keeping shapes tidy.
+        """
+        x = np.asarray(x, dtype=float)
+        if x.shape[-1] != self.dim:
+            raise DimensionError(
+                f"MVNAdapter expected last axis {self.dim}, got {x.shape[-1]}"
+            )
+
+        x_flat = x.reshape(-1, self.dim)
+        logp = self._mvn.logpdf(x_flat)
+        return np.asarray(logp).reshape(x.shape[:-1])
 
 
 class SciPyWrapper(AbstractTrueMeasure):
     r"""
-    Multivariate distribution with independent marginals from [`scipy.stats`](https://docs.scipy.org/doc/scipy/reference/stats.html#continuous-distributions).
+    True measure that wraps SciPy style distributions.
 
-    Examples:
-        >>> true_measure = SciPyWrapper(
-        ...     sampler = DigitalNetB2(3,seed=7),
-        ...     scipy_distribs = [
-        ...         scipy.stats.uniform(loc=1,scale=2),
-        ...         scipy.stats.norm(loc=3,scale=4),
[…]
