+++ demos/scipywrapper_dependence_custom/scipywrapper_dependence_custom_blog.md
@@ -1,0 +1,271 @@
+# Extending QMCPy's `SciPyWrapper` for Dependent and Custom Distributions
+
+When we first bump into `SciPyWrapper`, the examples live in very nice, controlled worlds. The true measures are clean, the distributions come straight from `scipy.stats`, and most variables behave independently of each other.
+
+My project started from the moments when that story breaks.
+* What if the variables you care about are strongly correlated?
+* What if your distribution is not in `scipy.stats` at all, but you still want to drive it with a low discrepancy point set?
+* And what if a user hands you a custom distribution that is quietly wrong?
+
+This blog walks through the changes I made to `qmcpy.true_measure.SciPyWrapper` to handle those cases, and it uses four visual examples plus one “failure” example to show that the new features really work.
+
+All of the code lives in:
+* `qmcpy/true_measure/scipy_wrapper.py`
+* `demos/scipy_wrapper_dependence_custom/scipywrapper_dependence_custom.ipynb`
+* `test/test_scipy_wrapper_custom.py`
+
+---
+
+## 1. What `SciPyWrapper` did before
+
+`SciPyWrapper` was originally designed for a very specific setting:
+
+> Take a discrete sampler that produces points in `[0, 1]^d` and transform each coordinate independently  
+> using a list of frozen `scipy.stats` univariate distributions.
+
+So if you gave it a 3 dimensional `DigitalNetB2` and three frozen continuous distributions, it would apply each inverse CDF (`ppf`) coordinatewise. This works nicely for independent marginals, but it has some limitations:
+1. It ignores any dependence between coordinates.  
+2. It assumes everything comes from `scipy.stats` continuous distributions.  
+3. There is no structured way to plug in a completely user defined distribution while still catching obvious mistakes.
+
+The goal of this project is to keep the original behaviour for simple cases, but extend it in two directions:
+- support joint distributions that encode dependence, including non trivial constructions like zero inflated models and acceptance rejection transforms;
+- support user defined marginals that follow a SciPy like interface, with sanity checks and clear warnings when something looks wrong.
+
+---
+
+## 2. New interface
+
+At a high level, the extended `SciPyWrapper` can now work in two modes.
+
+### 2.1 Independent marginal mode
+
+This is the original behaviour.
+* Input: list of frozen univariate distributions or a single frozen distribution from `scipy.stats`, or user defined univariate distributions that look similar.
+* Transform: apply each `ppf` coordinatewise.
+* Weight: multiply the marginal densities or log densities.
+
+This is still used in the examples for one dimensional custom marginals.
+
+
+### 2.2 Joint distribution mode
+
+This is the new piece.
+
+Instead of giving a list of marginals, you can now pass a *joint* object that knows how to turn a point `u` from `[0, 1]^d` into a sample `x` in the physical space.
+
+The joint object is expected to provide:
+- `dim`: the dimension `d`;
+- `transform(u)`: vectorised map from `u` in `[0, 1]^d` to `x` in `R^d`;
+- optional `logpdf(x)` or `pdf(x)`: to supply weights if needed.
+
+`SciPyWrapper` detects this type and routes all calls through the joint transform. This gives us the flexibility to encode dependence directly in `transform`.
+
+In the rest of the blog, all the “dependent” examples use this joint mode.
+
+---
+
+## 3. Example 1: independent vs dependent normals
+
+This is the example that checks that the joint transform behaves as expected for a familiar multivariate normal.
+
+We compare:
+1. `SciPyWrapper` with two independent standard normal marginals.
+2. `SciPyWrapper` with a 2-dimensional multivariate normal that has correlation `ρ = 0.8`, provided through a joint adapter.
+
[…]
