### 2025-12-05 88bf24a
In demos, move parsl_fest_2025/* to ParslFest2025/
