### 2025-12-05 c90ba89
Add description of presentation files
