### 2025-12-05 26ab286
try to fix the package version conflicts (#429)
Very minor changes.  Thanks, Aleksei.
