+++ test/booktests/READEME.md
@@ -40,5 +40,5 @@ To execute say two testbook files in parallel, run the following command in a te
     make booktests_parallel_no_docker TESTS="tb_Argonne_2023_Talk_Figures tb_Purdue_Talk_Figures"
 ```
 
-For a demo, see the Jupyter notebook, `demos/parsl_fest_2025.ipynb`.
+For a demo, see the Jupyter notebook, `demos/talk_paper_demos/ParslFest_2025/`.
 
