+++ demos/talk_paper_demos/parsl_fest_2025/output/parallel_output_2.txt
@@ -1,123 +1,0 @@
-rm -fr demos/.ipynb_checkpoints/*checkpoint.ipynb && \
-	find demos -name '*.ipynb' | while read nb; do \
-		base=$(basename "$nb" .ipynb); \
-		test_base=$(echo "$base" | sed 's/[-.]/_/g'); \
-		if echo "$nb" | grep -q "parsl_fest_2025"; then \
-			continue; \
-		fi; \
-		if ! ls test/booktests/tb_"$test_base".py > /dev/null 2>&1; then \
-			echo "    Missing test for: $nb -> Expected: test/booktests/tb_$test_base.py"; \
-		fi; \
-	done
-Total notebooks:        36
-Total test files:       32
-
-Generating missing booktest files...
-cd test/booktests/ && python generate_test.py --check-missing
-No missing test files found.
-rm -fr test/booktests/.ipynb_checkpoints/
-chmod +x scripts/find_local_only_folders.sh > /dev/null 2>&1
-for f in ; do \
-		rm -f "$f"; > /dev/null 2>&1; \
-	done
-
-Notebook tests with Parsl
-pip install -q -e ".[test]"  && \
-	cd test/booktests/ && \
-	rm -fr *.eps *.jpg *.pdf *.png *.part *.txt *.log && rm -fr logs && rm -fr runinfo prob_failure_gp_ci_plots && \
-	PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" \
-	python parsl_test_runner.py  -v --failfast && \
-	cd ../.. 
-Parsl configuration loaded successfully.
-Found 32 test modules to execute in parallel...
-All tests submitted to Parsl executor...
-[1/32] tb_elliptic_pde: PASSED
-tb_elliptic_pde ...     Memory used: 0.11 GB.  Test time: 28.54 s
-ok
-[2/32] tb_vectorized_qmc: PASSED
-tb_vectorized_qmc ...     Memory used: 0.09 GB.  Test time: 6.92 s
-ok
-[3/32] tb_pricing_options: PASSED
-tb_pricing_options ...     Memory used: 0.10 GB.  Test time: 5.64 s
-ok
-[4/32] tb_plot_proj_function: PASSED
-tb_plot_proj_function ...     Memory used: 0.11 GB.  Test time: 5.75 s
-ok
-[5/32] tb_asian_option_mlqmc: PASSED
-tb_asian_option_mlqmc ...     Memory used: 0.10 GB.  Test time: 42.18 s
-ok
-[6/32] tb_Purdue_Talk_Figures: PASSED
-[7/32] tb_sample_scatter_plots: PASSED
-tb_sample_scatter_plots ...     Memory used: 0.10 GB.  Test time: 3.92 s
-ok
-[8/32] tb_gaussian_diagnostics_demo: PASSED
-tb_gaussian_diagnostics_demo ...     Memory used: 0.13 GB.  Test time: 13.22 s
-ok
-[9/32] tb_qmcpy_intro: PASSED
-tb_qmcpy_intro ...     Memory used: 0.10 GB.  Test time: 2.48 s
-ok
-[10/32] tb_qei_demo_for_blog: PASSED
-tb_qei_demo_for_blog ...     Memory used: 0.10 GB.  Test time: 37.71 s
-ok
-[11/32] tb_digital_net_b2: PASSED
-tb_digital_net_b2 ...     Memory used: 0.10 GB.  Test time: 16.61 s
-ok
-[12/32] tb_joss2025: PASSED
[…]
