### 2025-12-05 a6130c2
restore strict test reqs w/ loosened dev reqs
