+++ pyproject.toml
@@ -80,16 +80,16 @@ test = [
     "matplotlib >= 3.9.0",
     ]
 test_torch = [
-    "torch >= 2.7.0",
+    "torch == 2.7.0",
     ]
 test_gpytorch = [
-    "gpytorch >= 1.11",
+    "gpytorch == 1.11",
     ]
 test_botorch = [
-    "botorch >= 0.10.0",
+    "botorch == 0.10.0",
     ]
 test_umbridge = [
-    "umbridge >= 1.2.4",
+    "umbridge == 1.2.4",
     ]
 docs = [
     "nbval >= 0.10.0",
@@ -107,7 +107,7 @@ docs = [
     "pylint >= 2.14.4",
 ]
 dev = [ # brew install weasyprint
-    "qmcpy[docs,test,test_gpytorch,test_botorch,test_umbridge]",
+    "qmcpy[docs,test,torch,gpytorch,botorch,umbridge]",
     "jupyterlab >= 4.0.0",
     "jupyterlab_execute_time >= 3.2.0",
     "ipynbname >= 2025.8.0.0",
