### 2025-12-05 7915f02
try to bump torch req
