### 2025-12-05 ab2a342
back to v2.7.0 torch tests
