### 2025-12-05 7a83a59
loosen the doctests themselves
