+++ pyproject.toml
@@ -80,16 +80,16 @@ test = [
     "matplotlib >= 3.9.0",
     ]
 test_torch = [
-    "torch == 2.7.0",
+    "torch >= 2.7.0",
     ]
 test_gpytorch = [
-    "gpytorch == 1.11",
+    "gpytorch >= 1.11",
     ]
 test_botorch = [
-    "botorch == 0.10.0",
+    "botorch >= 0.10.0",
     ]
 test_umbridge = [
-    "umbridge == 1.2.4",
+    "umbridge >= 1.2.4",
     ]
 docs = [
     "nbval >= 0.10.0",
