### 2025-12-06 7caaf84
gbm changes
