### 2025-12-06 b509563
Fix notebook errors in interactive plots and plots of QMCPy paths
