+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -6,12 +6,12 @@
 Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
  &  &  &   &  Error &  Error &   & &  \\
 \midrule
-QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.91370776 & 0.00000000 & - \\
-QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.64586080 & 0.00000000 & 1.41471314 \\
-QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.98026883 & 0.00000000 & - \\
-QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.64110380 & 0.00000000 & 1.42521033 \\
-QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.76989862 & 0.00000000 & 1.18678972 \\
-QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 3.47092987 & 0.00000000 & 0.26324581 \\
+QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.72002621 & 0.00000000 & - \\
+QMCPy & IIDStdUniform & 105.96842737 & 4.03136996 & 0.84131773 & 17.20606887 & 0.27234718 & 0.00000000 & 2.64378065 \\
+QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.78173499 & 0.00000000 & - \\
+QMCPy & Sobol & 110.04157186 & 5.80918796 & 4.91446222 & 15.42825087 & 0.27376281 & 0.00000000 & 2.63010964 \\
+QMCPy & Lattice & 102.68631550 & 6.15412494 & 2.44079413 & 15.08331388 & 0.27086386 & 0.00000000 & 2.65825872 \\
+QMCPy & Halton & 116.08198302 & 21.55953682 & 10.95487338 & 0.32209800 & 1.99024861 & 0.00000000 & 0.36177702 \\
 \bottomrule
 \end{tabular}
 \end{table}
