### 2025-12-06 650e04e
new plots
