### 2025-12-06 c1790c4
Update kernel output values in si_dsi_kernels.py
Make doctests reproducible by adding seed.
