+++ demos/GBM/averaged_mae.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
-from qmcpy_util import generate_qmcpy_paths
+from qmcpy_util_replications import generate_qmcpy_paths
 import config as cf
 import plot_util as pu
 from matplotlib.ticker import FixedLocator, FixedFormatter
