+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -6,12 +6,12 @@
 Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
  &  &  &   &  Error &  Error &   & &  \\
 \midrule
-QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.97977058 & 0.00000000 & - \\
-QMCPy & IIDStdUniform & 105.96842737 & 4.03136996 & 0.84131773 & 17.20606887 & 0.69803374 & 0.00000000 & 1.40361494 \\
-QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 1.03314749 & 0.00000000 & - \\
-QMCPy & Sobol & 110.04157186 & 5.80918796 & 4.91446222 & 15.42825087 & 0.68975755 & 0.00000000 & 1.42045647 \\
-QMCPy & Lattice & 102.68631550 & 6.15412494 & 2.44079413 & 15.08331388 & 0.82558300 & 0.00000000 & 1.18676206 \\
-QMCPy & Halton & 116.08198302 & 21.55953682 & 10.95487338 & 0.32209800 & 3.25633068 & 0.00000000 & 0.30088178 \\
+QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.93393791 & 0.00000000 & - \\
+QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.64938051 & 0.00000000 & 1.43819825 \\
+QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 1.03065195 & 0.00000000 & - \\
+QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.65544105 & 0.00000000 & 1.42489993 \\
+QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.82790568 & 0.00000000 & 1.12807283 \\
+QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 3.67974473 & 0.00000000 & 0.25380508 \\
 \bottomrule
 \end{tabular}
 \end{table}
