### 2025-12-06 e38be25
Take out useless functions
