+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -6,12 +6,12 @@
 Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
  &  &  &   &  Error &  Error &   & &  \\
 \midrule
-QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.93393791 & 0.00000000 & - \\
-QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.64938051 & 0.00000000 & 1.43819825 \\
-QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 1.03065195 & 0.00000000 & - \\
-QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.65544105 & 0.00000000 & 1.42489993 \\
-QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.82790568 & 0.00000000 & 1.12807283 \\
-QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 3.67974473 & 0.00000000 & 0.25380508 \\
+QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.64447216 & 0.00000000 & - \\
+QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.29036389 & 0.00000000 & 2.21953274 \\
+QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.72055169 & 0.00000000 & - \\
+QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.29408700 & 0.00000000 & 2.19143367 \\
+QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.28489466 & 0.00000000 & 2.26214197 \\
+QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 2.14989287 & 0.00000000 & 0.29976943 \\
 \bottomrule
 \end{tabular}
 \end{table}
