### 2025-12-06 af0d765
Remove useless functions from the code below.
