### 2025-12-06 5918719
Reproducing results from Larysa
