### 2025-12-07 19e2f45
averaged MAE for qmcpy + quantlib
