+++ demos/GBM/averaged_mae.py
@@ -1,68 +1,110 @@
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
-from qmcpy_util_replications import generate_qmcpy_paths
+import quantlib_util as qlu
+import qmcpy_util_replications as qpru
+import qmcpy_util as qpu
 import config as cf
 import plot_util as pu
 from matplotlib.ticker import FixedLocator, FixedFormatter
 
-def compute_mae_vs_paths(sampler_type='Sobol',
-                                replications=5):
+def compute_mae_vs_paths(method, sampler, replications=5, qp_seed = 42, ql_seed=7):
     """
-    Compute mean absolute error vs number of paths using independent QMCPy replications
+    Compute averaged MAE vs number of paths for all samplers
     """
-
-    # Get experiment configuration 
+    # Experiment configuration
     exp_cfg = cf.get_experiment_configurations()['paths']
     n_paths_range = exp_cfg['range']
     n_steps = exp_cfg['fixed_steps']
 
-    # Get GBM parameters
+    # GBM parameters
     gbm_params = cf.get_gbm_parameters()
     initial_value = gbm_params['initial_value']
     mu = gbm_params['mu']
-    diffusion = gbm_params['sigma']**2 # diffusion = sigma^2
+    sigma = gbm_params['sigma']
+    diffusion = sigma**2
     maturity = gbm_params['maturity']
 
-    # Compute theoretical mean
+    # Theoretical mean
     theoretical_mean = initial_value * np.exp(mu * maturity)
-    
+
     mean_errors = []
 
     for n_paths in n_paths_range:
-        # Generate GBM paths (with idependant replications) for each number of paths in n_paths_range
-        paths, gbm = generate_qmcpy_paths(
-            initial_value=initial_value,
-            mu=mu,
-            diffusion=diffusion,
-            maturity=maturity,
-            n_steps=n_steps,
-            n_paths=n_paths,
-            sampler_type=sampler_type,
-            replications=replications,
-        )
+        if method == "QMCPy":
+            # errors = []
+            # for r in range(replications):
+            #     seed = qp_seed + r
+            #     paths, _ = qpu.generate_qmcpy_paths(
+            #         initial_value = initial_value,
+            #         mu=mu,
+            #         diffusion=diffusion,
+            #         maturity=maturity,
+            #         n_steps=n_steps,
+            #         n_paths=n_paths,
+            #         sampler_type=sampler,
+            #         seed=seed
+            #     )
+            #     final_vals = paths[:, -1]
+            #     errors.append(np.abs(final_vals.mean() - theoretical_mean))
 
-        # Extract values at the final time
-        final_vals = paths[:, :, -1]  
+            paths, _ = qpru.generate_qmcpy_paths(
+                initial_value=initial_value,
+                mu=mu,
+                diffusion=diffusion,
+                maturity=maturity,
+                n_steps=n_steps,
+                n_paths=n_paths,
+                sampler_type=sampler,
+                replications=replications
+            )
+            # Extract values at the final time
+            final_vals = paths[:, :, -1]  
+            # Compute MAE for each replication and take the average
+            errors = np.abs(final_vals.mean(axis=1) - theoretical_mean)
 
-        # Compute MAE for each replication and take the average
-        errors_per_repl = np.abs(final_vals.mean(axis=1) - theoretical_mean)
-        mean_errors.append(errors_per_repl.mean())
+        elif method == "QuantLib":
+            errors = []
+            for r in range(replications):
[…]
