### 2025-12-07 3187bb9
Remove text output from Matplotlib and import statements that are not used.
