+++ demos/GBM/code/data_util.py
@@ -1,4 +1,3 @@
-import time
 import numpy as np
 import pandas as pd
 import quantlib_util as qlu
