+++ demos/GBM/code/plot_util.py
@@ -1,7 +1,6 @@
 
 import numpy as np
 import matplotlib.pyplot as plt
-import matplotlib.colors as mcolors
 from matplotlib.ticker import FixedLocator, FixedFormatter
 
 
