### 2025-12-07 78e9bb8
Increase readability and fix typos.
