+++ demos/GBM/code/plot_util.py
@@ -1,4 +1,3 @@
-
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.ticker import FixedLocator, FixedFormatter
@@ -15,7 +14,8 @@ def plot_error_comparison(ax, samplers, qmcpy_errors, quantlib_errors):
             ql_x.append(i)
             ql_errors.append(error)
     if ql_errors:
-        ax.bar([x - width/2 for x in ql_x], ql_errors, width, label='QuantLib', color='blue', alpha=0.8)
+        ax.bar([x - width/2 for x in ql_x], ql_errors, width,
+               label='QuantLib', color='blue', alpha=0.8)
     # Plot QMCPy data second (right side)
     ax.bar(x + width/2, qmcpy_errors, width, label='QMCPy', color='red', alpha=0.8)
 
@@ -40,21 +40,23 @@ def plot_performance_comparison(ax, samplers, qmcpy_times, quantlib_times):
                 ql_x.append(i)
                 ql_times.append(time)
         if ql_times:
-            ax.bar([x - width/2 for x in ql_x], ql_times, width, label='QuantLib', color='blue', alpha=0.8)
+            ax.bar([x - width/2 for x in ql_x], ql_times, width,
+                   label='QuantLib', color='blue', alpha=0.8)
         # Plot QMCPy data second (right side)
         ax.bar(x + width/2, qmcpy_times, width, label='QMCPy', color='red', alpha=0.8)
-        # Add speedup annotations where QuantLib data is available, positioned at center of QMCPy bars
+        # Add speedup annotations where QuantLib data is available, at center of QMCPy bars
         if len(ql_times) > 0:
             for i, (qmc_time, ql_time) in enumerate(zip(qmcpy_times, quantlib_times)):
                 if ql_time is not None:
                     speedup = ql_time / qmc_time
                     annotation_height = qmc_time + max(qmcpy_times) * 0.3
                     # Position arrow at center of QMCPy bar (i + width/2)
-                    ax.annotate(f'{speedup:.1f}x faster', 
-                               xy=(i + width/2, qmc_time), xytext=(i + width/2, annotation_height),
-                               ha='center', va='bottom', fontsize=9, fontweight='bold',
-                               arrowprops=dict(arrowstyle='->', color='blue', lw=1))
-        
+                    ax.annotate(f'{speedup:.1f}x faster',
+                                xy=(i + width/2, qmc_time),
+                                xytext=(i + width/2, annotation_height),
+                                ha='center', va='bottom', fontsize=9,
+                                fontweight='bold',
+                                arrowprops=dict(arrowstyle='->', color='blue', lw=1))
         ax.set_xlabel('Sampler Type')
         ax.set_ylabel('Execution Time (s)')
         ax.set_xticks(x)
@@ -62,7 +64,8 @@ def plot_performance_comparison(ax, samplers, qmcpy_times, quantlib_times):
         ax.legend()
         ax.grid(True, alpha=0.3)
     else:
-        ax.text(0.5, 0.5, 'Timing data not available\nRun previous cells to generate data', ha='center', va='center', transform=ax.transAxes, fontsize=12)
+        ax.text(0.5, 0.5, 'Timing data not available\nRun previous cells to generate data',
+                ha='center', va='center', transform=ax.transAxes, fontsize=12)
     
     ax.set_title('Performance Comparison', fontsize=16, fontweight='bold')
 
@@ -82,7 +85,8 @@ def get_plot_styling():
         }
     }
 
-def plot_single_series(ax, plot_data, series_name, x_col, y_col, title, xlabel, ylabel, log_scale=False, is_legend=False):
[…]
