### 2025-12-07 6b36a7e
Add README. Add headers to examples.ipynb.
