+++ demos/GBM/code/plot_util.py
@@ -1,6 +1,12 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.ticker import FixedLocator, FixedFormatter
+import os
+import scipy.stats as sc
+import qmcpy as qp
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.ticker import FixedLocator, FixedFormatter
 
 
 def plot_error_comparison(ax, samplers, qmcpy_errors, quantlib_errors):
@@ -68,7 +74,6 @@ def plot_performance_comparison(ax, samplers, qmcpy_times, quantlib_times):
                 ha='center', va='center', transform=ax.transAxes, fontsize=12)
     
     ax.set_title('Performance Comparison', fontsize=16, fontweight='bold')
-
 
 def get_plot_styling():
     """Define colors and markers for plotting"""
@@ -167,3 +172,175 @@ def create_parameter_sweep_plots(df):
     plt.tight_layout()
     plt.savefig('images/figure_7.png', bbox_inches='tight', dpi=150)
     plt.show()
+
+def plot_paths(motion_type, sampler, t_final, initial_value, drift, diffusion,
+               n, png_filename=None):
+    """
+    Plot realizations of Brownian Motion or Geometric Brownian Motion.
+
+    Args:
+        motion_type: 'BM' for Brownian Motion or 'GBM' for Geometric Brownian Motion
+        sampler: QMCPy sampler instance
+        t_final: Final time point
+        initial_value: Initial value S(0)
+        drift: Drift coefficient (mu)
+        diffusion: Diffusion coefficient (sigma^2 for GBM, sigma for BM)
+        n: Number of paths to generate
+        png_filename: Optional filename to save plot (saved to images/ directory)
+
+    Returns:
+        Motion object used for generation
+    """
+    if motion_type.upper() == 'BM':
+        motion = qp.BrownianMotion(sampler, t_final, initial_value, drift,
+                                   diffusion)
+        title = (f'Realizations of Brownian Motion using '
+                 f'{type(sampler).__name__} points')
+        ylabel = '$W(t)$'
+    elif motion_type.upper() == 'GBM':
+        motion = qp.GeometricBrownianMotion(sampler, t_final, initial_value,
+                                            drift, diffusion)
+        title = (f'Realizations of Geometric Brownian Motion using '
+                 f'{type(sampler).__name__} points')
+        ylabel = '$S(t)$'
+    else:
+        raise ValueError("motion_type must be 'BM' or 'GBM'")
+
+    t = motion.gen_samples(n)
+    initial_values = np.full((n, 1), motion.initial_value)
+    t_w_init = np.hstack((initial_values, t))
+    tvec_w_0 = np.hstack(([0], motion.time_vec))
+
+    plt.figure(figsize=(7, 4))
+    plt.plot(tvec_w_0, t_w_init.T)
+    plt.title(title)
+    plt.xlabel('$t$')
+    plt.ylabel(ylabel)
+    plt.xlim([tvec_w_0[0], tvec_w_0[-1]])
+    if png_filename:
+        os.makedirs('images', exist_ok=True)
+        plt.savefig(f'images/{png_filename}.png', bbox_inches='tight')
+    plt.show();
+
+    return motion
+
+def plot_gbm_paths_with_distribution(N, sampler, t_final, initial_value,
+                                     drift, diffusion, n):
+    """
+    Plot GBM paths with distribution of final values.
+
[…]
