### 2025-12-07 1434f19
Update date. Fix a bug in plot_util.py
