### 2025-12-07 0594ff4
Remove repeated lines
