### 2025-12-07 0ad071c
Fix GBM documentation at http://127.0.0.1:8000/demos/GBM/gbm_demo/
