### 2025-12-08 b72b3c7
fix typo and broken image of Fig 7 in gbm_demo's documentation
