### 2025-12-08 8d7dfba
Update README and averaged_mae.py
README.md:
* Corrected filename references.
* Removed references to non-existent files.

averaged_mae.py:
* Fixed PEP 8 spacing for parameter qp_seed and use the parameter in code.
*Enhanced documentation with.
*Removed an unused import.
* Cleaned up code by removing a commented-out block.
