### 2025-12-08 6dc8d1a
Update mkdocs.yml
