### 2025-12-08 9553816
Address copilot review comments
