### 2025-12-08 3540c03
Match parent class's method signature
