### 2025-12-08 a84eab0
Remove unnecessary space
