+++ demos/GBM/code/data_util.py
@@ -1,0 +1,216 @@
+import numpy as np
+import pandas as pd
+import quantlib_util as qlu
+import qmcpy_util as qpu
+import config as cf
+
+def add_theoretical_results(results_data, theoretical_mean, theoretical_std):
+    """Add theoretical values to results data"""
+    results_data.append({
+        'Method': 'Theoretical',
+        'Sampler': '-',
+        'Mean': theoretical_mean,
+        'Std Dev': theoretical_std,
+        'Mean Absolute Error': 0,
+        'Std Dev Error': 0,
+    })
+
+def add_quantlib_results(results_data, sampler_type, quantlib_final, theoretical_mean, theoretical_std):
+    """Add QuantLib results to results data"""
+    ql_emp_mean = np.mean(quantlib_final)
+    ql_emp_std = np.std(quantlib_final, ddof=1)
+    results_data.append({
+        'Method': 'QuantLib',
+        'Sampler': sampler_type,
+        'Mean': ql_emp_mean,
+        'Std Dev': ql_emp_std,
+        'Mean Absolute Error': abs(ql_emp_mean - theoretical_mean),
+        'Std Dev Error': abs(ql_emp_std - theoretical_std),
+    })
+
+def add_qmcpy_results(results_data, sampler_type, qmcpy_final, qp_emp_mean, theoretical_mean, theoretical_std):
+    """Add QMCPy results to results data"""
+    qp_emp_std = np.std(qmcpy_final, ddof=1)
+    results_data.append({
+        'Method': 'QMCPy',
+        'Sampler': sampler_type,
+        'Mean': qp_emp_mean,
+        'Std Dev': qp_emp_std,
+        'Mean Absolute Error': abs(qp_emp_mean - theoretical_mean),
+        'Std Dev Error': abs(qp_emp_std - theoretical_std),
+    })
+
+def process_sampler_data(sampler_type, results_data, theoretical_mean, theoretical_std, params_ql, params_qp):
+    """Process data for a single sampler type"""
+    # Initialize quantlib_paths to None
+    params_ql['sampler_type'] = sampler_type
+    params_qp['sampler_type'] = sampler_type
+    quantlib_paths, quantlib_final = None, None
+
+    # Generate paths for both libraries
+    if sampler_type in ['IIDStdUniform', 'Sobol']:
+        quantlib_paths, ql_gbm = qlu.generate_quantlib_paths(**params_ql)
+        quantlib_final = quantlib_paths[:, -1]
+    else:
+        quantlib_paths, ql_gbm = None, None
+        
+    qmcpy_paths, qp_gbm = qpu.generate_qmcpy_paths(**params_qp)
+
+    # Final value statistics
+    qmcpy_final = qmcpy_paths[:, -1]
+    qp_emp_mean = np.mean(qmcpy_final)
+
+    # Add results to data
+    if sampler_type in ['IIDStdUniform', 'Sobol'] and quantlib_final is not None:
+        add_quantlib_results(results_data, sampler_type, quantlib_final, theoretical_mean, theoretical_std)
+    add_qmcpy_results(results_data, sampler_type, qmcpy_final, qp_emp_mean, theoretical_mean, theoretical_std)
+    
+    return quantlib_paths, qmcpy_paths, ql_gbm, qp_gbm, params_ql, params_qp
+
+
+def create_timing_dataframe(quantlib_results, qmcpy_results, baseline_sampler):
+    """Create comprehensive timing table from benchmark results"""
+    timing_data = []
+    
+    # Add QuantLib data
+    for sampler_type, result in quantlib_results.items():
+        timing_data.append({
+            'Method': 'QuantLib',
+            'Sampler': sampler_type,
[…]
