### 2025-12-08 929be17
fix typos + update examples notebook
