### 2025-12-09 ec807b2
Fix typos.  Git rm .png.  Add testbook. Rename examples notebook.
