+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -1,17 +1,0 @@
-\begin{table}[tbp]
-\caption{GBM Final Value Statistics and Performance Comparison}
-\label{tab2}
-\begin{tabular}{ll@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r@{\hspace{0.4em}}r}
-\toprule
-Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
- &  &  &   &  Error &  Error &   & &  \\
-\midrule
-QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.60523902 & 0.00644429 & - \\
-QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.42061063 & 0.00697161 & 1.43895324 \\
-QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.70501040 & 0.00875881 & - \\
-QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.42916391 & 0.00279574 & 1.41027474 \\
-QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.57434715 & 0.00237917 & 1.05378606 \\
-QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 2.97119152 & 0.09537174 & 0.20370246 \\
-\bottomrule
-\end{tabular}
-\end{table}
