### 2025-12-09 64db53b
+doc strings. -test file and examples notebook.
