### 2025-12-09 e2ed4b1
final updates
