+++ demos/GBM/code/plot_util.py
@@ -144,7 +144,7 @@ def plot_single_series(ax, plot_data, series_name, x_col, y_col, title,
     if is_legend:
         ax.legend(fontsize=10)
 
-def create_parameter_sweep_plots(df):
+def create_parameter_sweep_plots(df,replications):
     """Create 4-panel plots from parameter sweep data"""
     # Filter out theoretical data
     plot_data = df[df['Method'] != 'Theoretical'].copy()
@@ -153,23 +153,23 @@ def create_parameter_sweep_plots(df):
     _, ((ax1, ax2), (ax3, ax4)) = plt.subplots(2, 2, figsize=(16, 12))
     # Panel 1: Mean Absolute Error vs n_steps (upper left)
     plot_single_series(ax1, plot_data, 'Time Steps', 'n_steps', 'Mean Absolute Error',
-                       'Mean Absolute Error vs Number of Time Steps\n(n_paths = 4,096)',
+                       f'Mean Absolute Error vs Number of Time Steps across {replications} Replications\n(n_paths = 4,096)',
                        'Number of Time Steps', 'Mean Absolute Error',
                        log_scale=True)
 
     # Panel 2: Runtime vs n_steps (upper right)
     plot_single_series(ax2, plot_data, 'Time Steps', 'n_steps', 'Runtime (s)',
-                       'Runtime vs Number of Time Steps\n(n_paths = 4,096)',
+                       f'Runtime vs Number of Time Steps\n(n_paths = 4,096)',
                        'Number of Time Steps', 'Runtime (seconds)', log_scale=True, is_legend=True)
     
     # Panel 3: Mean Absolute Error vs n_paths (lower left)
     plot_single_series(ax3, plot_data, 'Paths', 'n_paths', 'Mean Absolute Error',
-                       'Mean Absolute Error vs Number of Paths\n(n_steps = 252)',
+                       f'Mean Absolute Error vs Number of Paths across {replications} Replications\n(n_steps = 252)',
                        'Number of Paths', 'Mean Absolute Error', log_scale=True)
 
     # Panel 4: Runtime vs n_paths (lower right)
     plot_single_series(ax4, plot_data, 'Paths', 'n_paths', 'Runtime (s)',
-                       'Runtime vs Number of Paths\n(n_steps = 252)',
+                       f'Runtime vs Number of Paths\n(n_steps = 252)',
                        'Number of Paths', 'Runtime (seconds)', log_scale=True)
 
 def plot_paths(motion_type, sampler, t_final, initial_value, drift, diffusion,
