### 2025-12-10 cf1763c
Update assertion to use np.isclose for accuracy
