### 2025-12-10 0d333fa
Update test/test_integrate.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
