+++ demos/GBM/data_util/data_util.py
@@ -1,0 +1,271 @@
+import numpy as np
+import pandas as pd
+import quantlib_util as qlu
+import qmcpy_util as qpu
+import config as cf
+
+def add_theoretical_results(results_data, theoretical_mean, theoretical_std):
+    """
+    Add theoretical benchmark values to results data.
+    
+    Args:
+        results_data: List to append theoretical results to
+        theoretical_mean: Theoretical expected value E[S_T]
+        theoretical_std: Theoretical standard deviation of S_T
+    """
+    results_data.append({
+        'Method': 'Theoretical',
+        'Sampler': '-',
+        'Mean': theoretical_mean,
+        'Std Dev': theoretical_std,
+        'Mean Absolute Error': 0,
+        'Std Dev Error': 0,
+    })
+
+def add_quantlib_results(results_data, sampler_type, quantlib_final, theoretical_mean, theoretical_std):
+    """
+    Compute and add QuantLib simulation results to results data.
+    
+    Args:
+        results_data: List to append results to
+        sampler_type: Name of the sampler used
+        quantlib_final: Array of final values S_T from QuantLib simulation
+        theoretical_mean: Theoretical expected value for error computation
+        theoretical_std: Theoretical standard deviation for error computation
+    """
+    ql_emp_mean = np.mean(quantlib_final)
+    ql_emp_std = np.std(quantlib_final, ddof=1)
+    results_data.append({
+        'Method': 'QuantLib',
+        'Sampler': sampler_type,
+        'Mean': ql_emp_mean,
+        'Std Dev': ql_emp_std,
+        'Mean Absolute Error': abs(ql_emp_mean - theoretical_mean),
+        'Std Dev Error': abs(ql_emp_std - theoretical_std),
+    })
+
+def add_qmcpy_results(results_data, sampler_type, qmcpy_final, qp_emp_mean, theoretical_mean, theoretical_std):
+    """
+    Compute and add QMCPy simulation results to results data.
+    
+    Args:
+        results_data: List to append results to
+        sampler_type: Name of the sampler used
+        qmcpy_final: Array of final values S_T from QMCPy simulation
+        qp_emp_mean: Empirical mean already computed from qmcpy_final
+        theoretical_mean: Theoretical expected value for error computation
+        theoretical_std: Theoretical standard deviation for error computation
+    """
+    qp_emp_std = np.std(qmcpy_final, ddof=1)
+    results_data.append({
+        'Method': 'QMCPy',
+        'Sampler': sampler_type,
+        'Mean': qp_emp_mean,
+        'Std Dev': qp_emp_std,
+        'Mean Absolute Error': abs(qp_emp_mean - theoretical_mean),
+        'Std Dev Error': abs(qp_emp_std - theoretical_std),
+    })
+
+def process_sampler_data(sampler_type, results_data, theoretical_mean, theoretical_std, params_ql, params_qp):
+    """
+    Process and compare data for a single sampler type across both libraries.
+    
+    Args:
+        sampler_type: Type of sampler to test
+        results_data: List to append comparison results to
+        theoretical_mean: Theoretical expected value
+        theoretical_std: Theoretical standard deviation
+        params_ql: Dictionary of QuantLib parameters
+        params_qp: Dictionary of QMCPy parameters
+        
+    Returns:
+        tuple: (quantlib_paths, qmcpy_paths, ql_gbm, qp_gbm, params_ql, params_qp)
+    """
+    # Initialize quantlib_paths to None
+    params_ql['sampler_type'] = sampler_type
+    params_qp['sampler_type'] = sampler_type
[…]
