+++ demos/GBM/__init__/__init__.py
@@ -1,0 +1,5 @@
+"""
+GBM demonstration package.
+
+This package provides utilities for Geometric Brownian Motion demonstrations.
+"""
