### 2025-12-11 25fed42
Created draft of RELEASE.md which will be used by future contributors to be able to publish new releases to PyPi.
