### 2025-12-11 #9483 @alegresor
Thanks @lockwo, let me suggest the following reviewers from the [JOSS reviewers search](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=7&keywords=Monte+Carlo&name=&commit=Search) tool.

- [Vignesh Sathyaseelan](https://reviewers.joss.theoj.org/reviewers/18851) with GitHub username [vigneshS29](https://github.com/vigneshS29)
- [Adrien Corenflos](https://reviewers.joss.theoj.org/reviewers/8918) with GitHub username [AdrienCorenflos](https://github.com/AdrienCorenflos)
- [Matt Graham](https://reviewers.joss.theoj.org/reviewers/753)	with GitHub username [matt-graham](https://github.com/matt-graham)
- [shahmoradi](https://reviewers.joss.theoj.org/reviewers/1400) with GitHub username [shahmoradi](https://github.com/shahmoradi)

Outside the JOSS system, I would also suggest the following reviewers

- Pamphile Roy with GitHub username [tupui](https://github.com/tupui), the lead developer of `scipy`'s Quasi-Monte Carlo (QMC) library.
- Pierre L'Ecuyer with GitHub username [pierrelecuyer](https://github.com/pierrelecuyer), an expert in QMC with extensive (Q)MC software development experience. It may be easier to contact him through the email found on his website https://www.iro.umontreal.ca/~lecuyer/
- Dirk Nuyens with GitHub username [dnuy](https://github.com/dnuy), an expert in QMC with extensive (Q)MC software development experience. It may be easier to contact him through the email found on his website https://people.cs.kuleuven.be/~dirk.nuyens/
