### 2025-12-11 267363e
Added pip install command for when  fails.
