+++ demos/portfolio/portfolio_allocation_to_do.md
@@ -1,13 +1,13 @@
 # QMCPy 2.2 Portfolio allocation to-do:
 0. What are the differences between the two notebooks in `demos/portfolio`?
-1. Add Sobol and Halton samplers for weigths generation in the demo
+1. Add Sobol and Halton samplers for weights generation in the demo
 2. Add more text to the demo
-3. Create a single function to generate weigths for all sampler types
+3. Create a single function to generate weights for all sampler types
 4. Update visualizations in the demo
 5. Implement a more optimal simplex transformation for low-discrepancy sequences 
-6. Utilize QMCPy's 'replications' parameter for averaged results
+6. Utilize QMCPy's `replications` parameter for averaged results
 7. Make a function for collecting data (?)
 8. Clean up demo
 9. Add references to the demo
-10. Update poster   
+10. Update poster in Overleaf  
 
