### 2025-12-13 217514a
Fix a few bugs in notebook
