+++ demos/portfolio/portfolio_allocation_to_do.md
@@ -1,8 +1,6 @@
 # QMCPy 2.2 Portfolio allocation to-do
 
 0. What are the differences between the two notebooks in `demos/portfolio`?
-   - update end_date --- DONE
-   - save figures as .png in `images` --- TODO
 1. Add Sobol and Halton samplers for weights generation in the demo --- DONE
 2. Add more text in Markdown cells before code cells to the demo --- TODO
 3. Create a single function to generate weights for all sampler types --- DONE. See `gen_weights*()`.
@@ -10,8 +8,11 @@
 5. Implement a more optimal simplex transformation for low-discrepancy sequences --- TODO
 6. Utilize QMCPy's `replications` parameter for averaged results --- TODO
 7. Make a function for collecting data (?) --- TODO
-8. Clean up demo --- TODO
-9. Add references to the demo --- TODO
-10. Update poster in Overleaf --- TODO
+8. Update start_date and end_date --- DONE
+9. Save price datasets to `data` folder --- DONE
+10. Save figures as .png in `images` --- TODO
+11. Clean up demo --- TODO
+12. Add references to the demo --- TODO
+13. Update poster in Overleaf --- TODO
     
 
