### 2025-12-13 e895169
Make sure booktests.yml failed if failures occurred.
