### 2025-12-13 e78bff5
Fix booktests failures
