### 2025-12-13 2f5a78b
Add error log capture and display in parsl_test_runner.py
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
