### 2025-12-13 762d654
Simplify error logging by removing redundant stderr checks
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
