### 2025-12-13 f17f921
Add error handling for log file reading operations
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
