+++ test/booktests/parsl_test_runner.py
@@ -99,10 +99,13 @@ def execute_parallel_tests():
                 stdout_file = f'logs/test_{index}_{module}.out'
                 print(f"\n--- Error details for {module} (first attempt) ---")
                 if os.path.exists(stdout_file):
-                    with open(stdout_file, 'r') as f:
-                        content = f.read()
-                        # Print last 2000 chars to avoid log overflow
-                        print(content[-2000:] if len(content) > 2000 else content)
+                    try:
+                        with open(stdout_file, 'r') as f:
+                            content = f.read()
+                            # Print last 2000 chars to avoid log overflow
+                            print(content[-2000:] if len(content) > 2000 else content)
+                    except Exception as read_err:
+                        print(f"Could not read log file: {read_err}")
                 print(f"--- End of error details for {module} ---\n")
                 try:
                     # Resubmit the test for retry
@@ -125,10 +128,13 @@ def execute_parallel_tests():
                         retry_stdout_file = f'logs/test_{index}_{module}_retry.out'
                         print(f"\n--- Failure details for {module} (after retry) ---")
                         if os.path.exists(retry_stdout_file):
-                            with open(retry_stdout_file, 'r') as f:
-                                content = f.read()
-                                # Print last 2000 chars to avoid log overflow
-                                print(content[-2000:] if len(content) > 2000 else content)
+                            try:
+                                with open(retry_stdout_file, 'r') as f:
+                                    content = f.read()
+                                    # Print last 2000 chars to avoid log overflow
+                                    print(content[-2000:] if len(content) > 2000 else content)
+                            except Exception as read_err:
+                                print(f"Could not read log file: {read_err}")
                         print(f"--- End of failure details for {module} ---\n")
                         results.append((module, f'FAILED after retry: {e2}', 0))
                         status = 'FAILED'
