### 2025-12-13 5060083
Display actual Python exceptions in parsl_test_runner CI logs (#437)
* Initial plan

* Add error log capture and display in parsl_test_runner.py

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

* Simplify error logging by removing redundant stderr checks

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

* Add error handling for log file reading operations

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

---------

Co-authored-by: alegresor <agsorokin3@gmail.com>
Co-authored-by: copilot-swe-a
[…]
