### 2025-12-13 ca2b5aa
Move RELEASE.md to docs. Add to documentation.
