+++ mkdocs.yml
@@ -3,6 +3,7 @@ nav:
   - Quasi Monte Carlo Software: README.md
   - Components: components.md
   - Contributing: CONTRIBUTING.md
+  - Release: RELEASE.md
   - Community: community.md
   - Package Reference:
     - Discrete Distributions: api/discrete_distributions.md
