### 2025-12-14 cdb3974
Add .md as file extension to TODO file. DONE some items in there.
