+++ demos/GBM/averaged_mae.py
@@ -1,34 +1,49 @@
 import numpy as np
+import pandas as pd
 import matplotlib.pyplot as plt
 import quantlib_util as qlu
-import qmcpy_util_replications as qpru
+import qmcpy_util as qpu
 import config as cf
 import plot_util as pu
 from matplotlib.ticker import FixedLocator, FixedFormatter
 
-def compute_mae_vs_paths(method, sampler, replications=5, qp_seed=42, ql_seed=7):
+
+def _compute_mae(
+    method: str,
+    sampler: str,
+    sweep_type: str,
+    replications: int = 5,
+    qp_seed: int = 42,
+    ql_seed: int = 7
+) -> tuple:
     """
-    Compute averaged MAE vs number of paths for all samplers.
+    Compute averaged MAE for a given sweep type (paths or steps).
     
-    This computes the Mean Absolute Error (MAE) of the MEAN ESTIMATOR, defined as:
+    This is the unified internal function that computes Mean Absolute Error (MAE)
+    of the MEAN ESTIMATOR, defined as:
         MAE = E[|mean(paths) - theoretical_mean|]
-    where the expectation is taken over replications. This measures how accurately
-    the (quasi) Monte Carlo mean estimator approximates the true theoretical mean.
+    where the expectation is taken over replications.
     
     Args:
         method: Either "QMCPy" or "QuantLib"
         sampler: Sampler type (e.g., 'IIDStdUniform', 'Sobol', 'Lattice', 'Halton')
+        sweep_type: Either "paths" (sweep n_paths) or "steps" (sweep n_steps)
         replications: Number of independent replications to average over
         qp_seed: Random seed for QMCPy samplers
         ql_seed: Random seed for QuantLib samplers
         
     Returns:
-        tuple: (n_paths_range, mean_errors) - path counts and corresponding MAEs
+        tuple: (sweep_range, mean_errors) - sweep values and corresponding MAEs
     """
     # Experiment configuration
-    exp_cfg = cf.get_experiment_configurations()['paths']
-    n_paths_range = exp_cfg['range']
-    n_steps = exp_cfg['fixed_steps']
+    exp_cfg = cf.get_experiment_configurations()[sweep_type if sweep_type == "paths" else "time_steps"]
+    sweep_range = exp_cfg['range']
+    
+    # Determine fixed parameter based on sweep type
+    if sweep_type == "paths":
+        fixed_steps = exp_cfg['fixed_steps']
+    else:  # steps
+        fixed_paths = exp_cfg['fixed_paths']
 
     # GBM parameters
     gbm_params = cf.get_gbm_parameters()
@@ -43,10 +58,18 @@ def compute_mae_vs_paths(method, sampler, replications=5, qp_seed=42, ql_seed=7)
 
     mean_errors = []
 
-    for n_paths in n_paths_range:
+    for sweep_val in sweep_range:
+        # Set n_paths and n_steps based on sweep type
+        if sweep_type == "paths":
+            n_paths = sweep_val
+            n_steps = fixed_steps
+        else:
+            n_paths = fixed_paths
+            n_steps = sweep_val
+
         if method == "QMCPy":
             seed = qp_seed
-            paths, _ = qpru.generate_qmcpy_paths(
+            paths, _ = qpu.generate_qmcpy_paths(
                 initial_value=initial_value,
                 mu=mu,
                 diffusion=diffusion,
@@ -83,10 +106,38 @@ def compute_mae_vs_paths(method, sampler, replications=5, qp_seed=42, ql_seed=7)
 
         mean_errors.append(errors.mean())
 
-    return n_paths_range, mean_errors
+    return sweep_range, mean_errors
 
 
-def plot_mae_vs_paths(replications=5):
+def compute_mae_vs_paths(
+    method: str,
[…]
