+++ demos/GBM/gbm_todo.md
@@ -1,0 +1,11 @@
+# QMCPy 2.2 GBM To-do:
+
+1. Add 'replications' parameter to qmcpy_util.py --- DONE
+2. Delete qmcpy_util_replications.py --- DONE
+3. Combine averaged MAE computation functions to reduce repeated code --- DONE
+4. Look into Sobol randomization in QuantLib --- DONE
+5. Add replications to the MAE Comparison bar chart (figure_6) --- DONE
+6. Add more text to the demo --- TODO
+7. Add type hints to key functions --- DONE
+8. Update online blog in Overleaf and `qmcpy.org` --- TODO
+9. Prepare for PyData Chicago talk (title, abstract, speaker bio, and slides) --- TODO
