+++ docs/releases/QMCPy-2.1-announcement.md
@@ -1,0 +1,29 @@
+# QMCPy v2.1 Released – Quasi-Monte Carlo Community Software in Python
+
+**Date**: 2025-12-10
+
+We are happy to announce version 2.1 of QMCPy, our open-source Python library for quasi-Monte Carlo (QMC) methods used to approximate multivariate integrals. This release continues our work to provide reliable, modular, and well-tested QMC software for research and teaching.
+
+Notable additions and improvements in v2.1:
+
+1. Adaptive smoothness kernel and Cholesky option for multitask kernels.  
+2. Improved numerical stability (overcame cancellation errors) and other robustness fixes.  
+3. Improved demos of Geometric Brownian Motion (GBM) integrand support for financial applications.  
+4. Integration with [**Parsl**](https://parsl-project.org) for parallel workflows of unit tests using [**testbook**](https://github.com/nteract/testbook) for Jupyter notebooks to improve reliability of demos and code base.  
+5. Class installation option for Illinois Tech students and instructors using QMCPy in coursework.  
+6. Documentation updates, demo patches, and various bug fixes across the codebase.  
+
+We welcome your questions, feedback, and contributions.
+
+**The QMCPy Team**
+
+
+## Useful links
+
+- Documentation: https://qmcsoftware.github.io/QMCSoftware/
+- GitHub repository: https://github.com/QMCSoftware/QMCSoftware
+- PyPI: https://pypi.org/project/qmcpy/
+- Release notes / changelog: https://github.com/QMCSoftware/QMCSoftware/releases/tag/v2.1
+
+
+
