### 2025-12-14 83a67fe
Fix testbook failures
