+++ pyproject.toml
@@ -78,6 +78,8 @@ test = [
     "scikit-optimize",
     "ipykernel",
     "matplotlib >= 3.9.0",
+    "pandas >= 1.3.0",
+    "yfinance >= 0.2.0",
     ]
 test_torch = [
     "torch >= 2.7.0",
