### 2025-12-14 4977472
unified plotting function
