+++ pyproject.toml
@@ -80,6 +80,7 @@ test = [
     "matplotlib >= 3.9.0",
     "pandas >= 1.3.0",
     "yfinance >= 0.2.0",
+    "quantlib >= 1.38",
     ]
 test_torch = [
     "torch >= 2.7.0",
