### 2025-12-15 bee80ed
+ ipywidgets
