+++ pyproject.toml
@@ -80,7 +80,8 @@ test = [
     "matplotlib >= 3.9.0",
     "pandas >= 1.3.0",
     "yfinance >= 0.2.0",
-    "quantlib >= 1.38",
+    "quantlib == 1.38",
+    "ipywidgets == 8.1.7",
     ]
 test_torch = [
     "torch >= 2.7.0",
