+++ workouts/asset_allocation/PortfolioAllocation.py
@@ -1,640 +1,0 @@
-# -*- coding: utf-8 -*-
-"""portfolio-allocation.ipynb
-
-Original file is located at
-    https://colab.research.google.com/drive/1XOLxjJ71fsTl3DUtr0zOitcLuvszKu37
-
-MORE EFFICIENT PORTFOLIO ALLOCATION WITH QUASI-MONTE CARLO METHODS USING QMCPY
-Larysa Matiukha and Sou-Cheng T. Choi
-
-Illinois Institute of Technology
-
-Modification date: 8/13/2023
-Creation date: 8/1/2023
-"""
-
-import qmcpy as qp
-import pandas as pd
-import numpy as np
-import yfinance as yf
-import time
-import timeit
-import matplotlib.pyplot as plt
-from qmcpy import *
-import itertools
-
-
-# To display the entire DataFrame without truncation
-pd.set_option('display.max_columns', None)
-# Setting the display precision for pandas DataFrame
-pd.set_option('display.precision', 6)
-
-class PortfolioAllocation:
-
-    def __init__(self, start_date='2012-01-01', end_date='2023-08-04', split_date='2021-01-01',
-                tickers=["AAPL", "AMZN", "CSCO", "IBM"],
-                description=["Apple", "Amazon", "CISCO", "IBM"],
-                num_ports=[2**14],
-                seed=42, methods=["iid", "lattice", "sobol"]):
-        self.start_date = start_date
-        self.end_date = end_date
-        self.split_date = split_date
-        self.tickers = tickers
-        self.description = description
-        self.num_ports = num_ports
-        self.seed = seed
-        self.n = len(tickers)
-        self.methods = methods
-        self.output_filename = f"qmc_portfolio_{self.n}.csv"
-        self.outputs = {}
-
-    def _download_data(self):
-        tickers = self.tickers
-        description = self.description
-        col_names = ['Ticker', 'Company', 'Date', 'Adj Close Price', 'Volume']
-        df = pd.DataFrame(columns=col_names)
-
-        for i, ticker in enumerate(tickers):
-            company = description[i]
-            data = yf.download(ticker, start=self.start_date, end=self.end_date)
-            data['Ticker'] = ticker
-            data['Company'] = company
-            data.reset_index(inplace=True)
-            data = data[['Ticker', 'Company', 'Date', 'Adj Close', 'Volume']]
-            data.columns = col_names
-            df = pd.concat([df, data], ignore_index=True)
-
-        return df
-
-
-
-    def _get_log_ret(self, df):
-        stocks = df.pivot(index='Date', columns='Ticker', values='Adj Close Price')
-        self.stocks_df = stocks
-        log_ret = np.log(stocks / stocks.shift(1))
-        return log_ret
-
-
-    def gen_weights_lattice(self, n, num_ports):
-        """ generating weights with randomized lattice points"""
-        l = qp.Lattice(dimension=n, seed=self.seed, is_parallel=False)
-        weights = l.gen_samples(num_ports)  # using lattice points instead of iid
-        weights /= np.sum(weights, axis=1)[:, np.newaxis]
-        return weights
-
-    def gen_weights_sobol(self, n, num_ports):
-        """ generating weights with randomized lattice points"""
-        ld = qp.Sobol(n, seed=self.seed)  # define the generator
-        weights = ld.gen_samples(num_ports)  # generate points
-        weights /= np.sum(weights, axis=1)[:, np.newaxis]
-        return weights
-    def gen_weights_lattice_parallel(self, n, num_ports):
-        """ generating weights with randomized lattice points"""
-        l = qp.Lattice(dimension=n, seed=self.seed, order='mps', is_parallel=True)
[…]
