### 2025-12-15 8780327
Use helper functions
