### 2025-12-15 c3e0f18
Added information regarding 2FA when setting up account on PyPi in the Setup section.
