### 2025-12-15 67b2811
Added information regarding 2FA when setting up accounts on PyPI and Test PyPI in the Setup section.
