### 2025-12-15 ed2aa0e
Shorten testbook runtime by injecting is_debug=True
