### 2025-12-15 da3ff28
Update tb_gbm_examples.py
