### 2025-12-15 9186d4a
Increase timeout. Add nbconvert.
