+++ test/booktests/tb_gbm_demo.py
@@ -8,7 +8,7 @@ class NotebookTests(BaseNotebookTest):
         fix_gbm_symlinks(notebook_dir)
         self.execute_notebook_file(notebook_path, timeout=600)
         # Toggle code cell [3] cf.is_debug -> True before executing
-        run_notebook(notebook_path, notebook_dir, change_value=True)
+        run_notebook(notebook_path, notebook_dir, change_value=True, timeout=600)
         
         
 if __name__ == '__main__':
