### 2025-12-16 755e81d
Update gbm to-do list
