### 2025-12-16 b460317
Stop tracking PNGs: update .gitignore and remove PNGs from index
