+++ test/booktests/tb_archive.py
@@ -1,0 +1,12 @@
+import unittest
+from testbook import testbook
+from __init__ import TB_TIMEOUT, BaseNotebookTest
+
+class NotebookTests(BaseNotebookTest):
+
+    @testbook('../../demos/portfolio/archive.ipynb', execute=True, timeout=TB_TIMEOUT)
+    def test_archive_notebook(self, tb):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
