### 2025-12-16 e6cbd21
Shorten run time
