### 2025-12-16 db3cb3c
Eliminate legacy functions. Add seed for reproducibility.
