+++ test/booktests/tb_portfolio_allocation_demo.py
@@ -17,6 +17,10 @@ class NotebookTests(BaseNotebookTest):
             "n_ports = [2**13, 2**14, 2**15]",
             "start_date = '2014-01-01'", 
             "dimensions = [5, 10, 20, 50, 100, 200, 500, 1000]",
+            'tickers1 = ["AAPL", "AMZN", "CSCO", "IBM", "TSLA", "META", "ABNB", "UPS", "NFLX", "MRNA"]',
+            'description1 = ["Apple", "Amazon", "CISCO", "IBM", "Tesla", "Meta", "Airbnb", "UPS", "Netflix", "Moderna"]"',
+            'tickers2 = ["AAPL", "AMZN", "CSCO","IBM","TSLA","META","ABNB","UPS","NFLX","MRNA","^IXIC", "T","GE","FMC","AMC","JPM","DIS","CVX","GOOGL","BA"]',
+            'description2 = ["Apple", "Amazon", "CISCO", "IBM","Tesla","Meta","Airbnb","UPS","Netflix","Moderna","NASDAQ","AT&T","General Electric","FMC","AMC","JPMorgan","Disney","Chevron","Google","Boeing"]',
         ]
         new_value = [
             """params = [
@@ -24,7 +28,11 @@ class NotebookTests(BaseNotebookTest):
     (10, 2**8)]""",
             "n_ports = [2**7, 2**8]",
             "start_date = '2025-01-01'", 
-            "dimensions = [5, 10]"
+            "dimensions = [5, 10]",
+            'tickers1 = ["AAPL", "AMZN"]',
+            'description1 = ["Apple", "Amazon"]"',
+             'tickers2 = ["AAPL", "AMZN", "CSCO"]',
+            'description2 = ["Apple", "Amazon", "CISCO"]',
         ]
         run_notebook(notebook_path, notebook_dir, change_value=True, 
                      value=value, new_value=new_value, timeout=TB_TIMEOUT)
