+++ test/booktests/generate_test.py
@@ -101,7 +101,7 @@ def generate_missing_tests(demos_dir="../../demos", output_dir=None):
         
         # Skip all notebooks in demos/talk_paper_demos/Parslfest_2025/
         # These are presentation/demo notebooks and should not have tests generated
-        if "Parslfest_2025" in str(notebook_path):
+        if "Parslfest_2025" in str(notebook_path) or notebook_path.name in ["archive.ipynb"]:
             continue
             
         # Convert notebook name to test file name
