### 2025-12-17 358f5c1
Change Lookback to include start price
