+++ qmcpy/integrand/financial_option.py
@@ -423,7 +423,7 @@ class FinancialOption(AbstractIntegrand):
         v[flag] = np.maximum(self.strike_price-v[flag],0)
         return v
     
-    def payoff_lookback_call(self, gbm): #incllude start price in min
+    def payoff_lookback_call(self, gbm): #include start price in min
         min_path = np.minimum(gbm.min(-1), self.start_price)
         return gbm[..., -1] - min_path
     
