### 2025-12-17 813605d
Shorten run time of tb_joss.py
