+++ test/booktests/__init__.py
@@ -8,6 +8,8 @@ import gc
 import time
 import os
 import subprocess
+from testbook import testbook
+import nbformat
 
 TB_TIMEOUT = 3600
 subprocess.run(['pip', 'install', '-q', 'psutil', 'testbook', 'parsl'], check=False)
@@ -42,16 +44,56 @@ class BaseNotebookTest(unittest.TestCase):
         notebook_path = os.path.abspath(notebook_path)
         return notebook_path, os.path.dirname(notebook_path)
 
-    def execute_notebook_file(self, notebook_path, timeout=TB_TIMEOUT):
-        """Execute a notebook file using nbconvert's ExecutePreprocessor."""
+    def fix_gbm_symlinks(self, notebook_dir, symlinks_to_fix=None):
+        """Fix or create symlinks inside a demo notebook directory."""
+        code_dir = os.path.join(notebook_dir, 'gbm_code')
+        if not symlinks_to_fix:
+            return
+        for module in symlinks_to_fix:
+            symlink_path = os.path.join(notebook_dir, module)
+            target_path = os.path.join(code_dir, module)
+            if os.path.islink(symlink_path):
+                link_target = os.readlink(symlink_path)
+                if link_target.startswith('code/') or not os.path.exists(symlink_path):
+                    os.remove(symlink_path)
+                    os.symlink(f'gbm_code/{module}', symlink_path)
+            elif not os.path.exists(symlink_path) and os.path.exists(target_path):
+                os.symlink(f'gbm_code/{module}', symlink_path)
+
+        
+    def change_notebook_cells(self, notebook_path, replacements=None):
+        """Modify code-cell sources in a notebook file
+        """
+        notebook_dir = os.path.dirname(notebook_path)
+        original_dir = os.getcwd()
+        try:
+            os.chdir(notebook_dir)
+            if not replacements:
+                return
+            with testbook(os.path.basename(notebook_path), execute=False) as tb:
+                for cell in tb.nb.cells:
+                    if cell.get('cell_type') == 'code':
+                        src = cell.get('source', '')
+                        for old, new in replacements.items():
+                            if old in src:
+                                src = src.replace(old, new)
+                        cell['source'] = src
+                #nbformat.write(tb.nb, os.path.basename(notebook_path))
+        except Exception as e:
+            print(f"Error modifying notebook cells: {e}")
+        finally:
+            os.chdir(original_dir)
+
+
+    def run_notebook(self, notebook_path, timeout=TB_TIMEOUT):
+        """Execute a notebook file using nbconvert's ExecutePreprocessor.
+        """
         import nbformat
         from nbconvert.preprocessors import ExecutePreprocessor
 
         with open(notebook_path) as f:
             nb = nbformat.read(f, as_version=4)
-
         ep = ExecutePreprocessor(timeout=timeout, kernel_name='python3')
-
         try:
             ep.preprocess(nb, {'metadata': {'path': os.path.dirname(notebook_path)}})
         except Exception as e:
