### 2025-12-17 338af98
Update SciPyWrapper demo/docs and Student t joint example
