+++ qmcpy/true_measure/scipy_wrapper.py
@@ -1,6 +1,8 @@
 from .abstract_true_measure import AbstractTrueMeasure
 from ..util import DimensionError, ParameterError
-from ..discrete_distribution.abstract_discrete_distribution import AbstractDiscreteDistribution
+from ..discrete_distribution.abstract_discrete_distribution import (
+    AbstractDiscreteDistribution,
+)
 from ..discrete_distribution import DigitalNetB2
 import numpy as np
 import scipy.stats
@@ -10,19 +12,19 @@ import warnings
 class _MVNAdapter:
     """
     Small adapter that turns a SciPy multivariate normal like object into
-    something with a simple "transform(u)" interface.
+    something with a simple ``transform(u)`` interface.
 
     Idea:
       1. Start from u in (0,1)^d.
-      2. Map to standard normals z via norm.ppf.
-      3. Apply a Cholesky factor to inject the correlation structure.
+      2. Map to standard normals z via ``norm.ppf``.
+      3. Apply Cholesky to inject the correlation structure.
     """
 
     def __init__(self, mvn_like):
-        # Keep the original object so we can still call logpdf later.
+        # Keep the original object around so that we can still call logpdf.
         self._mvn = mvn_like
 
-        # SciPy's multivariate_normal has attributes "mean" and "cov".
+        # SciPy's multivariate_normal has attributes ``mean`` and ``cov``.
         mean = np.asarray(mvn_like.mean)
         cov = np.asarray(mvn_like.cov)
 
@@ -61,7 +63,7 @@ class _MVNAdapter:
         # Map to i.i.d. standard normals.
         z = scipy.stats.norm.ppf(u_clip)
 
-        # Flatten, apply Cholesky, then reshape back.
+        # Flatten, apply Cholesky, then reshape back to original shape.
         z_flat = z.reshape(-1, self.dim)
         x_flat = z_flat @ self._chol.T + self._mean
         return x_flat.reshape(z.shape)
@@ -85,102 +87,48 @@ class SciPyWrapper(AbstractTrueMeasure):
     r"""
     True measure that wraps SciPy style distributions.
 
-    It supports two main modes:
+    This class keeps the original behaviour of SciPyWrapper with
+    independent 1D marginals and adds an optional "joint" mode for
+    dependent distributions.
 
-      1. Independent marginals (original behavior):
-         - You pass a single 1D SciPy frozen distribution, or a list of them.
-         - Each dimension is transformed independently via ppf.
+    Examples:
+        Independent marginals from ``scipy.stats``:
 
-      2. Joint distributions with dependence (new behavior):
-         - You pass a single "joint" object with a method transform(u).
-         - The transform takes u in (0,1)^d and returns x in R^d.
-         - Optional logpdf(x) lets us define weights.
+        >>> from qmcpy.discrete_distribution import DigitalNetB2
+        >>> import scipy.stats as stats
+        >>> tm = SciPyWrapper(
+        ...     sampler=DigitalNetB2(3, seed=7),
+        ...     scipy_distribs=[
+        ...         stats.uniform(loc=1, scale=2),
+        ...         stats.norm(loc=0, scale=1),
+        ...         stats.gamma(a=5, loc=0, scale=2)])
+        >>> x = tm(2)
+        >>> x.shape
+        (2, 3)
 
-    We also support user defined custom distributions that look SciPy like:
-    anything with ppf and pdf or logpdf can be plugged in. For custom
-    distributions that are not native SciPy frozen distributions, we emit
-    a warning and run light sanity checks so the user knows to double
-    check their implementation.
+        Joint multivariate normal passed as a single object:
 
-    ------------------------------------------------------------------------
[…]
