### 2025-12-17 eed7103
Clarify comment in tb_gbm_demo.py
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
