### 2025-12-17 0e75866
fixed docs for lookback
