+++ demos/GBM/quantlib_util.py
@@ -1,1 +1,1 @@
-code/quantlib_util.py
+gbm_code/quantlib_util.py
