+++ demos/GBM/gbm_code/qmcpy_util.py
@@ -1,7 +1,19 @@
 import qmcpy as qp
 
-def create_qmcpy_sampler(sampler_type, dimension, replications, seed=42):
-    """Create a sampler instance based on type and dimension"""
+
+def create_qmcpy_sampler(sampler_type: str, dimension: int, replications: int = 1, seed: int = 42):
+    """
+    Create a sampler instance based on type and dimension.
+    
+    Args:
+        sampler_type: Type of sampler ('IIDStdUniform', 'Sobol', 'Lattice', 'Halton')
+        dimension: Dimension of the sampler (typically n_steps)
+        replications: Number of independent replications
+        seed: Random seed for reproducibility
+        
+    Returns:
+        QMCPy sampler instance
+    """
     if sampler_type == 'IIDStdUniform':
         return qp.IIDStdUniform(dimension, replications, seed=seed)
     elif sampler_type == 'Sobol':
@@ -14,9 +26,17 @@ def create_qmcpy_sampler(sampler_type, dimension, replications, seed=42):
         raise ValueError(f"Unsupported sampler type: {sampler_type}")
 
 
-def generate_qmcpy_paths(initial_value, mu, diffusion, maturity, n_steps,
-                         n_paths, sampler_type='IIDStdUniform',
-                         replications=1, seed=42):
+def generate_qmcpy_paths(
+    initial_value: float,
+    mu: float,
+    diffusion: float,
+    maturity: float,
+    n_steps: int,
+    n_paths: int,
+    sampler_type: str = 'IIDStdUniform',
+    replications: int = 1,
+    seed: int = 42
+):
     """
     Generate Geometric Brownian Motion paths using QMCPy with multiple replications.
     
@@ -32,11 +52,11 @@ def generate_qmcpy_paths(initial_value, mu, diffusion, maturity, n_steps,
         seed: Random seed for reproducibility
         
     Returns:
-        tuple: (paths, gbm) where paths has shape (replications, n_paths, n_steps)
+        tuple: (paths, gbm) where paths has shape (n_paths, n_steps) if replications=1,
+               or (replications, n_paths, n_steps) if replications>1,
                and gbm is the GeometricBrownianMotion object
     """
-    sampler = create_qmcpy_sampler(
-        sampler_type, n_steps, replications, seed)
+    sampler = create_qmcpy_sampler(sampler_type, n_steps, replications, seed)
     gbm = qp.GeometricBrownianMotion(
         sampler, t_final=maturity, initial_value=initial_value,
         drift=mu, diffusion=diffusion)
