+++ pyproject.toml
@@ -78,6 +78,11 @@ test = [
     "scikit-optimize",
     "ipykernel",
     "matplotlib >= 3.9.0",
+    "pandas >= 1.3.0",
+    "yfinance >= 0.2.0",
+    "quantlib == 1.38",
+    "ipywidgets == 8.1.7",
+    "nbconvert >= 7.2.9",
     ]
 test_torch = [
     "torch >= 2.7.0",

[…]
