### 2025-12-18 7a76f6c
Fix a bug and speedup tb_jboss2025
