+++ test/booktests/tb_joss2025.py
@@ -6,7 +6,8 @@ from __init__ import TB_TIMEOUT, BaseNotebookTest
 
 class NotebookTests(BaseNotebookTest):
 
-    notebook_path = '../../demos/talk_paper_demos/JOSS2025/joss2025.ipynb'
+    notebook_dir = '../../demos/talk_paper_demos/JOSS2025/'
+    notebook_path = f'{notebook_dir}joss2025.ipynb'
     
     def setUp(self):
         super().setUp()  # Call parent setUp first to initialize timing attributes
@@ -16,7 +17,7 @@ class NotebookTests(BaseNotebookTest):
             'tueplots'
         ], check=False)
         # Create the JOSS2025.outputs directory that the notebook expects
-        self.output_dir = os.path.join(self.notebook_path, 'JOSS2025.outputs')
+        self.output_dir = os.path.join(self.notebook_dir, 'JOSS2025.outputs')
         if not os.path.exists(self.output_dir):
             os.makedirs(self.output_dir, exist_ok=True)
             self._created_output_dir = True
