### 2025-12-18 ada7b74
Speed up tests. Add README.
