+++ test/README.md
@@ -1,0 +1,263 @@
+# Test Targets Guide
+
+This document describes the available test targets in the Makefile for QMCSoftware. All targets use pytest with parallel execution (via `pytest-xdist`) when available.
+
+## Quick Reference
+
+| Target | Purpose | Speed | Use Case |
+|--------|---------|-------|----------|
+| `make tests_fast` ⚡ | **Recommended**: All tests in parallel | Fast | Default choice: doctests + unittests + booktests concurrently |
+| `make tests_no_docker` | All tests sequentially (no Docker) | Slow | Conservative validation; isolates flaky tests |
+| `make tests` | All tests with Docker support | Very Slow | Complete validation with GPU/Docker dependencies |
+| `make unittests` | Unit tests only | Fast | Quick feedback on code changes |
+| `make doctests_no_docker` | Doctests (excludes GPU/Docker) | Moderate | Validate docstring examples |
+| `make doctests` | All doctests with Docker | Slow | Full docstring validation |
+| `make booktests_no_docker` | Jupyter notebook tests | Slow | Validate demo notebooks |
+| `make booktests_parallel_no_docker` | Notebook tests with Parsl parallelization | Variable | Distributed notebook execution |
+| `make coverage` | Display coverage report | Instant | View test coverage summary |
+| `make delcoverage` | Reset coverage tracking | Instant | Start fresh coverage analysis |
+
+
+## Detailed Descriptions
+
+### Core Test Targets (Recommended)
+
+#### `make tests_fast` ⚡ **RECOMMENDED**
+**Fastest option**: Runs doctests, unittests, and booktests **concurrently** in background processes.
+- **Parallelization**: All three test families run simultaneously, not sequentially
+- **Cleanup**: Removes invalid distribution artifacts before running
+- **Time**: ~30–60 seconds (depending on CPU cores and notebook complexity)
+- **Coverage**: Full summary report at the end
+- **Use when**: You want comprehensive testing with maximum speed
+
+#### `make tests_no_docker`
+Runs all tests sequentially: doctests, unittests, and generates coverage reports (excludes Docker-dependent tests).
+- **Cleans** invalid distribution artifacts before running (via `scripts/cleanup_invalid_dist.py`)
+- **Sequence**: doctests_no_docker → unittests → coverage report
+- **Time**: ~60–120 seconds
+- **Coverage**: Full summary report
+- **Use when**: Pre-commit or CI/CD validation (without Docker)
+
+#### `make tests`
+Runs all tests **sequentially** with full Docker support (for GPU-heavy and UMBridge tests).
+- **Includes**: Full doctests suite (with umbridge and markdown validation)
+- **Time**: ~120–180+ seconds
+- **Coverage**: Full summary report
+- **Use when**: Complete validation with Docker dependencies available
+
+---
+
+### Doctest Targets
+
+#### `make doctests_no_docker` (Composite)
+Runs doctests excluding GPU and Docker dependencies.
+- **Composition**: `doctests_minimal` + `doctests_torch` + `doctests_gpytorch` + `doctests_botorch`
+- **Time**: ~15–30 seconds
+- **Use when**: Validating docstring examples
+
+#### `make doctests` (Composite)
+Runs all doctests including Docker-dependent UMBridge and markdown validation.
+- **Composition**: `doctests_markdown` + `doctests_minimal` + `doctests_torch` + `doctests_gpytorch` + `doctests_botorch` + `doctests_umbridge`
+- **Time**: Variable (Docker startup overhead)
[…]
