### 2025-12-18 c5f0526
Add alltests,yml
