+++ .github/workflows/alltests.yml
@@ -1,0 +1,55 @@
+name: All Tests
+# on:
+#   push:
+#     branches:
+#       - master 
+on: [push]
+
+jobs:
+  tests:
+    name: All Tests on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    env:
+      PYTEST_XDIST: "-n auto"
+    strategy:
+        matrix:
+            os: ["macos-latest", "ubuntu-latest", "windows-latest"]
+    steps:
+      - uses: actions/checkout@v4
+      - uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniconda-version: "latest"
+          auto-activate-base: true
+      - run: pip install -e .[test]
+      - run: pip freeze
+      - run: make doctests_minimal
+      - run: pip install -e .[test_torch] 
+      - run: make doctests_torch 
+      - run: pip install -e .[test_gpytorch]
+      - run: make doctests_gpytorch
+      - run: pip install -e .[test_botorch]
+      - run: make doctests_botorch
+      - run: pip install -e .[test_umbridge]
+      - name: run umbridge doctests on Linux only
+        shell: bash -l {0}
+        run: |
+          if [ "$RUNNER_OS" == "Linux" ]; then
+            make doctests_umbridge
+          else
+            echo "umbridge doctests only run on Linux"
+          fi
+      - run: make doctests_markdown
+      - run: make unittests
+      - name: Run booktests (parallel) on macOS/Ubuntu
+        if: runner.os != 'Windows'
+        shell: bash -l {0}
+        run: |
+          make booktests_parallel_no_docker
+
+      - name: Run booktests (sequential) on Windows
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: |
+          make booktests_no_docker
+      - run: make coverage
+
