+++ .github/workflows/alltests.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
         matrix:
-            os: ["macos-latest", "ubuntu-latest", "windows-latest"]
+            os: ["macos-15-intel", "ubuntu-latest", "windows-latest"]
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -48,6 +48,6 @@ jobs:
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
-          make booktests_no_docker
+          python -m unittest discover -p "tb_*.py" -v
       - run: make coverage
 
