### 2025-12-19 a32bec3
Fix Windows LaTeX
