### 2025-12-19 bd2a8d3
Update alltests.yml
