### 2025-12-19 09199d4
Fix latex on windows
