### 2025-12-19 8651951
One more attempt to fix Windows booktest error
