### 2025-12-19 254d864
Change Lookback to include start price (#440)
* Change Lookback to include start price

* fix spelling error

* fixed docs for lookback
