### 2025-12-20 0a9dfe8
Fix latex issues on WIndows
