+++ .github/workflows/alltests.yml
@@ -100,56 +100,26 @@ jobs:
           sudo tlmgr update --self
           sudo tlmgr install latexmk dvipng collection-fontsrecommended type1cm
           
-      - name: Install MiKTeX + deps (Windows)
+      - name: Install minimal LaTeX (Windows)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
-          $ErrorActionPreference = "Stop"
           choco install miktex -y
           choco install ghostscript -y
           choco install strawberryperl -y
-
-      - name: Configure MiKTeX (Windows)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          $ErrorActionPreference = "Stop"
-
-          # Try to find MiKTeX binaries robustly
-          $miktexBin = "C:\Program Files\MiKTeX\miktex\bin\x64"
-          if (-not (Test-Path $miktexBin)) {
-            # fallback: search typical locations
-            $miktexBin = (Get-ChildItem "C:\Program Files\MiKTeX" -Recurse -Directory -ErrorAction SilentlyContinue |
-              Where-Object { $_.FullName -match "\\miktex\\bin\\x64$" } |
-              Select-Object -First 1 -ExpandProperty FullName)
-          }
-          if (-not $miktexBin) { throw "Could not locate MiKTeX bin directory." }
-          $env:PATH = "$miktexBin;$env:PATH"
-
-          # Enable install-on-the-fly (shared/admin)
-          initexmf --admin --set-config-value "[MPM]AutoInstall=1"
-          initexmf --admin --set-config-value "[MPM]AutoAdmin=1"
-
-          # Refresh FNDB + links (often needed after CI installs)
-          initexmf --admin --update-fndb
-          initexmf --admin --mklinks --force
-
-          # Update package database + packages using the supported CLI
-          miktex --admin packages update-package-database
-          miktex --admin packages update
-
-      - name: Verify LaTeX toolchain (Windows)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          $ErrorActionPreference = "Stop"
+          # Put MiKTeX on PATH
+          $env:PATH = "C:\Program Files\MiKTeX\miktex\bin\x64;$env:PATH"
+          # Enable auto-install of missing packages (MiKTeX "just enough TeX" approach)
+          initexmf --admin --set-config-value [MPM]AutoInstall=1
+          # Refresh package database + upgrade base
+          mpm --admin --update-db
+          mpm --admin --upgrade
+          # Install a few common “baseline” packages explicitly (names are MiKTeX packages)
+          mpm --admin --install=latexmk
+          mpm --admin --install=dvipng
           latex --version
-          pdflatex --version
-          perl -v
-          gswin64c -v
           latexmk -v
           dvipng --version
-        
 
       - run: pip freeze
       - run: make doctests_minimal
@@ -184,11 +154,10 @@ jobs:
         run: |
           make booktests_parallel_no_docker
 
-      - name: Run booktests (parallel) on Windows
+      - name: Run booktests (sequential) on Windows
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
           # Execute notebook integration tests in parallel
           $env:MPLBACKEND = "Agg"
-          #python -m pytest -n auto -q test/booktests
-          #python -m pytest -q test/booktests --maxfail=1 -vv     # sequential
+          make booktests_no_docker TESTS="tb_*.py"
