### 2025-12-20 2df2d00
Take out ghostcript installation to save time
