+++ .github/workflows/alltests.yml
@@ -107,14 +107,16 @@ jobs:
           choco install miktex -y
           #choco install ghostscript -y
           choco install strawberryperl -y
-          # Put MiKTeX on PATH
+          # Put MiKTeX on PATH for all subsequent steps
+          echo "C:\Program Files\MiKTeX\miktex\bin\x64" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
+          # Refresh PATH in current session
           $env:PATH = "C:\Program Files\MiKTeX\miktex\bin\x64;$env:PATH"
           # Enable auto-install of missing packages (MiKTeX "just enough TeX" approach)
           initexmf --admin --set-config-value [MPM]AutoInstall=1
           # Refresh package database + upgrade base
           mpm --admin --update-db
           mpm --admin --upgrade
-          # Install a few common “baseline” packages explicitly (names are MiKTeX packages)
+          # Install a few common "baseline" packages explicitly (names are MiKTeX packages)
           mpm --admin --install=latexmk
           mpm --admin --install=dvipng
           latex --version
