+++ .github/workflows/alltests.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
         matrix:
-            os: ["windows-latest"m "macos-latest", "ubuntu-latest"]
+            os: ["windows-latest", "macos-latest", "ubuntu-latest"]
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
