### 2025-12-20 b4c5732
Add documentation. Remove duplicate statements.
