+++ .github/workflows/alltests.yml
@@ -38,13 +38,13 @@ jobs:
           sudo mkfile 12g /private/var/vm/tempswapfile
           sudo chmod 600 /private/var/vm/tempswapfile
 
-      - name: Add 8GB swap space (Windows)  # pagefile
+      - name: Add 12GB pagefile (Windows) (best effort)
         if: runner.os == 'Windows'
+        continue-on-error: true
         uses: al-cheb/configure-pagefile-action@v1.5
         with:
-          minimum-size: 8GB
+          minimum-size: 12GB
           maximum-size: 12GB
-          # D: is the default for this action; keep it unless you specifically want C:
           disk-root: "D:"
 
       # -----------------------------------------------------------
