### 2025-12-20 7fef80a
Speed up tb_elliptic_pde.py. Suppress matplotlib outputs in Jupyter Notebook.
