+++ test/booktests/tb_elliptic_pde.py
@@ -3,10 +3,13 @@ from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
 class NotebookTests(BaseNotebookTest):
-
-    @testbook('../../demos/elliptic-pde.ipynb', execute=True, timeout=TB_TIMEOUT)
+  
+    notebook_path = f'../../demos/elliptic-pde.ipynb'
+ 
     def test_elliptic_pde_notebook(self, tb):
-        pass
+        replacements = {"plot_convergence(execute_convergence_test","#plot_convergence(execute_convergence_test"}
+        self.change_notebook_cells(self.notebook_path, replacements)        
+        self.run_notebook(self.notebook_path, timeout=TB_TIMEOUT)
 
 if __name__ == '__main__':
     unittest.main()
