### 2025-12-20 ac2e478
Update tb_elliptic_pde.py
