+++ demos/portfolio/portfolio_allocation_to_do.md
@@ -1,6 +1,6 @@
 # QMCPy 2.2 Portfolio allocation to-do
 
-0. What are the differences between the two notebooks in `demos/portfolio`? Keep `portfolio_allocation_demo` and merge content from `archive.ipynb` into it if necessary. Afterwards, delete `archive.ipynb`. --- TODO 
+0. What are the differences between the two notebooks in `demos/portfolio`? Keep `portfolio_allocation_demo` and merge content from `archive.ipynb` into it if necessary. Afterwards, delete `archive.ipynb`. --- DONE
 1. Add Sobol and Halton samplers for weights generation in the demo --- DONE
 2. Add more text in Markdown cells before code cells to the demo --- TODO
 3. . Create a single function to generate weights for all sampler types --- DONE. See `gen_weights*()`.
