### 2025-12-21 a34b097
Speed up a bit
