### 2025-12-21 9f54b9b
Skip unnecessary installation for Windows
