### 2025-12-21 e22cc36
More caching. Pytest-xdist for all OS.
