+++ makefile
@@ -106,7 +106,7 @@ booktests_parallel_pytest: check_booktests generate_booktests clean_local_only_f
 	pip install -q -e ".[test]"  && \
 	cd test/booktests/ && \
 	PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" \
-	python -W ignore -m pytest -n auto -v tb_*.py && \
+	python -W ignore -m pytest $(PYTEST_XDIST) -v tb_*.py --cov=qmcpy --cov-append --cov-report=term --cov-report=json && \
 	cd ../.. 
 
 tests: 
