### 2025-12-21 47104f1
Comment out redundant statements
