+++ .github/workflows/alltests.yml
@@ -240,6 +240,7 @@ jobs:
       - run: make doctests_torch
       - run: make doctests_gpytorch
       - run: make doctests_botorch
+      - run: make doctests_markdown
       - name: run umbridge doctests on Linux only
         shell: bash -l {0}
         run: |
@@ -248,7 +249,6 @@ jobs:
           else
             echo "umbridge doctests only run on Linux"
           fi
-      - run: make doctests_markdown
       # -----------------------------------------------------------
       # Run unittests for Python source files
       # -----------------------------------------------------------   
@@ -256,15 +256,15 @@ jobs:
         run: |
           python -m pytest -n auto -q
       # -----------------------------------------------------------
-      # Run unittests for Jupyter notebooks (OS-specific)
+      # Run unittests for Jupyter notebooks  141.71s (0:02:21)
       # ----------------------------------------------------------- 
-      - name: Run booktests (parallel) on macOS and Ubuntu (~11 minutes)
+      - name: Run booktests (parallel) on macOS and Ubuntu (~13 mins)
         if: runner.os != 'Windows'
         shell: bash -l {0}
         run: |
           make booktests_parallel_pytest
 
-      - name: Run booktests (parallel) on Windows (~11 minutes with pytest-xdist)
+      - name: Run booktests (parallel) on Windows (~26 mins)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
