### 2025-12-21 c354905
Rm announcement. Update pahts in mkdocs.yml.
