### 2025-12-21 a4842c4
Dynamically detect number of processors (workers)
