### 2025-12-21 953e56f
Revert and use 4 processors for WIndows Booktests
