+++ scripts/compute_pytest_workers.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-"""Deprecated stub for compute_pytest_workers.
-
-Worker computation is now inlined in the `Makefile`. This stub exists only
-to avoid surprises if callers still reference the script; it prints `1`.
-"""
-import sys
-
-print(1)
-sys.exit(0)
