### 2025-12-21 abbfe6a
Resolve reviews of copilot
