### 2025-12-21 fb546f9
Delay LaTeX Installations
