### 2025-12-21 ab77cec
Fix LaTeX error involving geometry.sty not found
