### 2025-12-22 b3f1441
Release Documentation (#439)
* Created draft of RELEASE.md which will be used by future contributors to be able to publish new releases to PyPi.

* edit RELEASE.md

* Added pip install command for when  fails.

* Fix a typo

* Move RELEASE.md to docs. Add to documentation.

* +seed for reproduciblity

* Release announcement draft

* Move RELEASE.md. Add Post-Release Tasks

* Added information regarding 2FA when setting up account on PyPi in the Setup section.

* Added information regarding 2FA when setting up accounts on PyPI and Test PyPI in the Setup section.

* Initial plan (#441)

Co-autho
[…]
