+++ test/test_integrate.py
@@ -17,13 +17,13 @@ class IntegrationExampleTest(unittest.TestCase):
         dimensions = [1, 2, 3]
         true_values = [1.3803884470431430, 1.808186429263620, 2.168309102165481]
         for i in range(len(dimensions)):
-            integrand = Keister(IIDStdUniform(dimension=dimensions[i]))
+            integrand = Keister(IIDStdUniform(dimension=dimensions[i], seed=7))
             solution,data = CubMCCLT(integrand,abs_tol=abs_tol).integrate()
             self.assertTrue(abs(solution - true_values[i]) < abs_tol)
 
     def test_asian_option_multi_level(self):
         abs_tol =.01
-        integrand = FinancialOption(IIDStdUniform(64))
+        integrand = FinancialOption(IIDStdUniform(64, seed=7))
         solution,data = CubMCCLT(integrand, abs_tol).integrate()
         true_value = 1.7845
         self.assertTrue(np.isclose(solution, true_value, atol=abs_tol))
@@ -45,7 +45,7 @@ class IntegrationExampleTest(unittest.TestCase):
     
     def test_lebesgue_inf_measure(self):
         abs_tol = .1
-        true_measure = Lebesgue(Gaussian(Lattice(1)))
+        true_measure = Lebesgue(Gaussian(Lattice(1, seed=7)))
         myfunc = lambda x: np.exp(-x**2).sum(1)
         integrand = CustomFun(true_measure, myfunc)
         solution,data = CubQMCLatticeG(integrand,abs_tol=abs_tol).integrate()
@@ -54,7 +54,7 @@ class IntegrationExampleTest(unittest.TestCase):
     
     def test_lebesgue_inf_measure_2d(self):
         abs_tol = .1
-        true_measure = Lebesgue(Gaussian(Lattice(2,replications=32),mean=1,covariance=2))
+        true_measure = Lebesgue(Gaussian(Lattice(2,replications=32, seed=7),mean=1,covariance=2))
         myfunc = lambda x: np.exp(-x**2).prod(-1)
         integrand = CustomFun(true_measure,myfunc)
         solution,data = CubQMCCLT(integrand,abs_tol=abs_tol).integrate()
@@ -64,7 +64,7 @@ class IntegrationExampleTest(unittest.TestCase):
     def test_uniform_measure(self):
         """ Mathematica: Integrate[(x^3 y^3)/6, {x, 1, 3}, {y, 3, 6}] """
         abs_tol = 1
-        true_measure = Uniform(Lattice(2,replications=32), lower_bound=[1,3], upper_bound=[3,6])
+        true_measure = Uniform(Lattice(2, replications=32, seed=7), lower_bound=[1,3], upper_bound=[3,6])
         myfunc = lambda x: (x.prod(-1))**3
         integrand = CustomFun(true_measure,myfunc)
         solution,data = CubQMCCLT(integrand, abs_tol=abs_tol).integrate()
@@ -82,7 +82,7 @@ class IntegrationExampleTest(unittest.TestCase):
         true_value = 0
         for i in range(len(dimensions)):
             d = dimensions[i]
-            integrand = Linear0(DigitalNetB2(d,replications=32))
+            integrand = Linear0(DigitalNetB2(d,replications=32, seed=7))
             solution,data = CubQMCCLT(integrand, abs_tol=abs_tol).integrate()
             self.assertTrue(abs(solution - true_value) < abs_tol)
 
@@ -101,7 +101,7 @@ class IntegrationExampleTest(unittest.TestCase):
         for i in range(len(dimensions)):
             d = dimensions[i]
             integrand = CustomFun(
-                true_measure = Uniform(IIDStdUniform(d),lower_bound=0,upper_bound=1),
[…]
