+++ mkdocs.yml
@@ -61,6 +61,8 @@ nav:
       - Importance Sampling with True Measures: demos/some_true_measures.ipynb 
       - Sensitivity Analysis for the Iris Dataset: demos/iris.ipynb 
       - Ray Tracing: demos/ray_tracing.ipynb 
+  - For Developers:
+    - Release Manual: RELEASE.md
   - Printable Docs: print_page/
 
 plugins:
