### 2025-12-22 200179e
Install LaTeX after Python dependencies
