### 2025-12-23 278b3e7
Fix windows LaTeX failures
