### 2025-12-23 99d4562
Added Booktests for Windows and combined all tests in 1 workflow (#444)
* GBM code

* Fix doctest and inline documentation

* Fix typo. Add GBM to doc.

* GBM unit test

* Add _validate_input()

* Add fast unit tests for GBM

* Create gbm_demo.ipynb

* Add mean and covariance of GBM

* Speed up gen_samples using Parsl

* Interactive visualization.

* Add demo to documentation

* Minor improvements

* Revert "Minor improvements"

This reverts commit aeb3d5cb734172146c6c8f171c6dd169b4e4cd4d.

* Revert "Interactive visualization."

This reverts commit 3b3f7958c1e6f0070cf3f786310506205e70f5df.

* 
[…]
