+++ test/booktests/tb_joss2025.py
@@ -5,6 +5,10 @@ import subprocess
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
 class NotebookTests(BaseNotebookTest):
+
+    notebook_dir = '../../demos/talk_paper_demos/JOSS2025/'
+    notebook_path = f'{notebook_dir}joss2025.ipynb'
+    
     def setUp(self):
         super().setUp()  # Call parent setUp first to initialize timing attributes
         subprocess.run([
@@ -13,7 +17,7 @@ class NotebookTests(BaseNotebookTest):
             'tueplots'
         ], check=False)
         # Create the JOSS2025.outputs directory that the notebook expects
-        self.output_dir = os.path.join(os.path.dirname(__file__), 'JOSS2025.outputs')
+        self.output_dir = os.path.join(self.notebook_dir, 'JOSS2025.outputs')
         if not os.path.exists(self.output_dir):
             os.makedirs(self.output_dir, exist_ok=True)
             self._created_output_dir = True
@@ -30,15 +34,13 @@ class NotebookTests(BaseNotebookTest):
 
 
     def test_joss2025_notebook(self):
-        # locate_notebook will raise unittest.SkipTest if notebook not present
-        notebook_path, notebook_dir = self.locate_notebook('../../demos/talk_paper_demos/JOSS2025/joss2025.ipynb')
-        self.change_notebook_cells(notebook_path, 
-                                   replacements={"trials = 100": "trials = 2",
-                                                 "assert os.path.isdir(OUTDIR)":"",
-                                                 'fig.savefig':'#fig.savefig',
-                                                 'np.save':'#np.save',
-                                                 })        
-        self.run_notebook(notebook_path, timeout=TB_TIMEOUT)
+        replacements={"trials = 100": "trials = 2",
+                      "assert os.path.isdir(OUTDIR)":"",
+                      'fig.savefig':'#fig.savefig',
+                      'np.save':'#np.save',
+                     }
+        self.change_notebook_cells(self.notebook_path, replacements)        
+        self.run_notebook(self.notebook_path, timeout=TB_TIMEOUT)
 
 if __name__ == '__main__':
     unittest.main()
