### 2025-12-23 5c1d861
Introduce debug mode. Delay LaTeX installation.
