### 2025-12-23 e8b6256
Add memory
