+++ .github/workflows/alltests.yml
@@ -1,16 +1,5 @@
 name: All Tests
-on:
-  push:
-  workflow_dispatch:
-    inputs:
-      debug:
-        description: 'Enable debug/diagnostic steps (swap, free space, resource info)'
-        required: false
-        default: 'false'   # 'true' to debug, 'false' to run faster
-        type: choice
-        options:
-          - 'false'
-          - 'true'
+on: [push]
 
 jobs:
   tests:
@@ -27,48 +16,48 @@ jobs:
           auto-activate-base: true
           conda-remove-defaults: true
           use-only-tar-bz2: true
+      
       # -----------------------------------------------------------
-      # Add 12GB swap, i.e., virtual memory (OS-specific) 
+      # Clean old coverage files
+      # -----------------------------------------------------------
+      - name: Remove old coverage data
+        run: |
+          rm -f .coverage* coverage.json || true
+      # -----------------------------------------------------------
+      # Add 12GB swap, i.e., virtual memory (OS-specific)
       # -----------------------------------------------------------
       - name: Add 12GB swap (Linux)
-        if: runner.os == 'Linux'  
+        if: runner.os == 'Linux'
         run: |
           sudo fallocate -l 12G /swapfile
           sudo chmod 600 /swapfile
           sudo mkswap /swapfile
           sudo swapon /swapfile
           free -h
-
       - name: Add 12GB swap (macOS)
-        if: runner.os == 'macOS'  
+        if: runner.os == 'macOS'
         run: |
           # Create and mount a temporary swapfile
           sudo mkfile 12g /private/var/vm/tempswapfile
           sudo chmod 600 /private/var/vm/tempswapfile
-
       - name: Configure pagefile (Windows) via CIM (12GB)
-        if: runner.os == 'Windows'  
+        if: runner.os == 'Windows'
         shell: pwsh
         run: |
           $ErrorActionPreference = "Stop"
-
           $drive    = "D:"
           $initialMB = 12288
           $maxMB     = 12288
           $pagefile  = "$drive\pagefile.sys"
-
           Write-Host "Configuring pagefile: $pagefile (Initial=${initialMB}MB, Max=${maxMB}MB)"
-
           # Disable automatic management
           $cs = Get-CimInstance Win32_ComputerSystem
           if ($cs.AutomaticManagedPagefile) {
             Set-CimInstance -InputObject $cs -Property @{ AutomaticManagedPagefile = $false } | Out-Null
           }
-
           # Create or update the pagefile setting
           $escaped = $pagefile.Replace('\','\\')
           $pfs = Get-CimInstance Win32_PageFileSetting -Filter "Name='$escaped'" -ErrorAction SilentlyContinue
-
           if ($null -eq $pfs) {
             New-CimInstance -ClassName Win32_PageFileSetting -Property @{
               Name        = $pagefile
@@ -81,54 +70,47 @@ jobs:
               MaximumSize = $maxMB
             } | Out-Null
           }
-
           "=== Pagefile settings ==="
           Get-CimInstance Win32_PageFileSetting | Format-Table Name,InitialSize,MaximumSize -Auto
-
       # -----------------------------------------------------------
-      # Free disk + conda cache early (OS-specific) 
+      # Free disk + conda cache early (OS-specific)
       # -----------------------------------------------------------
       - name: Free space (Linux)
-        if: runner.os == 'Linux' 
+        if: runner.os == 'Linux'
         run: |
           sudo apt-get clean
           sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
           conda clean --all --yes
           pip cache purge || true
           df -h
-
       - name: Free space (macOS)
-        if: runner.os == 'macOS' 
[…]
