### 2025-12-23 89c929a
Correctly remove .coverage on windows
