+++ .github/workflows/alltests.yml
@@ -1,7 +1,11 @@
 name: All Tests
 on: [push]
 
+
 jobs:
+  #============================================================
+  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
+  #============================================================
   tests:
     name: All Tests on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
@@ -240,9 +244,32 @@ jobs:
       # Run unittests for Python source files
       # -----------------------------------------------------------   
       - name: Run unittests (parallel)
-        run: |
-          python -m pytest -n auto -q
+        run: make unittests
       
+      # -----------------------------------------------------------
+      # Free disk + conda cache before booktests (OS-specific)
+      # -----------------------------------------------------------
+      - name: Free space (Linux)
+        if: runner.os == 'Linux'
+        run: |
+          sudo apt-get clean
+          sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
+          conda clean --all --yes
+          pip cache purge || true
+          df -h
+      - name: Free space (macOS)
+        if: runner.os == 'macOS'
+        run: |
+          conda clean --all --yes
+          pip cache purge || true
+          df -h
+      - name: Free space (Windows)
+        if: runner.os == 'Windows'
+        run: |
+          conda clean --all --yes || true
+          pip cache purge || true
+          Get-Volume | Select-Object DriveLetter, Size, SizeRemaining | Format-Table
+
       # -----------------------------------------------------------
       # Run unittests for Jupyter notebooks (OS-specific)
       # ----------------------------------------------------------- 
@@ -270,12 +297,12 @@ jobs:
             coverage.json
           retention-days: 1
 
-  # -----------------------------------------------------------
+  # =============================================================
   # Combine coverage from all OS runners and produce final report
-  # -----------------------------------------------------------
+  # =============================================================
   combine-coverage:
     runs-on: ubuntu-latest
-    needs: tests
+    needs: tests  # Wait for 3 test jobs to complete
     steps:
       - uses: actions/checkout@v4
       
