### 2025-12-23 9a4fc71
Append coverage for Parsl BookTests. Update README.
