### 2025-12-23 5336600
Simplify README. Fix combine coverage job
