### 2025-12-23 8e678bc
More robut MikTex installation on Windows
