+++ .github/workflows/alltests.yml
@@ -201,28 +201,29 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-miktex-
           
-      - name: Install minimal LaTeX (Windows)
+      - name: Install MiKTeX (retry if needed)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
-          choco install miktex -y
-          choco install strawberryperl -y
+          choco install miktex --no-progress --limit-output -y || choco install miktex --no-progress --limit-output -y
           # Put MiKTeX on PATH for all subsequent steps
           echo "C:\Program Files\MiKTeX\miktex\bin\x64" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          # Refresh PATH in current session
           $env:PATH = "C:\Program Files\MiKTeX\miktex\bin\x64;$env:PATH"
-          # Enable auto-install of missing packages (MiKTeX "just enough TeX" approach)
-          initexmf --admin --set-config-value [MPM]AutoInstall=1
-          # Refresh package database + upgrade base
-          mpm --admin --update-db
-          mpm --admin --upgrade
-          # Install a few common "baseline" MiKTeX packages explicitly
-          mpm --admin --install=latexmk
-          mpm --admin --install=dvipng
-          mpm --admin --install=cm-super
-          mpm --admin --install=lmodern
-          mpm --admin --install=type1cm
-          mpm --admin --install=tex-gyre
+      
+      - name: Install required MiKTeX packages
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: |
+          miktexsetup finish
+          mpm --admin --verbose --require=cm,cm-super,cmex10,luxi,times,thailatex,latexmk,dvipng
+      
+      - name: Update MiKTeX font map
+        if: runner.os == 'Windows'
+        shell: pwsh
+        run: |
+          initexmf --admin --update-fndb
+          initexmf --admin --mkmaps
+          
           latex --version
           latexmk -v
           dvipng --version
