+++ .github/workflows/alltests.yml
@@ -322,6 +322,7 @@ jobs:
   combine-coverage:
     runs-on: ubuntu-latest
     needs: tests  # Wait for 3 test jobs to complete
+    if: always()  # Run even if some test jobs failed
     steps:
       - uses: actions/checkout@v4
       
@@ -338,22 +339,31 @@ jobs:
         with:
           path: coverage-data
           pattern: coverage-*
+          merge-multiple: true
+        continue-on-error: true
       
       - name: Combine coverage and generate reports
         run: |
           # List downloaded artifacts for debugging
-          ls -la coverage-data/
-          find coverage-data -type f -name '.coverage*' -o -name 'coverage.json'
+          echo "=== Coverage data directory ==="
+          ls -la coverage-data/ || echo "No coverage-data directory"
+          echo "=== Finding coverage files ==="
+          find coverage-data -type f -name '.coverage*' -o -name 'coverage.json' || echo "No coverage files found"
           # Move all coverage files to current directory
-          find coverage-data -name '.coverage*' -exec cp {} . \;
+          find coverage-data -name '.coverage*' -exec cp {} . \; 2>/dev/null || true
           # List coverage files to combine
-          ls -la .coverage*
-          # Combine coverage data from all runners
-          python -m coverage combine --keep
-          # Generate reports
-          python -m coverage report -m
-          python -m coverage xml -o coverage.xml
-          python -m coverage html -d coverage_html
+          echo "=== Coverage files to combine ==="
+          ls -la .coverage* 2>/dev/null || echo "No .coverage files to combine"
+          # Combine coverage data from all runners (only if files exist)
+          if ls .coverage* 1> /dev/null 2>&1; then
+            python -m coverage combine --keep
+            python -m coverage report -m
+            python -m coverage xml -o coverage.xml
+            python -m coverage html -d coverage_html
+          else
+            echo "No coverage data to combine"
+            exit 1
+          fi
       
       - name: Upload combined coverage HTML report
         uses: actions/upload-artifact@v4
