### 2025-12-24 8ba938d
Fix seed inconsistencies and add trailing commas per review feedback
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
