### 2025-12-24 1068b2a
Find coverage files before combining
