### 2025-12-24 22bcf8c
fix failure of coverage combined report
