### 2025-12-25 4310545
Fix coverage file detection to exclude .coveragerc
Changed line 371 to properly detect coverage data files while excluding .coveragerc configuration file, which was causing "No data to combine" errors.
