+++ .github/workflows/alltests.yml
@@ -1,393 +1,361 @@
-name: All Tests
-on: [push]
+name: Run all tests
 
+on:
+  push:
+    branches: [ master, develop ]
+  pull_request:
+    branches: [ master, develop ]
+  workflow_dispatch:
 
 jobs:
-  #============================================================
-  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
-  #============================================================
-  tests:
-    name: All Tests on ${{ matrix.os }}
+  test:
     runs-on: ${{ matrix.os }}
     strategy:
-        matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
-            os: ["windows-latest", "macos-latest", "ubuntu-latest"] 
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.8', '3.9', '3.10', '3.11']
+        exclude:
+          # Exclude some combinations to reduce CI time
+          - os: macos-latest
+            python-version: '3.8'
+          - os: macos-latest
+            python-version: '3.9'
+          - os: windows-latest
+            python-version: '3.8'
+          - os: windows-latest
+            python-version: '3.9'
+
     steps:
-      - uses: actions/checkout@v4
-      - uses: conda-incubator/setup-miniconda@v3
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
         with:
-          miniconda-version: "latest"
-          auto-activate-base: true
-          conda-remove-defaults: true
-          use-only-tar-bz2: true
-      
-      # -----------------------------------------------------------
-      # Clean old coverage files
-      # -----------------------------------------------------------
-      - name: Remove old coverage data (Unix)
-        if: runner.os != 'Windows'
+          python-version: ${{ matrix.python-version }}
+
+      - name: Get pip cache dir
+        id: pip-cache
+        shell: bash
         run: |
-          rm -f .coverage* coverage.json || true
-      - name: Remove old coverage data (Windows)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          Remove-Item -Path .coverage* -Force -ErrorAction SilentlyContinue -Confirm:$false
-          Remove-Item -Path coverage.json -Force -ErrorAction SilentlyContinue -Confirm:$false
-      # -----------------------------------------------------------
-      # Add 12GB swap, i.e., virtual memory (OS-specific)
-      # -----------------------------------------------------------
-      - name: Add 12GB swap (Linux)
-        if: runner.os == 'Linux'
-        run: |
-          sudo fallocate -l 12G /swapfile
-          sudo chmod 600 /swapfile
-          sudo mkswap /swapfile
-          sudo swapon /swapfile
-          free -h
-      - name: Add 12GB swap (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          # Create and mount a temporary swapfile
-          sudo mkfile 12g /private/var/vm/tempswapfile
-          sudo chmod 600 /private/var/vm/tempswapfile
-      - name: Configure pagefile (Windows) via CIM (12GB)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          $ErrorActionPreference = "Stop"
-          $drive    = "D:"
-          $initialMB = 12288
-          $maxMB     = 12288
-          $pagefile  = "$drive\pagefile.sys"
-          Write-Host "Configuring pagefile: $pagefile (Initial=${initialMB}MB, Max=${maxMB}MB)"
-          # Disable automatic management
-          $cs = Get-CimInstance Win32_ComputerSystem
-          if ($cs.AutomaticManagedPagefile) {
[…]
