### 2025-12-25 80de9cb
Fix coverage file detection in alltests.yml line 371
Changed line 371 from using compgen -G to using ls with grep to properly exclude .coveragerc configuration file and only match actual coverage data files (.coverage and .coverage.*).

This prevents the workflow from incorrectly detecting .coveragerc as a coverage data file.
