+++ .github/workflows/alltests.yml
@@ -2,9 +2,9 @@ name: Run all tests
 
 on:
   push:
-    branches: [ master, develop ]
+    branches: [ main, develop, booktests_choi ]
   pull_request:
-    branches: [ master, develop ]
+    branches: [ main, develop ]
   workflow_dispatch:
 
 jobs:
@@ -16,186 +16,244 @@ jobs:
         os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ['3.8', '3.9', '3.10', '3.11']
         exclude:
-          # Exclude some combinations to reduce CI time
+          # Exclude some combinations to reduce build time
           - os: macos-latest
             python-version: '3.8'
-          - os: macos-latest
-            python-version: '3.9'
           - os: windows-latest
             python-version: '3.8'
-          - os: windows-latest
-            python-version: '3.9'
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Get pip cache dir
-        id: pip-cache
-        shell: bash
-        run: |
-          echo "dir=$(pip cache dir)" >> $GITHUB_OUTPUT
-
-      - name: Cache pip packages
-        uses: actions/cache@v3
-        with:
-          path: ${{ steps.pip-cache.outputs.dir }}
-          key: ${{ runner.os }}-pip-${{ matrix.python-version }}-${{ hashFiles('**/requirements.txt', '**/setup.py') }}
-          restore-keys: |
-            ${{ runner.os }}-pip-${{ matrix.python-version }}-
-            ${{ runner.os }}-pip-
-
-      - name: Install dependencies
-        shell: bash
-        run: |
-          python -m pip install --upgrade pip
-          pip install -e .
-          pip install pytest pytest-cov coverage
-
-      - name: Run unit tests
-        shell: bash
-        run: |
-          echo "Running unit tests..."
-          python -m pytest test/fasttests --cov=qmcpy --cov-report=xml --cov-report=term -v
-
-      - name: Run integration tests
-        shell: bash
-        run: |
-          echo "Running integration tests..."
-          python -m pytest test/longtests --cov=qmcpy --cov-append --cov-report=xml --cov-report=term -v
-
-      - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
-        with:
-          file: ./coverage.xml
-          flags: unittests
-          env_vars: OS,PYTHON
-          name: codecov-umbrella
-          fail_ci_if_error: false
+    - uses: actions/checkout@v3
+    
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+    
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .[dev]
+    
+    - name: Run tests with coverage
+      run: |
+        coverage run -m pytest
+        coverage report
+    
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v3
+      with:
+        file: ./coverage.xml
+        flags: unittests
+        name: codecov-umbrella
+        fail_ci_if_error: false
 
   test-notebooks:
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
       matrix:
-        python-version: ['3.9', '3.10', '3.11']
+        notebook-group: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
 
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
+    - uses: actions/checkout@v3
+    
+    - name: Set up Python
[…]
