+++ .github/workflows/alltests.yml
@@ -23,194 +23,325 @@ jobs:
             python-version: '3.8'
 
     steps:
-    - uses: actions/checkout@v3
-    
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e .[dev]
-    
-    - name: Run tests with coverage
-      run: |
-        coverage run -m pytest
-        coverage report
-    
-    - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v3
-      with:
-        file: ./coverage.xml
-        flags: unittests
-        name: codecov-umbrella
-        fail_ci_if_error: false
-
-  test-notebooks:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        notebook-group: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
-
-    steps:
-    - uses: actions/checkout@v3
-    
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.11'
-    
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -e .[dev]
-        pip install jupyter nbconvert
-    
-    - name: Get notebook list
-      id: notebooks
-      run: |
-        # Get all Jupyter notebooks, excluding certain directories
-        notebooks=$(find . -name "*.ipynb" -not -path "*/.*" -not -path "*/_*" | sort)
-        echo "All notebooks:"
-        echo "$notebooks"
-        
-        # Count total notebooks
-        total=$(echo "$notebooks" | wc -l)
-        echo "Total notebooks: $total"
-        
-        # Calculate notebooks per group (ceiling division)
-        per_group=$(( ($total + 9) / 10 ))
-        echo "Notebooks per group: $per_group"
-        
-        # Get notebooks for this group
-        start=$(( (${{ matrix.notebook-group }} - 1) * per_group + 1 ))
-        end=$(( ${{ matrix.notebook-group }} * per_group ))
-        
-        group_notebooks=$(echo "$notebooks" | sed -n "${start},${end}p")
-        echo "Group ${{ matrix.notebook-group }} notebooks (lines $start-$end):"
-        echo "$group_notebooks"
-        
-        # Save to file for next step
-        echo "$group_notebooks" > notebooks_to_test.txt
-    
-    - name: Test notebooks
-      run: |
-        echo "Testing notebooks in group ${{ matrix.notebook-group }}:"
-        cat notebooks_to_test.txt
-        
-        failed=0
-        while IFS= read -r notebook; do
-          if [ -n "$notebook" ]; then
-            echo "Testing: $notebook"
-            if jupyter nbconvert --to notebook --execute --inplace "$notebook"; then
-              echo "✓ Passed: $notebook"
-            else
-              echo "✗ Failed: $notebook"
-              failed=$((failed + 1))
-            fi
+      - uses: actions/checkout@v4
+      - uses: conda-incubator/setup-miniconda@v3
+        with:
+          miniconda-version: "latest"
+          auto-activate-base: true
+          conda-remove-defaults: true
+          use-only-tar-bz2: true
+      
+      # -----------------------------------------------------------
+      # Clean old coverage files
+      # -----------------------------------------------------------
+      - name: Remove old coverage data (Unix)
+        if: runner.os != 'Windows'
+        run: |
+          rm -f .coverage* coverage.json || true
+      - name: Remove old coverage data (Windows)
+        if: runner.os == 'Windows'
[…]
