+++ .github/workflows/alltests.yml
@@ -1,36 +1,32 @@
-name: Run all tests
-
-on:
-  push:
-    branches: [ main, develop, booktests_choi ]
-  pull_request:
-    branches: [ main, develop ]
-  workflow_dispatch:
+name: All Tests
+on: [push]
 
 jobs:
-  test:
+  #============================================================
+  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
+  #============================================================
+  tests:
+    name: All Tests on ${{ matrix.os }}
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.8', '3.9', '3.10', '3.11']
-        exclude:
-          # Exclude some combinations to reduce build time
-          - os: macos-latest
-            python-version: '3.8'
-          - os: windows-latest
-            python-version: '3.8'
+      matrix: # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
+        os: ["windows-latest", "macos-latest", "ubuntu-latest"]
+
+    # Ensure each matrix runner writes a distinct coverage data file (.coverage.<os>)
+    # so we can safely combine them later.
+    env:
+      COVERAGE_FILE: .coverage.${{ matrix.os }}
 
     steps:
       - uses: actions/checkout@v4
+
       - uses: conda-incubator/setup-miniconda@v3
         with:
           miniconda-version: "latest"
           auto-activate-base: true
           conda-remove-defaults: true
           use-only-tar-bz2: true
-      
+
       # -----------------------------------------------------------
       # Clean old coverage files
       # -----------------------------------------------------------
@@ -38,12 +34,14 @@ jobs:
         if: runner.os != 'Windows'
         run: |
           rm -f .coverage* coverage.json || true
+
       - name: Remove old coverage data (Windows)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
           Remove-Item -Path .coverage* -Force -ErrorAction SilentlyContinue -Confirm:$false
           Remove-Item -Path coverage.json -Force -ErrorAction SilentlyContinue -Confirm:$false
+
       # -----------------------------------------------------------
       # Add 12GB swap, i.e., virtual memory (OS-specific)
       # -----------------------------------------------------------
@@ -55,30 +53,37 @@ jobs:
           sudo mkswap /swapfile
           sudo swapon /swapfile
           free -h
+
       - name: Add 12GB swap (macOS)
         if: runner.os == 'macOS'
         run: |
           # Create and mount a temporary swapfile
           sudo mkfile 12g /private/var/vm/tempswapfile
           sudo chmod 600 /private/var/vm/tempswapfile
+
       - name: Configure pagefile (Windows) via CIM (12GB)
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
           $ErrorActionPreference = "Stop"
-          $drive    = "D:"
+
+          $drive     = "D:"
           $initialMB = 12288
           $maxMB     = 12288
           $pagefile  = "$drive\pagefile.sys"
+
           Write-Host "Configuring pagefile: $pagefile (Initial=${initialMB}MB, Max=${maxMB}MB)"
+
           # Disable automatic management
           $cs = Get-CimInstance Win32_ComputerSystem
           if ($cs.AutomaticManagedPagefile) {
             Set-CimInstance -InputObject $cs -Property @{ AutomaticManagedPagefile = $false } | Out-Null
           }
+
           # Create or update the pagefile setting
-          $escaped = $pagefile.Replace('\','\\')
+          $escaped = $pagefile.Replace('','\')
[…]
