### 2025-12-25 e6c75af
Remove combining report
