+++ .github/workflows/alltests.yml
@@ -352,70 +352,5 @@ jobs:
           if-no-files-found: warn
           retention-days: 1
 
-  # =============================================================
-  # Combine coverage from all OS runners and produce final report
-  # =============================================================
-  combine-coverage:
-    runs-on: ubuntu-latest
-    needs: tests # Wait for 3 test jobs to complete
-    if: always() # Run even if some test jobs failed
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.12'
-
-      - name: Install coverage
-        run: python -m pip install coverage
-
-      - name: Download all coverage artifacts (into separate folders)
-        uses: actions/download-artifact@v4
-        continue-on-error: true
-        with:
-          path: coverage-data
-          pattern: coverage-*
-
-      # FIX: Combine ONLY real coverage data files (.coverage.*), not coverage.json.
-      # Also avoid overwriting by ensuring files are already unique from the matrix job.
-      - name: Combine coverage and generate reports
-        shell: bash
-        run: |
-          set -euo pipefail
-          shopt -s globstar nullglob
-
-          echo "=== Downloaded artifacts ==="
-          ls -laR coverage-data/ || true
-
-          files=(coverage-data/**/.coverage.*)
-          if (( ${#files[@]} == 0 )); then
-            echo "No .coverage.* files found; skipping combine."
-            exit 0
-          fi
-
-          cp "${files[@]}" .
-          echo "=== Coverage files copied to workspace ==="
-          ls -la .coverage.* || true
-
-          python -m coverage combine --keep
-          python -m coverage report -m
-          python -m coverage xml -o coverage.xml
-          python -m coverage html -d coverage_html
-
-      - name: Upload combined coverage HTML report
-        if: ${{ hashFiles('coverage_html/**') != '' }}
-        uses: actions/upload-artifact@v4
-        with:
-          name: coverage-report-html
-          path: coverage_html
-
-      - name: Upload coverage XML for external tools
-        if: ${{ hashFiles('coverage.xml') != '' }}
-        uses: actions/upload-artifact@v4
-        with:
-          name: coverage-xml
-          path: coverage.xml
-
+  
 # For success reports, see https://github.com/QMCSoftware/QMCSoftware/actions?query=is%3Asuccess+workflow%3AAll+Tests
