### 2025-12-25 8c84e9a
Summarize OS vs coverage
