### 2025-12-25 d7729bc
Display final reports for Mac and Linux
