### 2025-12-26 e0a71dd
Remove coverage reports in CI
