### 2025-12-26 bbec8f5
Speedup
