### 2025-12-26 eb002b0
Revert "Speedup" commit from booktests_choi branch (#447)
* Initial plan

* Undo Speedup commit from booktests_choi

This commit represents the state of tb_vectorized_qmc_bayes.py before the
'Speedup' commit (bbec8f5c). When merged into booktests_choi, this will
effectively revert the Speedup changes.

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
