### 2025-12-26 d502201
Remove and ignore coverage files. Update README.
