+++ test/README.md
@@ -20,6 +20,15 @@ This document describes the available test targets in the Makefile for QMCSoftwa
 
 ## Detailed Descriptions
 
+## Scope
+
+This short guide highlights four practical areas:
+- Local tests in the `test/` directory and the Makefile targets that run them (for example `make unittests`, `make booktests_no_docker`, and `make doctests_no_docker`).
+- Local coverage targets in the Makefile (for example `make coverage` and `make delcoverage`) and how to collect coverage locally using `--cov-append` or `coverage run --append`.
+- Remote CI: the GitHub Actions workflow at `.github/workflows/alltests.yml` (referred to here as `alltests.yml`) which runs matrix jobs and invokes the Makefile targets for full validation.
+- Remote coverage publishing to Codecov and how to add a Codecov badge to the project `README.md`.
+
+
 ### Core Test Targets (Recommended)
 
 #### `make tests_fast` ⚡ **RECOMMENDED**
@@ -261,7 +270,7 @@ make tests_no_docker        # Sequential, safe (60–120s)
 ## Coverage Report Strategy
 
 ### Overview
-QMCSoftware uses a **multi-platform unified coverage report** approach in GitHub Actions CI. Coverage data from all test types (doctests, unittests, booktests) running on all platforms (Ubuntu, macOS, Windows) is combined into a single comprehensive report.
+QMCSoftware uses a **multi-platform unified coverage report** approach in GitHub Actions CI. Coverage data from all test types (doctests, unittests, booktests) running on all platforms (Ubuntu, macOS, Windows) is combined into a single  coverage percentage.
 
 ### Architecture
 
@@ -275,18 +284,6 @@ QMCSoftware uses a **multi-platform unified coverage report** approach in GitHub
 │  4. Run booktests (with --cov-append)                   │
 │  5. Upload .coverage & coverage.json as artifacts       │
 └─────────────────────────────────────────────────────────┘
-                          ↓
-┌─────────────────────────────────────────────────────────┐
-│  Final Job: combine-coverage (ubuntu-latest)            │
-├─────────────────────────────────────────────────────────┤
-│  1. Download all coverage artifacts (3 OS runners)      │
-│  2. Combine with: coverage combine                      │
-│  3. Generate reports:                                   │
-│     - coverage report -m (terminal)                     │
-│     - coverage xml -o coverage.xml                      │
-│     - coverage html -d coverage_html                    │
-│  4. Upload final HTML & XML as artifacts                │
-└─────────────────────────────────────────────────────────┘
 ```
 
 ### Key Syntax & Configuration
@@ -297,7 +294,6 @@ Cross-platform coverage combining requires relative paths to handle different OS
 **`.coveragerc` settings:**
 - `relative_files = True` – Store paths relative to project root
 - `[paths]` section maps all OS path variants to common source location
-- Enables successful `coverage combine` across Windows, macOS, and Ubuntu runners
 
 #### 2. Makefile Test Targets (Coverage Append Mode)
 All test targets use `--cov-append` (pytest) or `coverage run --append` to accumulate coverage within each OS runner:
@@ -321,8 +317,14 @@ All test targets use `--cov-append` (pytest) or `coverage run --append` to accum
 #### 3. GitHub Actions Workflow
 
 - **Clean coverage at start of each matrix job:**
-- **Upload coverage artifacts after all tests:**
-- **Combine in separate job:**
+
+## CI & Coverage (summary)
+
[…]
