+++ test/booktests/tb_elliptic_pde.py
@@ -6,7 +6,7 @@ class NotebookTests(BaseNotebookTest):
     notebook_path = f'../../demos/elliptic-pde.ipynb'
  
     def test_elliptic_pde_notebook(self):
-        replacements = {"plot_convergence(execute_convergence_test","#plot_convergence(execute_convergence_test"}
+        replacements = {"plot_convergence(execute_convergence_test": "#plot_convergence(execute_convergence_test",}
         self.change_notebook_cells(self.notebook_path, replacements)        
         self.run_notebook(self.notebook_path, timeout=TB_TIMEOUT)
 
