### 2025-12-27 aaf4015
Skip slow notebook tests on windows
