+++ test/booktests/tb_joss2025.py
@@ -4,6 +4,7 @@ import shutil
 import subprocess
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     notebook_dir = '../../demos/talk_paper_demos/JOSS2025/'
