+++ test/booktests/tb_asian_option_mlqmc.py
@@ -2,6 +2,7 @@ import unittest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/asian-option-mlqmc.ipynb', execute=True, timeout=TB_TIMEOUT)
