+++ test/booktests/tb_vectorized_qmc.py
@@ -3,6 +3,7 @@ import subprocess
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
