+++ test/booktests/tb_vectorized_qmc_bayes.py
@@ -1,6 +1,8 @@
 import unittest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
+
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/vectorized_qmc_bayes.ipynb', execute=True, timeout=TB_TIMEOUT)
