+++ test/booktests/tb_gaussian_diagnostics_demo.py
@@ -4,6 +4,7 @@ import os
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     def setUp(self):
