+++ test/booktests/tb_gaussian_diagnostics_demo.py
@@ -1,4 +1,4 @@
-import unittest
+import unittest, pytest
 import subprocess
 import os
 from testbook import testbook
