+++ test/booktests/tb_joss2025.py
@@ -1,4 +1,4 @@
-import unittest
+import unittest, pytest
 import os
 import shutil
 import subprocess
