+++ test/booktests/tb_vectorized_qmc.py
@@ -1,4 +1,4 @@
-import unittest
+import unittest, pytest
 import subprocess
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
