### 2025-12-27 5320a9b
Add seed values for reproducible test results
