+++ test/booktests/tb_iris.py
@@ -1,7 +1,8 @@
-import unittest
+import unittest, pytest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
+@pytest.mark.slow 
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/iris.ipynb', execute=True, timeout=TB_TIMEOUT) 
