### 2025-12-27 6f4cf4c
Add two targets in Makefile for reuse in alltests.yml.
