### 2025-12-27 a0ab494
Revert "Add two targets in Makefile for reuse in alltests.yml."
This reverts commit 6f4cf4c1013189903800b4bcf3a93cbb707f2da3.
