+++ makefile
@@ -106,13 +106,6 @@ booktests_parallel_pytest: check_booktests generate_booktests clean_local_only_f
 	python -W ignore -m pytest $(PYTEST_XDIST) -v tb_*.py --cov=qmcpy --cov-append --cov-report=term --cov-report=json && \
 	cd ../.. 
 
-# Windows-compatible parallel booktests using pytest-xdist instead of Parsl
-booktests_parallel_pytest: check_booktests generate_booktests clean_local_only_files
-	cd test/booktests/ && \
-	PYTHONWARNINGS="ignore::UserWarning,ignore::DeprecationWarning,ignore::FutureWarning,ignore::ImportWarning" \
-	python -W ignore -m pytest $(PYTEST_XDIST) -v tb_*.py --cov=qmcpy --cov-append --cov-report=term --cov-report=json && \
-	cd ../.. 
-
 tests: 
 	set -e && $(MAKE) doctests && $(MAKE) unittests && $(MAKE) coverage
 
