### 2025-12-28 b85707a
Increase coverage to 100% for qmcpy codebase
