+++ test/booktests/tb_vectorized_qmc_bayes.py
@@ -2,7 +2,7 @@ import unittest, pytest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
-@pytest.mark.slow 
+@pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/vectorized_qmc_bayes.ipynb', execute=True, timeout=TB_TIMEOUT)
