+++ test/booktests/tb_iris.py
@@ -2,7 +2,7 @@ import unittest, pytest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
-@pytest.mark.slow 
+@pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/iris.ipynb', execute=True, timeout=TB_TIMEOUT) 
