+++ test/booktests/tb_asian_option_mlqmc.py
@@ -2,7 +2,7 @@ import unittest, pytest
 from testbook import testbook
 from __init__ import TB_TIMEOUT, BaseNotebookTest
 
-@pytest.mark.slow 
+@pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
 
     @testbook('../../demos/asian-option-mlqmc.ipynb', execute=True, timeout=TB_TIMEOUT)
