+++ mkdocs.yml
@@ -63,6 +63,7 @@ nav:
       - Ray Tracing: demos/ray_tracing.ipynb 
   - For Developers:
     - Release Manual: RELEASE.md
+    - Testing Guidelines: test/README.md
   - Printable Docs: print_page/
 
 plugins:
