### 2025-12-29 96c2726
backtesting with replications + one more todo item
