+++ demos/portfolio/portfolio_allocation_to_do.md
@@ -13,7 +13,8 @@
 10. Save figures as .png in `images` --- DONE
 11. Use same colors for same samplers across all visualizations --- DONE
 12. Use white background for all visualizations --- DONE
-13. Stopping checking in .png or binary files into repository --- TODO
+13. Stopping checking in .png or binary files into repository --- DONE
 14. Clean up demo --- TODO
 15. Add references to the demo --- TODO
 16. Update poster in Overleaf --- TODO
+17. Implement out-of-sample backtesting 
