### 2025-12-29 #9705 @editorialbot
[REVIEW]: QMCPy: An Open-Source Python Framework for (Quasi-)Monte Carlo Algorithms
**Submitting author:** @alegresor (<a href="http://orcid.org/0000-0003-1004-4632">Aleksei Sorokin</a>)
**Repository:** https://github.com/QMCSoftware/QMCSoftware
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.2
