### 2025-12-29 #9705 @lockwo
@matt-graham, @tupui Thanks for agreeing to review this submission!

As you can see above, you each should use the command @editorialbot generate my checklist to create your review checklist. @editorialbot commands need to be the first thing in a new comment. [I see @tupui has already done this, thanks!].

As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/9705 so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 4-5 weeks [this week is a holiday week in the US, so I added a week to the goal timelines]. Please let me know if either of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Note that the main/top issue text also contains information, such as how to generate your checklists.
