### 2025-12-29 #9705 @tupui
Hi all!

A very straightforward review from my side. I know the software and I have been playing with it already when I was implementing the QMC module in SciPy itself. This is a very good addition to JOSS.

The one thing I am missing from the paper is a bit more comparison with SciPy feature set itself. I think this is important because that would highlight even more the difference in terms of audience specifically. Since we have a JOSS article for that work too, I think it's very relevant to cite and discuss.

Otherwise it's a warm recommendation for publication from me. Great job team 👏
