### 2025-12-30 6e706fd
Enhance code and doc readability
