+++ qmcpy/discrete_distribution/abstract_discrete_distribution.py
@@ -1,35 +1,57 @@
-from ..util import ParameterError, ParameterWarning, MethodImplementationError, _univ_repr, DimensionError
+from ..util import (
+    ParameterError,
+    ParameterWarning,
+    MethodImplementationError,
+    _univ_repr,
+    DimensionError,
+)
 import warnings
 import numpy as np
-from typing import Union 
+from typing import Union
+
 
 class AbstractDiscreteDistribution(object):
 
     def __init__(self, dimension, replications, seed, d_limit, n_limit):
-        self.mimics = 'StdUniform'
-        if not hasattr(self,'parameters'):
+        self.mimics = "StdUniform"
+        if not hasattr(self, "parameters"):
             self.parameters = []
         self.d_limit = d_limit
-        self.n_limit = n_limit 
-        if not (np.isscalar(self.d_limit) and self.d_limit>0 and np.isscalar(self.n_limit) and self.n_limit>0):
+        self.n_limit = n_limit
+        if not (
+            np.isscalar(self.d_limit)
+            and self.d_limit > 0
+            and np.isscalar(self.n_limit)
+            and self.n_limit > 0
+        ):
             raise ParameterError("d_limit and n_limit must be greater than 0")
-        self.d_limit = np.inf if self.d_limit==np.inf else int(self.d_limit)
-        self.n_limit = np.inf if self.n_limit==np.inf else int(self.n_limit)
-        self.no_replications = replications is None 
+        self.d_limit = np.inf if self.d_limit == np.inf else int(self.d_limit)
+        self.n_limit = np.inf if self.n_limit == np.inf else int(self.n_limit)
+        self.no_replications = replications is None
         self.replications = 1 if self.no_replications else int(replications)
-        if self.replications<0:
+        if self.replications < 0:
             raise ParameterError("replications must be None or a postive int")
-        if isinstance(dimension,list) or isinstance(dimension,tuple) or isinstance(dimension,np.ndarray):
-            self.dvec = np.array(dimension,dtype=int)
+        if (
+            isinstance(dimension, list)
+            or isinstance(dimension, tuple)
+            or isinstance(dimension, np.ndarray)
+        ):
+            self.dvec = np.array(dimension, dtype=int)
             self.d = len(self.dvec)
-            if not (self.dvec.ndim==1 and len(np.unique(self.dvec))==self.d):
+            if not (self.dvec.ndim == 1 and len(np.unique(self.dvec)) == self.d):
                 raise ParameterError("dimension must be a 1d array of unique values")
         else:
             self.d = int(dimension)
             self.dvec = np.arange(self.d)
-        if any(self.dvec>self.d_limit):
-            raise ParameterError('dimension greater than dimension limit %d'%self.d_limit)
-        self._base_seed = seed if isinstance(seed,np.random.SeedSequence) else np.random.SeedSequence(seed)
+        if any(self.dvec > self.d_limit):
+            raise ParameterError(
+                "dimension greater than dimension limit %d" % self.d_limit
+            )
+        self._base_seed = (
+            seed
+            if isinstance(seed, np.random.SeedSequence)
+            else np.random.SeedSequence(seed)
+        )
         self.entropy = self._base_seed.entropy
         self.spawn_key = self._base_seed.spawn_key
         self.rng = np.random.Generator(np.random.SFC64(self._base_seed))
@@ -44,49 +66,66 @@ class AbstractDiscreteDistribution(object):
             n (Union[None,int]): Number of points to generate.
             n_min (Union[None,int]): Starting index of sequence.
             n_max (Union[None,int]): Final index of sequence.
-            return_binary (bool): Only used for `DigitalNetB2`.  
[…]
