+++ demos/GBM/gbm_code/data_util.py
@@ -5,30 +5,40 @@ import quantlib_util as qlu
 import qmcpy_util as qpu
 import config as cf
 
-def add_theoretical_results(results_data: list, theoretical_mean: float, theoretical_std: float) -> None:
+
+def add_theoretical_results(
+    results_data: list, theoretical_mean: float, theoretical_std: float
+) -> None:
     """
     Add theoretical benchmark values to results data.
-    
+
     Args:
         results_data: List to append theoretical results to
         theoretical_mean: Theoretical expected value E[S_T]
         theoretical_std: Theoretical standard deviation of S_T
     """
-    results_data.append({
-        'Method': 'Theoretical',
-        'Sampler': '-',
-        'Mean': theoretical_mean,
-        'Std Dev': theoretical_std,
-        'Mean Absolute Error': 0,
-        'Std Dev Error': 0,
-    })
+    results_data.append(
+        {
+            "Method": "Theoretical",
+            "Sampler": "-",
+            "Mean": theoretical_mean,
+            "Std Dev": theoretical_std,
+            "Mean Absolute Error": 0,
+            "Std Dev Error": 0,
+        }
+    )
 
-def add_quantlib_results(results_data: list, sampler_type: str, 
-                         quantlib_final: npt.NDArray[np.floating], 
-                         theoretical_mean: float, theoretical_std: float) -> None:
+
+def add_quantlib_results(
+    results_data: list,
+    sampler_type: str,
+    quantlib_final: npt.NDArray[np.floating],
+    theoretical_mean: float,
+    theoretical_std: float,
+) -> None:
     """
     Compute and add QuantLib simulation results to results data.
-    
+
     Args:
         results_data: List to append results to
         sampler_type: Name of the sampler used
@@ -38,21 +48,29 @@ def add_quantlib_results(results_data: list, sampler_type: str,
     """
     ql_emp_mean = np.mean(quantlib_final)
     ql_emp_std = np.std(quantlib_final, ddof=1)
-    results_data.append({
-        'Method': 'QuantLib',
-        'Sampler': sampler_type,
-        'Mean': ql_emp_mean,
-        'Std Dev': ql_emp_std,
-        'Mean Absolute Error': abs(ql_emp_mean - theoretical_mean),
-        'Std Dev Error': abs(ql_emp_std - theoretical_std),
-    })
+    results_data.append(
+        {
+            "Method": "QuantLib",
+            "Sampler": sampler_type,
+            "Mean": ql_emp_mean,
+            "Std Dev": ql_emp_std,
+            "Mean Absolute Error": abs(ql_emp_mean - theoretical_mean),
+            "Std Dev Error": abs(ql_emp_std - theoretical_std),
+        }
+    )
 
-def add_qmcpy_results(results_data: list, sampler_type: str, 
-                      qmcpy_final: npt.NDArray[np.floating], qp_emp_mean: float, 
-                      theoretical_mean: float, theoretical_std: float) -> None:
+
+def add_qmcpy_results(
+    results_data: list,
+    sampler_type: str,
+    qmcpy_final: npt.NDArray[np.floating],
+    qp_emp_mean: float,
+    theoretical_mean: float,
+    theoretical_std: float,
+) -> None:
     """
     Compute and add QMCPy simulation results to results data.
-    
+
     Args:
         results_data: List to append results to
         sampler_type: Name of the sampler used
@@ -62,21 +80,29 @@ def add_qmcpy_results(results_data: list, sampler_type: str,
         theoretical_std: Theoretical standard deviation for error computation
     """
     qp_emp_std = np.std(qmcpy_final, ddof=1)
-    results_data.append({
-        'Method': 'QMCPy',
[…]
