+++ test/test_keister.py
@@ -57,7 +57,7 @@ class KeisterProblemTest(unittest.TestCase):
     
     def test_cub_qmc_clt_dnb2(self):
         atol = 1e-4
-        integrand = Keister(DigitalNet(self.d,seed=7,replications=32))
+        integrand = Keister(DigitalNetB2(self.d,seed=7,replications=32))
         true_value = integrand.get_exact_value(self.d)
         solution,data = CubQMCCLT(integrand, abs_tol=atol).integrate()
         self.assertTrue(abs(solution-true_value) < atol)
