+++ qmcpy/discrete_distribution/__init__.py
@@ -3,10 +3,12 @@ from .iid_std_uniform import IIDStdUniform
 from .lattice import Lattice
 from .digital_net_b2 import DigitalNetB2
 from .halton import Halton
+from .digital_net_any_bases import DigitalNetAnyBases
 
 DiscreteDistribution = AbstractDiscreteDistribution
 _DiscreteDistribution = AbstractDiscreteDistribution
 Sobol = DigitalNetB2
 DigitalNet = DigitalNetB2
 Net = DigitalNetB2
+NetB2 = DigitalNetB2
 
