+++ qmcpy/discrete_distribution/digital_net_any_bases/digital_net_any_bases.py
@@ -1,0 +1,4 @@
+from ..abstract_discrete_distribution import AbstractLDDiscreteDistribution
+
+class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
+    pass
