+++ qmcpy/discrete_distribution/digital_net_any_bases/__init__.py
@@ -1,1 +1,0 @@
-from .digital_net_any_bases import DigitalNetAnyBases
