+++ qmcpy/discrete_distribution/__init__.py
@@ -2,8 +2,7 @@ from .abstract_discrete_distribution import AbstractDiscreteDistribution
 from .iid_std_uniform import IIDStdUniform
 from .lattice import Lattice
 from .digital_net_b2 import DigitalNetB2
-from .halton import Halton
-from .digital_net_any_bases import DigitalNetAnyBases
+from .halton import Halton,DigitalNetAnyBases
 
 DiscreteDistribution = AbstractDiscreteDistribution
 _DiscreteDistribution = AbstractDiscreteDistribution
