+++ qmcpy/discrete_distribution/halton.py
@@ -7,7 +7,180 @@ from math import *
 from copy import deepcopy
 
 
-class Halton(AbstractLDDiscreteDistribution):
+class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
+    r"""
+    Low discrpancy digital net with arbitrary bases for each dimension. 
+    
+    Note: 
+        - Digital net samples sizes should be products of powers of bases, 
+        i.e., a digital net with bases $(b_1,\dots,b_d)$ 
+        will prefer sample sizes $n = b_1^{p_1} \cdots b_d^{p_d}$ for some $p_1,\dots,p_d \in \mathbb{N}_0$.
+        - The first point of an unrandomized digital net is the origin. 
+        - The construction of higher order digital nets requires the same base for each dimension. 
+        To construct higher order digital nets, either: 
+        
+            - Pass in `generating_matrices` *without* interlacing and supply `alpha>1` to apply interlacing, or 
+            - Pass in `generating_matrices` *with* interlacing and set `alpha=1` to avoid additional interlacing. 
+            
+            i.e. do *not* pass in interlaced `generating_matrices` and set `alpha>1`, this will apply additional interlacing. 
+    
+    Examples:
+        >>> pass
+    
+    **References:**
+
+    1.  Dick, Josef, and Friedrich Pillichshammer.  
+        Digital nets and sequences: discrepancy theory and quasi–Monte Carlo integration.  
+        Cambridge University Press, 2010.
+    
+    2.  Sorokin, Aleksei.  
+        "QMCPy: A Python Software for Randomized Low-Discrepancy Sequences, Quasi-Monte Carlo, and Fast Kernel Methods"  
+        arXiv preprint arXiv:2502.14256 (2025).
+    """
+    
+    DEFAULT_GENERATING_MATRICES = None 
+    
+    def __init__(self,
+                 dimension = 1,
+                 replications = None,
+                 seed = None, 
+                 randomize = 'LMS DP',
+                 bases_generating_matrices = None,
+                 t = 63,
+                 alpha = 1,
+                 n_lim = 2**32):
+        r"""
+        Args:
+            dimension (Union[int,np.ndarray]): Dimension of the generator.
+
+                - If an `int` is passed in, use generating vector components at indices 0,...,`dimension`-1.
+                - If an `np.ndarray` is passed in, use generating vector components at these indices.
+            
+            replications (int): Number of independent randomizations of a pointset.
+            seed (Union[None,int,np.random.SeedSeq): Seed the random number generator for reproducibility.
+            randomize (str): Options are
+                
+                - `'LMS DP'`: Linear matrix scramble with digital permutation.
+                - `'LMS DS'`: Linear matrix scramble with digital shift.
+                - `'LMS'`: Linear matrix scramble only.
+                - `'DP'`: Digital permutation scramble only.
+                - `'DS'`: Digital shift only.
[…]
