### 2025-12-31 fd1af55
recover Halton results with GDN
