### 2025-12-31 c225718
before accounting for alpha>1
