### 2025-12-31 b73a98f
check constant base for alpha>1
