### 2025-12-31 49d49ad
restructure API for DN any base
