+++ qmcpy/discrete_distribution/digital_net_any_base/digital_net_any_base.py
@@ -1,6 +1,6 @@
 import warnings
-from .abstract_discrete_distribution import AbstractLDDiscreteDistribution
-from ..util import ParameterError,ParameterWarning
+from ..abstract_discrete_distribution import AbstractLDDiscreteDistribution
+from ...util import ParameterError,ParameterWarning
 import qmctoolscl
 import numpy as np
 from math import *
@@ -119,7 +119,7 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
         assert self.randomize in ["LMS DP","LMS DS","LMS","DP","DS","NUS","QRNG","FALSE"]
         if self.randomize=="QRNG":
             assert self.type_bases_generating_matrices=="HALTON", "QRNG randomization is only applicable for the Halton generator."
-            from ._c_lib import _load_c_lib
+            from .._c_lib import _load_c_lib
             assert self.replications==1, "QRNG requires replications=1"
             self.randu_d_32 = self.rng.uniform(size=(self.d,32))
             _c_lib = _load_c_lib()
@@ -238,184 +238,3 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             randomize = self.randomize,
             t = self.input_t,
             n_lim = self.n_limit)
-
-class Halton(DigitalNetAnyBases):
-    r"""
-    Low discrepancy Halton points.
-
-    Note:
-        - The first point of an unrandomized Halton sequence is the origin.
-        - QRNG does *not* support multiple replications (independent randomizations).
-    
-    Examples:
-        >>> discrete_distrib = Halton(2,seed=7)
-        >>> discrete_distrib(4)
-        array([[0.83790457, 0.89981478],
-               [0.00986102, 0.4610941 ],
-               [0.62236343, 0.02796307],
-               [0.29427505, 0.79909098]])
-        >>> discrete_distrib
-        Halton (AbstractLDDiscreteDistribution)
-            d               2^(1)
-            replications    1
-            randomize       LMS DP
-            t               63
-            n_limit         2^(32)
-            entropy         7
-        
-        Replications of independent randomizations 
-
-        >>> x = Halton(3,seed=7,replications=2)(4)
-        >>> x.shape
-        (2, 4, 3)
-        >>> x
-        array([[[0.70988236, 0.18180876, 0.54073621],
-                [0.38178158, 0.61168824, 0.64684354],
-                [0.98597752, 0.70650871, 0.31479029],
-                [0.15795399, 0.28162992, 0.98945647]],
-        <BLANKLINE>
-               [[0.620398  , 0.57025403, 0.46336542],
-                [0.44021889, 0.69926312, 0.60133428],
-                [0.89132308, 0.12030255, 0.35715804],
-                [0.04025218, 0.44304244, 0.10724799]]])
-
-        Unrandomized Halton 
-
-        >>> Halton(2,randomize="FALSE",seed=7)(4,warn=False)
-        array([[0.        , 0.        ],
-               [0.5       , 0.33333333],
-               [0.25      , 0.66666667],
-               [0.75      , 0.11111111]])
-        
-        All randomizations 
-
-        >>> Halton(2,randomize="LMS DP",seed=7)(4)
-        array([[0.83790457, 0.89981478],
-               [0.00986102, 0.4610941 ],
-               [0.62236343, 0.02796307],
-               [0.29427505, 0.79909098]])
-        >>> Halton(2,randomize="LMS DS",seed=7)(4)
[…]
