+++ qmcpy/discrete_distribution/digital_net_any_base/halton.py
@@ -1,0 +1,183 @@
+from .digital_net_any_base import DigitalNetAnyBases
+
+
+class Halton(DigitalNetAnyBases):
+    r"""
+    Low discrepancy Halton points.
+
+    Note:
+        - The first point of an unrandomized Halton sequence is the origin.
+        - QRNG does *not* support multiple replications (independent randomizations).
+    
+    Examples:
+        >>> discrete_distrib = Halton(2,seed=7)
+        >>> discrete_distrib(4)
+        array([[0.83790457, 0.89981478],
+               [0.00986102, 0.4610941 ],
+               [0.62236343, 0.02796307],
+               [0.29427505, 0.79909098]])
+        >>> discrete_distrib
+        Halton (AbstractLDDiscreteDistribution)
+            d               2^(1)
+            replications    1
+            randomize       LMS DP
+            t               63
+            n_limit         2^(32)
+            entropy         7
+        
+        Replications of independent randomizations 
+
+        >>> x = Halton(3,seed=7,replications=2)(4)
+        >>> x.shape
+        (2, 4, 3)
+        >>> x
+        array([[[0.70988236, 0.18180876, 0.54073621],
+                [0.38178158, 0.61168824, 0.64684354],
+                [0.98597752, 0.70650871, 0.31479029],
+                [0.15795399, 0.28162992, 0.98945647]],
+        <BLANKLINE>
+               [[0.620398  , 0.57025403, 0.46336542],
+                [0.44021889, 0.69926312, 0.60133428],
+                [0.89132308, 0.12030255, 0.35715804],
+                [0.04025218, 0.44304244, 0.10724799]]])
+
+        Unrandomized Halton 
+
+        >>> Halton(2,randomize="FALSE",seed=7)(4,warn=False)
+        array([[0.        , 0.        ],
+               [0.5       , 0.33333333],
+               [0.25      , 0.66666667],
+               [0.75      , 0.11111111]])
+        
+        All randomizations 
+
+        >>> Halton(2,randomize="LMS DP",seed=7)(4)
+        array([[0.83790457, 0.89981478],
+               [0.00986102, 0.4610941 ],
+               [0.62236343, 0.02796307],
+               [0.29427505, 0.79909098]])
+        >>> Halton(2,randomize="LMS DS",seed=7)(4)
+        array([[0.82718745, 0.90603116],
+               [0.0303368 , 0.44704107],
+               [0.60182684, 0.03580544],
+               [0.30505343, 0.78367016]])
+        >>> Halton(2,randomize="LMS",seed=7)(4,warn=False)
+        array([[0.        , 0.        ],
+               [0.82822666, 0.92392942],
+               [0.28838899, 0.46493682],
+               [0.6165384 , 0.2493814 ]])
+        >>> Halton(2,randomize="DP",seed=7)(4)
+        array([[0.11593484, 0.99232505],
+               [0.61593484, 0.65899172],
+               [0.36593484, 0.32565839],
+               [0.86593484, 0.77010283]])
+        >>> Halton(2,randomize="DS",seed=7)(4)
+        array([[0.56793849, 0.04063513],
+               [0.06793849, 0.37396846],
+               [0.81793849, 0.7073018 ],
+               [0.31793849, 0.15174624]])
+        >>> Halton(2,randomize="NUS",seed=7)(4)
+        array([[0.141964  , 0.99285569],
+               [0.65536579, 0.51938353],
+               [0.46955206, 0.11342811],
+               [0.78505432, 0.87032345]])
[…]
