+++ qmcpy/discrete_distribution/digital_net_any_base/faure.py
@@ -1,0 +1,16 @@
+from .digital_net_any_base import DigitalNetAnyBases
+
+
+class Faure(DigitalNetAnyBases):
+    r"""
+    Low discrepancy Faure points.
+
+    Note:
+        - The first point of an unrandomized Faure sequence is the origin.
+    
+    Examples:
+        >>> pass
+
+    """
+    
+    DEFAULT_GENERATING_MATRICES = "FAURE"
