### 2025-12-31 c46e098
Update tb_gbm_demo.py
