### 2025-12-31 2f8d4bf
draft some code for higher-order GDNs
