+++ qmcpy/discrete_distribution/digital_net_any_base/digital_net_any_base.py
@@ -134,6 +134,12 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
                 np.ctypeslib.ndpointer(ctypes.c_double, flags='C_CONTIGUOUS'), # randu_d_32
                 np.ctypeslib.ndpointer(ctypes.c_int, flags='C_CONTIGUOUS')]  # dvec
             self.halton_cf_qrng.restype = None
+        self.alpha = alpha
+        assert self.alpha>=1
+        assert self.alpha%1==0
+        if self.alpha>1:
+            assert (self.dvec==np.arange(self.d)).all(), "digital interlacing requires dimension is an int"
+        self.dtalpha = self.alpha*self.d 
         if self.type_bases_generating_matrices=="HALTON":
             self.bases = self.all_primes[self.dvec][None,:]
             self.m_max = int(np.ceil(np.log(self.n_limit)/np.log(self.bases.min())))
@@ -141,58 +147,97 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             self.t = self.m_max if self.m_max>t else t
             self.C = qmctoolscl.gdn_get_halton_generating_matrix(np.uint64(1),np.uint64(self.d),np.uint64(self._t_curr))
         elif self.type_bases_generating_matrices=="FAURE":
+            assert (self.dvec==np.arange(self.d)).all(), "Faure requires dimension is an int"
             p = self.all_primes[np.argmax(self.all_primes>=self.d)]
-            self.bases = p*np.ones((1,self.d),dtype=np.uint64)
+            self.bases = p*np.ones((1,self.dtalpha),dtype=np.uint64)
             self.m_max = int(np.ceil(np.log(self.n_limit)/np.log(p)))
             self._t_curr = self.m_max
             self.t = self.m_max if self.m_max>t else t
-            self.C = np.ones((self.dvec.max(),1,1),dtype=np.uint64)*np.eye(self._t_curr,dtype=np.uint64)
-            if self.dvec.max()>1:
+            self.C = np.ones((self.dtalpha,1,1),dtype=np.uint64)*np.eye(self._t_curr,dtype=np.uint64)
+            if self.dtalpha>1:
                 for a in range(self._t_curr):
                     for b in range(a+1):
                         self.C[1,a,b] = comb(a,b)%p
-            if self.dvec.max()>2:
-                for k in range(2,self.dvec.max()):
+            if self.dtalpha>2:
+                for k in range(2,self.dtalpha):
                     for a in range(self._t_curr):
                         for b in range(a+1):
                             self.C[k,a,b] = (int(self.C[1,a,b])*((k**(a-b))%p))%p
-            self.C = self.C[self.dvec][None,:,:,:]
+            self.C = self.C[None,:,:,:]
         else:
             self.bases = bases.astype(np.uint64)
             if self.bases.ndim==1: self.bases = self.bases[None,:]
-            assert self.bases.shape[1]>=self.d 
-            self.bases = self.bases[:,self.dvec]
+            assert self.bases.shape[1]>=self.dtalpha
+            if self.alpha==1:
+                self.bases = self.bases[:,self.dvec]
+            else:
+                self.bases = self.bases[:,:self.dtalpha]
             self.C = generating_matrices.astype(np.uint64)
             if self.C.ndim==3: self.C = self.C[None,:,:,:]
-            assert self.C.shape[1]>=self.d 
-            self.C = self.C[:,self.dvec,:,:]
+            assert self.C.shape[1]>=self.dtalpha
+            if self.alpha==1:
+                self.C = self.C[:,self.dvec,:,:]
[…]
