### 2025-12-31 d049abe
first try for Faure points doctests
