+++ qmcpy/discrete_distribution/digital_net_any_base/faure.py
@@ -9,7 +9,144 @@ class Faure(DigitalNetAnyBases):
         - The first point of an unrandomized Faure sequence is the origin.
     
     Examples:
-        >>> pass
+        >>> discrete_distrib = Faure(4,seed=7)
+        >>> discrete_distrib(4)
+        array([[0.1450272 , 0.587522  , 0.53005481, 0.72297587],
+               [0.91002739, 0.9486738 , 0.99520878, 0.26209473],
+               [0.38835418, 0.62726593, 0.71130669, 0.09272597],
+               [0.43149581, 0.06796957, 0.01715528, 0.9998576 ]])
+        >>> discrete_distrib
+        Faure (AbstractLDDiscreteDistribution)
+            d               2^(2)
+            replications    1
+            randomize       LMS DP
+            t               63
+            n_limit         2^(32)
+            entropy         7
+        
+        Replications of independent randomizations 
+
+        >>> x = Faure(3,seed=7,replications=2)(4)
+        >>> x.shape
+        (2, 4, 3)
+        >>> x
+        array([[[0.68991158, 0.61000002, 0.67989243],
+                [0.58982495, 0.7175662 , 0.50462275],
+                [0.12333072, 0.17197967, 0.31563725],
+                [0.90367942, 0.99885531, 0.02770573]],
+        <BLANKLINE>
+               [[0.81876583, 0.85609978, 0.76079113],
+                [0.49583728, 0.29945951, 0.0986577 ],
+                [0.1730512 , 0.41731239, 0.422907  ],
+                [0.97827231, 0.04710566, 0.54335177]]])
+
+        Unrandomized Faure 
+
+        >>> Faure(6,randomize="FALSE",seed=7)(4,warn=False)
+        array([[0.        , 0.        , 0.        , 0.        , 0.        ,
+                0.        ],
+               [0.14285714, 0.14285714, 0.14285714, 0.14285714, 0.14285714,
+                0.14285714],
+               [0.28571429, 0.28571429, 0.28571429, 0.28571429, 0.28571429,
+                0.28571429],
+               [0.42857143, 0.42857143, 0.42857143, 0.42857143, 0.42857143,
+                0.42857143]])
+        
+        All randomizations 
+
+        >>> Faure(3,randomize="LMS DP",seed=7)(4)
+        array([[0.6894943 , 0.31139262, 0.78744381],
+               [0.45345276, 0.87022239, 0.04198496],
+               [0.25871653, 0.42903509, 0.63353422],
+               [0.79204932, 0.95957485, 0.46904675]])
+        >>> Faure(3,randomize="LMS DS",seed=7)(4)
+        array([[0.05631564, 0.3701308 , 0.27474945],
+               [0.62289628, 0.13782096, 0.76230462],
+               [0.81895352, 0.91797542, 0.4629459 ],
+               [0.28289384, 0.05405794, 0.59451755]])
+        >>> Faure(3,randomize="LMS",seed=7)(4,warn=False)
+        array([[0.        , 0.        , 0.        ],
+               [0.5706961 , 0.78569224, 0.82552878],
+               [0.80394278, 0.56569421, 0.63728174],
+               [0.22678173, 0.73927861, 0.36137831]])
+        >>> Faure(3,randomize="DP",seed=7)(4)
+        array([[0.62331102, 0.06843544, 0.08388064],
+               [0.28997769, 0.7351021 , 0.7505473 ],
+               [0.95664436, 0.40176877, 0.41721397],
+               [0.51219991, 0.95732433, 0.63943619]])
+        >>> Faure(3,randomize="DS",seed=7)(4)
[…]
