### 2025-12-31 f02e9f6
conservative GDN t + start HO GDN linting
