+++ qmcpy/discrete_distribution/digital_net_any_base/faure.py
@@ -11,16 +11,16 @@ class Faure(DigitalNetAnyBases):
     Examples:
         >>> discrete_distrib = Faure(4,seed=7)
         >>> discrete_distrib(4)
-        array([[0.1450272 , 0.587522  , 0.53005481, 0.72297587],
-               [0.91002739, 0.9486738 , 0.99520878, 0.26209473],
-               [0.38835418, 0.62726593, 0.71130669, 0.09272597],
-               [0.43149581, 0.06796957, 0.01715528, 0.9998576 ]])
+        array([[0.93821576, 0.98657146, 0.32975962, 0.03942865],
+               [0.58283196, 0.16959024, 0.53811273, 0.23004558],
+               [0.01636438, 0.59543606, 0.95458401, 0.81716783],
+               [0.69975534, 0.36527546, 0.74717761, 0.40420262]])
         >>> discrete_distrib
         Faure (AbstractLDDiscreteDistribution)
             d               2^(2)
             replications    1
             randomize       LMS DP
-            t               63
+            t               14
             n_limit         2^(32)
             entropy         7
         
@@ -30,123 +30,125 @@ class Faure(DigitalNetAnyBases):
         >>> x.shape
         (2, 4, 3)
         >>> x
-        array([[[0.68991158, 0.61000002, 0.67989243],
-                [0.58982495, 0.7175662 , 0.50462275],
-                [0.12333072, 0.17197967, 0.31563725],
-                [0.90367942, 0.99885531, 0.02770573]],
+        array([[[0.07060326, 0.73528722, 0.24685699],
+                [0.83745824, 0.59870887, 0.82773772],
+                [0.6043124 , 0.16585117, 0.42540718],
+                [0.17372299, 0.45045811, 0.59981538]],
         <BLANKLINE>
-               [[0.81876583, 0.85609978, 0.76079113],
-                [0.49583728, 0.29945951, 0.0986577 ],
-                [0.1730512 , 0.41731239, 0.422907  ],
-                [0.97827231, 0.04710566, 0.54335177]]])
+               [[0.80940329, 0.4768136 , 0.79978683],
+                [0.21034443, 0.93042812, 0.40497512],
+                [0.51683229, 0.09321636, 0.3061612 ],
+                [0.72632804, 0.80064811, 0.08157659]]])
 
         Unrandomized Faure 
 
-        >>> Faure(6,randomize="FALSE",seed=7)(4,warn=False)
-        array([[0.        , 0.        , 0.        , 0.        , 0.        ,
-                0.        ],
-               [0.14285714, 0.14285714, 0.14285714, 0.14285714, 0.14285714,
-                0.14285714],
-               [0.28571429, 0.28571429, 0.28571429, 0.28571429, 0.28571429,
-                0.28571429],
-               [0.42857143, 0.42857143, 0.42857143, 0.42857143, 0.42857143,
-                0.42857143]])
+        >>> Faure(4,randomize="FALSE",seed=7)(10,warn=False)
+        array([[0.  , 0.  , 0.  , 0.  ],
+               [0.2 , 0.2 , 0.2 , 0.2 ],
+               [0.4 , 0.4 , 0.4 , 0.4 ],
+               [0.6 , 0.6 , 0.6 , 0.6 ],
+               [0.8 , 0.8 , 0.8 , 0.8 ],
+               [0.04, 0.24, 0.44, 0.64],
+               [0.24, 0.44, 0.64, 0.84],
+               [0.44, 0.64, 0.84, 0.04],
+               [0.64, 0.84, 0.04, 0.24],
+               [0.84, 0.04, 0.24, 0.44]])
         
         All randomizations 
 
         >>> Faure(3,randomize="LMS DP",seed=7)(4)
[…]
