### 2025-12-31 2528200
technical t digits + HO NUS patches
