+++ qmcpy/discrete_distribution/digital_net_any_base/digital_net_any_base.py
@@ -144,7 +144,7 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             self.bases = self.all_primes[self.dvec][None,:]
             self.m_max = int(np.ceil(np.log(self.n_limit)/np.log(self.bases.min())))
             self._t_curr = self.m_max
-            if t is None: t = 63
+            if t is None: t = int(np.ceil(-np.log(2**(-63))/np.log(self.bases.min())))
             self.t = self.m_max if self.m_max>t else t
             self.C = qmctoolscl.gdn_get_halton_generating_matrix(np.uint64(1),np.uint64(self.d),np.uint64(self._t_curr))
         elif self.type_bases_generating_matrices=="FAURE":
@@ -152,8 +152,8 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             p = self.all_primes[np.argmax(self.all_primes>=self.d)]
             self.bases = p*np.ones((1,self.dtalpha),dtype=np.uint64)
             self.m_max = int(np.ceil(np.log(self.n_limit)/np.log(p)))
+            if t is None: t = int(np.ceil(-np.log(2**(-63))/np.log(p)))
             self._t_curr = self.m_max
-            if t is None: t = self.m_max
             self.t = self.m_max if self.m_max>t else t
             self.C = np.ones((self.dtalpha,1,1),dtype=np.uint64)*np.eye(self._t_curr,dtype=np.uint64)
             if self.dtalpha>1:
@@ -182,6 +182,7 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             else:
                 self.C = self.C[:,:self.dtalpha,:,:]
             self.m_max,self._t_curr = self.C.shape[-2:]
+            if t is None: t = int(np.ceil(-np.log(2**(-63))/np.log(self.bases.min())))
             self.t = self.m_max if self.m_max>t else t
         if self.alpha>1:
             assert (self.bases==self.bases[0,0]).all(), "alpha>1 performs digital interlacing which requires the same base across dimensions and replications."
@@ -230,7 +231,7 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
                 self.perms = qmctoolscl.gdn_get_digital_permutations(self.rng,np.uint64(self.replications),np.uint64(self.d),self.t,np.uint64(r_b),self.bases)
             elif self.randomize=="LMS DS":
                 self.rshift = qmctoolscl.gdn_get_digital_shifts(self.rng,np.uint64(self.replications),np.uint64(self.d),self.t,np.uint64(r_b),self.bases)
-        elif "NUS" in self.randomize:
+        elif self.randomize=="NUS":
             if self.alpha==1:
                 new_seeds = self._base_seed.spawn(self.replications*self.d)
                 self.rngs = np.array([np.random.Generator(np.random.SFC64(new_seeds[j])) for j in range(self.replications*self.d)]).reshape(self.replications,self.d)
@@ -239,8 +240,8 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
                 new_seeds = self._base_seed.spawn(self.replications*self.dtalpha)
                 self.rngs = np.array([np.random.Generator(np.random.SFC64(new_seeds[j])) for j in range(self.replications*self.dtalpha)]).reshape(self.replications,self.dtalpha)
[…]
