+++ qmcpy/discrete_distribution/digital_net_any_base/faure.py
@@ -11,16 +11,16 @@ class Faure(DigitalNetAnyBases):
     Examples:
         >>> discrete_distrib = Faure(4,seed=7)
         >>> discrete_distrib(4)
-        array([[0.93821576, 0.98657146, 0.32975962, 0.03942865],
-               [0.58283196, 0.16959024, 0.53811273, 0.23004558],
-               [0.01636438, 0.59543606, 0.95458401, 0.81716783],
-               [0.69975534, 0.36527546, 0.74717761, 0.40420262]])
+        array([[0.33534484, 0.26765076, 0.49135906, 0.6052543 ],
+               [0.45194823, 0.53502655, 0.93195248, 0.5024511 ],
+               [0.69367212, 0.09800441, 0.77275989, 0.35304609],
+               [0.0086832 , 0.96517227, 0.01157255, 0.85055011]])
         >>> discrete_distrib
         Faure (AbstractLDDiscreteDistribution)
             d               2^(2)
             replications    1
             randomize       LMS DP
-            t               14
+            t               28
             n_limit         2^(32)
             entropy         7
         
@@ -30,15 +30,15 @@ class Faure(DigitalNetAnyBases):
         >>> x.shape
         (2, 4, 3)
         >>> x
-        array([[[0.07060326, 0.73528722, 0.24685699],
-                [0.83745824, 0.59870887, 0.82773772],
-                [0.6043124 , 0.16585117, 0.42540718],
-                [0.17372299, 0.45045811, 0.59981538]],
+        array([[[0.46995809, 0.81347921, 0.84921511],
+                [0.22573669, 0.29702269, 0.03044283],
+                [0.68125682, 0.38949622, 0.65611464],
+                [0.58494536, 0.14367836, 0.43908813]],
         <BLANKLINE>
-               [[0.80940329, 0.4768136 , 0.79978683],
-                [0.21034443, 0.93042812, 0.40497512],
-                [0.51683229, 0.09321636, 0.3061612 ],
-                [0.72632804, 0.80064811, 0.08157659]]])
+               [[0.44633162, 0.60684282, 0.96393795],
+                [0.13904548, 0.73709456, 0.37711571],
+                [0.79070871, 0.15560342, 0.12328662],
+                [0.40389155, 0.97863405, 0.0114599 ]]])
 
         Unrandomized Faure 
 
@@ -57,98 +57,98 @@ class Faure(DigitalNetAnyBases):
         All randomizations 
 
         >>> Faure(3,randomize="LMS DP",seed=7)(4)
-        array([[0.09276526, 0.67280281, 0.05133827],
-               [0.99257681, 0.47726151, 0.79454369],
-               [0.52628644, 0.23872389, 0.65411616],
-               [0.31075747, 0.66239464, 0.36868717]])
+        array([[0.60869072, 0.76096155, 0.79807281],
+               [0.03946861, 0.56443723, 0.39462162],
+               [0.84046405, 0.17459934, 0.31263449],
+               [0.39080182, 0.42253604, 0.09713822]])
         >>> Faure(3,randomize="LMS DS",seed=7)(4)
-        array([[0.27002269, 0.26003102, 0.63911171],
-               [0.50326958, 0.38652922, 0.11399055],
-               [0.74047972, 0.84089074, 0.70985883],
-               [0.1504592 , 0.4608547 , 0.90612583]])
+        array([[0.09363194, 0.63822128, 0.52606134],
+               [0.66021198, 0.02837507, 0.7155204 ],
+               [0.85642224, 0.83340554, 0.22549083],
+               [0.31989976, 0.30119144, 0.00893035]])
         >>> Faure(3,randomize="LMS",seed=7)(4,warn=False)
[…]
