### 2025-12-31 8dbdcfe
faure doctests
