+++ qmcpy/discrete_distribution/digital_net_any_base/faure.py
@@ -10,11 +10,32 @@ class Faure(DigitalNetAnyBases):
     
     Examples:
         >>> discrete_distrib = Faure(4,seed=7)
-        >>> discrete_distrib(4)
+        >>> discrete_distrib(25)
         array([[0.33534484, 0.26765076, 0.49135906, 0.6052543 ],
                [0.45194823, 0.53502655, 0.93195248, 0.5024511 ],
                [0.69367212, 0.09800441, 0.77275989, 0.35304609],
-               [0.0086832 , 0.96517227, 0.01157255, 0.85055011]])
+               [0.0086832 , 0.96517227, 0.01157255, 0.85055011],
+               [0.97041531, 0.63254601, 0.25235643, 0.18869841],
+               [0.20862319, 0.45844039, 0.64733082, 0.97109116],
+               [0.57003274, 0.12560935, 0.08772865, 0.02771165],
+               [0.73495613, 0.91292101, 0.32692686, 0.68483526],
+               [0.05156003, 0.78796242, 0.56754229, 0.54342666],
+               [0.89329109, 0.21506683, 0.80647015, 0.27293526],
+               [0.25277941, 0.07296122, 0.23030907, 0.46332812],
+               [0.4941915 , 0.94839898, 0.46928813, 0.39348452],
+               [0.60952104, 0.69550381, 0.90983954, 0.81009581],
+               [0.17092444, 0.37887902, 0.74911456, 0.10721583],
+               [0.93584834, 0.40585698, 0.18944839, 0.72587572],
+               [0.36989645, 0.85447136, 0.97641039, 0.14826857],
+               [0.53450751, 0.7376398 , 0.61725827, 0.64572148],
+               [0.65143961, 0.28501343, 0.05601974, 0.58393268],
+               [0.09284913, 0.59218182, 0.29689116, 0.23247628],
+               [0.80817253, 0.02749358, 0.53742056, 0.88960099],
+               [0.29406032, 0.64634796, 0.15459067, 0.31205785],
+               [0.40906423, 0.35338995, 0.39383649, 0.93066193],
+               [0.77079531, 0.50849426, 0.43438996, 0.06810554],
+               [0.13540741, 0.17593145, 0.87500744, 0.76631324],
+               [0.85233694, 0.81903638, 0.71417647, 0.42286144]])
         >>> discrete_distrib
         Faure (AbstractLDDiscreteDistribution)
             d               2^(2)
@@ -26,23 +47,33 @@ class Faure(DigitalNetAnyBases):
         
         Replications of independent randomizations 
 
-        >>> x = Faure(3,seed=7,replications=2)(4)
+        >>> x = Faure(3,seed=7,replications=2)(9)
         >>> x.shape
-        (2, 4, 3)
+        (2, 9, 3)
         >>> x
         array([[[0.46995809, 0.81347921, 0.84921511],
                 [0.22573669, 0.29702269, 0.03044283],
                 [0.68125682, 0.38949622, 0.65611464],
-                [0.58494536, 0.14367836, 0.43908813]],
+                [0.58494536, 0.14367836, 0.43908813],
+                [0.00739063, 0.63809438, 0.95363225],
+                [0.79609189, 0.71822727, 0.1472069 ],
+                [0.36690976, 0.47240941, 0.2470566 ],
+                [0.12284101, 0.96819212, 0.55174327],
+                [0.91154206, 0.05940035, 0.73295342]],
         <BLANKLINE>
                [[0.44633162, 0.60684282, 0.96393795],
                 [0.13904548, 0.73709456, 0.37711571],
                 [0.79070871, 0.15560342, 0.12328662],
[…]
