### 2025-12-31 6a707c4
digital net any base comparison
