+++ qmcpy/discrete_distribution/digital_net_any_base/digital_net_any_base.py
@@ -1,6 +1,7 @@
 import warnings
 from ..abstract_discrete_distribution import AbstractLDDiscreteDistribution
 from ...util import ParameterError,ParameterWarning
+from ..digital_net_b2 import DigitalNetB2
 import qmctoolscl
 import numpy as np
 from math import *
@@ -24,8 +25,120 @@ class DigitalNetAnyBases(AbstractLDDiscreteDistribution):
             
             i.e. do *not* pass in interlaced `generating_matrices` and set `alpha>1`, this will apply additional interlacing. 
     
+    A few examples below showcase how to pass in custom bases and generating matrices. Many other examples can be found in the Halton and Faure implementations
+    
     Examples:
-        >>> pass
+        >>> bases = 3 
+        >>> generating_matrices = np.array(
+        ...     [
+        ...         [[1, 0, 0],
+        ...          [0, 1, 0],
+        ...          [1, 2, 1]],
+        ...         [[2, 0, 0],
+        ...          [1, 2, 0],
+        ...          [0, 2, 2]]
+        ...     ],
+        ...     dtype=np.uint64)
+        >>> DigitalNetAnyBases(2,randomize="False",bases_generating_matrices=(bases,generating_matrices))(9,warn=False)
+        array([[0.        , 0.        ],
+               [0.33333333, 0.66666667],
+               [0.66666667, 0.33333333],
+               [0.11111111, 0.55555556],
+               [0.44444444, 0.22222222],
+               [0.77777778, 0.88888889],
+               [0.22222222, 0.77777778],
+               [0.55555556, 0.44444444],
+               [0.88888889, 0.11111111]])
+        >>> DigitalNetAnyBases(1,randomize="False",alpha=2,bases_generating_matrices=(bases,generating_matrices))(9,warn=False)
+        array([[0.        ],
+               [0.55555556],
+               [0.77777778],
+               [0.17283951],
+               [0.39506173],
+               [0.95061728],
+               [0.30864198],
+               [0.5308642 ],
+               [0.75308642]])
+    
+        >>> rng = np.random.Generator(np.random.PCG64(7))
+        >>> bases = np.array(
+        ...     [[2,5,7,23],
+        ...      [3,11,13,17]],
+        ...     dtype=np.uint64)
+        >>> generating_matrices = np.stack(
+        ...     [
+        ...         np.stack(
+        ...             [   np.tril(rng.integers(0,2,(10,10))),
+        ...                 np.tril(rng.integers(0,5,(10,10))),
+        ...                 np.tril(rng.integers(0,7,(10,10))),
+        ...                 np.tril(rng.integers(0,23,(10,10))),],axis=0),
+        ...         np.stack(
+        ...             [   np.tril(rng.integers(0,3,(10,10))),
+        ...                 np.tril(rng.integers(0,11,(10,10))),
+        ...                 np.tril(rng.integers(0,13,(10,10))),
+        ...                 np.tril(rng.integers(0,17,(10,10))),],axis=0),
+        ...     ],axis=0).astype(np.uint64)
+        >>> discrete_distrib = DigitalNetAnyBases(
+        ...     dimension = 4,
+        ...     randomize = "NUS",
+        ...     seed = 7,
+        ...     replications = 2,
+        ...     bases_generating_matrices=(bases,generating_matrices),
+        ...     )
+        >>> discrete_distrib(10,warn=False)
+        array([[[0.141964  , 0.50711584, 0.75766621, 0.43359373],
+                [0.65536579, 0.66327673, 0.21825902, 0.93901529],
[…]
