### 2025-12-31 5768146
doc updates + reorg
