+++ qmcpy/true_measure/geometric_brownian_motion.py
@@ -248,6 +248,6 @@ class GeometricBrownianMotion(BrownianMotion):
                 warn (bool): whether to warn about sample generation
                 
             Returns:
-                ndarray or tuple: GBM samples, optionally with weights if return_weights=True
+                samples Union[ndarray,tuple]: GBM samples, optionally with weights if return_weights=True
             """
             return super().gen_samples(n=n, n_min=n_min, n_max=n_max, return_weights=return_weights, warn=warn)
