+++ qmcpy/discrete_distribution/digital_net_any_bases/halton.py
@@ -1,4 +1,4 @@
-from .digital_net_any_base import DigitalNetAnyBases
+from .digital_net_any_bases import DigitalNetAnyBases
 
 
 class Halton(DigitalNetAnyBases):
@@ -167,17 +167,3 @@ class Halton(DigitalNetAnyBases):
     """
     
     DEFAULT_GENERATING_MATRICES = "HALTON"
-
-class Faure(DigitalNetAnyBases):
-    r"""
-    Low discrepancy Faure points.
-
-    Note:
-        - The first point of an unrandomized Faure sequence is the origin.
-    
-    Examples:
-        >>> pass
-
-    """
-    
-    DEFAULT_GENERATING_MATRICES = "FAURE"
