### 2025-12-31 14fcdc8
patch doctests
