### 2026-01-01 057b432
Add and Enhance Test Documenation  (#450)
* Interactive visualization.

* Add demo to documentation

* Minor improvements

* Revert "Minor improvements"

This reverts commit aeb3d5cb734172146c6c8f171c6dd169b4e4cd4d.

* Revert "Interactive visualization."

This reverts commit 3b3f7958c1e6f0070cf3f786310506205e70f5df.

* Update gbm_demo.ipynb

* Update gbm_demo.ipynb

* Add interactive visualization

* Add section title and some text before code

* Change input parameter to 2**n

* Add help

* clean up

* Delete workout

* Update gbm_demo.ipynb

* added to do list

* Update gbm_demo.ipynb

* Upd
[…]
