+++ qmcpy/true_measure/uniform.py
@@ -8,7 +8,7 @@ from typing import Union
 
 class Uniform(AbstractTrueMeasure):
     r"""
-    Uniform distribution, see [https://en.wikipedia.org/wiki/Continuous_uniform_distribution](https://en.wikipedia.org/wiki/Continuous_uniform_distribution). 
+    Uniform distribution, see [https://en.wikipedia.org/wiki/Continuous_uniform_distribution](https://en.wikipedia.org/wiki/Continuous_uniform_distribution).
 
     Examples:
         >>> true_measure = Uniform(DigitalNetB2(2,seed=7),lower_bound=[0,.5],upper_bound=[2,3])
@@ -22,10 +22,10 @@ class Uniform(AbstractTrueMeasure):
             lower_bound     [0.  0.5]
             upper_bound     [2 3]
 
-        With independent replications 
+        With independent replications
 
         >>> x = Uniform(DigitalNetB2(3,seed=7,replications=2),lower_bound=[.25,.5,.75],upper_bound=[1.75,1.5,1.25])(4)
-        >>> x.shape 
+        >>> x.shape
         (2, 4, 3)
         >>> x
         array([[[0.61979915, 0.6821862 , 1.12366296],
@@ -38,53 +38,58 @@ class Uniform(AbstractTrueMeasure):
                 [0.59821215, 0.79281506, 0.78420518],
                 [1.37943573, 1.10241448, 1.13481488]]])
     """
-    
+
     def __init__(self, sampler, lower_bound=0, upper_bound=1):
         r"""
         Args:
-            sampler (Union[AbstractDiscreteDistribution,AbstractTrueMeasure]): Either  
-                
+            sampler (Union[AbstractDiscreteDistribution,AbstractTrueMeasure]): Either
+
                 - a discrete distribution from which to transform samples, or
                 - a true measure by which to compose a transform.
             lower_bound (Union[float,np.ndarray]): Lower bound.
             upper_bound (Union[float,np.ndarray]): Upper bound.
         """
-        self.parameters = ['lower_bound', 'upper_bound']
-        self.domain = np.array([[0,1]])
+        self.parameters = ["lower_bound", "upper_bound"]
+        self.domain = np.array([[0, 1]])
         self._parse_sampler(sampler)
         self.lower_bound = lower_bound
         self.upper_bound = upper_bound
         if np.isscalar(self.lower_bound):
-            lower_bound = np.tile(self.lower_bound,self.d)
+            lower_bound = np.tile(self.lower_bound, self.d)
         if np.isscalar(self.upper_bound):
-            upper_bound = np.tile(self.upper_bound,self.d)
+            upper_bound = np.tile(self.upper_bound, self.d)
         self.a = np.array(lower_bound)
         self.b = np.array(upper_bound)
-        if len(self.a)!=self.d or len(self.b)!=self.d:
-            raise DimensionError('upper bound and lower bound must be of length dimension')
+        if len(self.a) != self.d or len(self.b) != self.d:
+            raise DimensionError(
+                "upper bound and lower bound must be of length dimension"
+            )
         self.delta = self.b - self.a
-        self.inv_delta_prod = 1/self.delta.prod()
-        self.range = np.hstack((self.a.reshape((self.d,1)),self.b.reshape((self.d,1))))
-        super(Uniform,self).__init__()
-        assert self.a.shape==(self.d,) and self.b.shape==(self.d,)
+        self.inv_delta_prod = 1 / self.delta.prod()
+        self.range = np.hstack(
+            (self.a.reshape((self.d, 1)), self.b.reshape((self.d, 1)))
+        )
+        super(Uniform, self).__init__()
[…]
