### 2026-01-01 cf13fee
Fix notebook tests
