+++ test/booktests/tb_portfolio_allocation_demo.py
@@ -1,41 +1,29 @@
 import unittest
-from __init__ import TB_TIMEOUT, BaseNotebookTest, run_notebook
+from __init__ import TB_TIMEOUT, BaseNotebookTest
 
 class NotebookTests(BaseNotebookTest):
 
     def test_portfolio_allocation_demo_notebook(self):
-        notebook_path, notebook_dir = self.locate_notebook('../../demos/portfolio/portfolio_allocation_demo.ipynb')
-        value = [
-            """params = [
-    (4, 2**14),     
-    (10, 2**15),   
-    (20, 2**16),    
-    (100, 2**17),   
-    (500, 2**18),   
-    (1000, 2**19)   
-]""",
-            "n_ports = [2**13, 2**14, 2**15]",
-            "start_date = '2014-01-01'", 
-            "dimensions = [5, 10, 20, 50, 100, 200, 500, 1000]",
-            'tickers1 = ["AAPL", "AMZN", "CSCO", "IBM", "TSLA", "META", "ABNB", "UPS", "NFLX", "MRNA"]',
-            'description1 = ["Apple", "Amazon", "CISCO", "IBM", "Tesla", "Meta", "Airbnb", "UPS", "Netflix", "Moderna"]"',
-            'tickers2 = ["AAPL", "AMZN", "CSCO","IBM","TSLA","META","ABNB","UPS","NFLX","MRNA","^IXIC", "T","GE","FMC","AMC","JPM","DIS","CVX","GOOGL","BA"]',
-            'description2 = ["Apple", "Amazon", "CISCO", "IBM","Tesla","Meta","Airbnb","UPS","Netflix","Moderna","NASDAQ","AT&T","General Electric","FMC","AMC","JPMorgan","Disney","Chevron","Google","Boeing"]',
-        ]
-        new_value = [
-            """params = [
-    (4, 2**7),     
-    (10, 2**8)]""",
-            "n_ports = [2**7, 2**8]",
-            "start_date = '2025-01-01'", 
-            "dimensions = [5, 10]",
-            'tickers1 = ["AAPL", "AMZN"]',
-            'description1 = ["Apple", "Amazon"]"',
-             'tickers2 = ["AAPL", "AMZN", "CSCO"]',
-            'description2 = ["Apple", "Amazon", "CISCO"]',
-        ]
-        run_notebook(notebook_path, notebook_dir, change_value=True, 
-                     value=value, new_value=new_value, timeout=TB_TIMEOUT)
+        notebook_path, _ = self.locate_notebook('../../demos/portfolio/portfolio_allocation_demo.ipynb')
+        
+        # Define ticker patterns for readability (wildcard: [...] represents variable list items)
+        old_tick1 = '"CSCO", "IBM", "TSLA", "META", "ABNB", "UPS", "NFLX", "MRNA"'
+        old_desc1 = '"CISCO", "IBM", "Tesla", "Meta", "Airbnb", "UPS", "Netflix", "Moderna"'
+        old_tick2 = '"IBM","TSLA","META","ABNB","UPS","NFLX","MRNA","^IXIC", "T","GE","FMC","AMC","JPM","DIS","CVX","GOOGL","BA"'
+        old_desc2 = '"IBM","Tesla","Meta","Airbnb","UPS","Netflix","Moderna","NASDAQ","AT&T","General Electric","FMC","AMC","JPMorgan","Disney","Chevron","Google","Boeing"'
+        
+        replacements = {
+            "n_ports = [2**13, 2**14, 2**15]": "n_ports = [2**7, 2**8]",
+            "start_date = '2014-01-01'": "start_date = '2025-01-01'",
[…]
