### 2026-01-01 0cb04b8
Fix broken booktest
