+++ test/booktests/tb_portfolio_allocation_demo.py
@@ -16,12 +16,14 @@ class NotebookTests(BaseNotebookTest):
             "n_ports = [2**13, 2**14, 2**15]": "n_ports = [2**7, 2**8]",
             "start_date = '2014-01-01'": "start_date = '2025-01-01'",
             "dimensions = [5, 10, 20, 50, 100, 200, 500, 1000]": "dimensions = [5, 10]",
+            "num_ports = 2**14": "num_ports = 2**8",
             f'tickers1 = ["AAPL", "AMZN", {old_tick1}]': 'tickers1 = ["AAPL", "AMZN"]',
             f'description1 = ["Apple", "Amazon", {old_desc1}]': 'description1 = ["Apple", "Amazon"]',
             f'tickers2 = ["AAPL", "AMZN", "CSCO",{old_tick2}]': 'tickers2 = ["AAPL", "AMZN", "CSCO"]',
             f'description2 = ["Apple", "Amazon", "CISCO", {old_desc2}]': 'description2 = ["Apple", "Amazon", "CISCO"]',
         }
-
+        
+        # Add a line to notebook cell [44] after the line
         self.run_notebook(notebook_path, replacements)
 
 
