### 2026-01-02 0e8ccb2
Update pyproject.toml
