### 2026-01-02 a43d57d
Update tb_portfolio_allocation_demo.py
