### 2026-01-02 26c28d9
Remove a redundant function. Change notebook in memory or on disk.
