+++ .github/workflows/alltests.yml
@@ -199,7 +199,18 @@ jobs:
         if: runner.os == 'Windows'
         shell: pwsh
         run: |
-          choco install miktex -y
+          # Install MiKTeX with retries
+          $maxTries = 2
+          $try = 0
+          while ($try -lt $maxTries) {
+            choco install miktex --no-progress && exit 0
+            $try++
+            Write-Host "Retry $try of $maxTries..."
+            Start-Sleep -Seconds 30
+          }
+          Write-Error "MiKTeX installation failed after $maxTries attempts."
+
+          # Install Strawberry Perl (required by MiKTeX)
           choco install strawberryperl -y
           # Put MiKTeX on PATH for all subsequent steps
           echo "C:\Program Files\MiKTeX\miktex\bin\x64" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
