### 2026-01-02 ef4d3c8
Add MultivariateStudentTJoint true measure and SciPyWrapper joint support
