+++ qmcpy/true_measure/scipy_wrapper.py
@@ -1,15 +1,11 @@
 from .abstract_true_measure import AbstractTrueMeasure
 from ..util import DimensionError, ParameterError
-from ..discrete_distribution.abstract_discrete_distribution import (
-    AbstractDiscreteDistribution,
-)
-from ..discrete_distribution.abstract_discrete_distribution import (
-    AbstractDiscreteDistribution,
-)
-from ..discrete_distribution import DigitalNetB2
+from ..discrete_distribution.abstract_discrete_distribution import AbstractDiscreteDistribution
+
 import numpy as np
 import scipy.stats
 import warnings
+
 
 
 class _MVNAdapter:
@@ -97,7 +93,9 @@ class SciPyWrapper(AbstractTrueMeasure):
     Examples:
         Independent marginals from ``scipy.stats``:
 
-        >>> from qmcpy.discrete_distribution import DigitalNetB2
+        >>> from importlib import import_module
+        >>> _base = SciPyWrapper.__module__.split(".true_measure.", 1)[0]
+        >>> DigitalNetB2 = import_module(_base + ".discrete_distribution").DigitalNetB2
         >>> import scipy.stats as stats
         >>> tm = SciPyWrapper(
         ...     sampler=DigitalNetB2(3, seed=7),
@@ -125,8 +123,8 @@ class SciPyWrapper(AbstractTrueMeasure):
         >>> true_measure = SciPyWrapper(
         ...     sampler=DigitalNetB2(2, seed=13),
         ...     scipy_distribs=[
-        ...         scipy.stats.t(df=df, loc=0.0, scale=1.0),
-        ...         scipy.stats.t(df=df, loc=1.0, scale=2.0),
+        ...         stats.t(df=df, loc=0.0, scale=1.0),
+        ...         stats.t(df=df, loc=1.0, scale=2.0),
         ...     ],
         ... )
         >>> xs = true_measure(4)
@@ -134,70 +132,6 @@ class SciPyWrapper(AbstractTrueMeasure):
         (4, 2)
     """
 
-    def __init__(self, sampler, scipy_distribs):
-        """
-        Parameters
-        ----------
-        sampler : AbstractDiscreteDistribution
-            Low discrepancy or iid sampler in dimension d, living on [0,1)^d.
-        scipy_distribs :
-            One of the following:
-
-            - A single SciPy 1D continuous frozen distribution.
-            - A list of such frozen distributions (independent marginals).
-            - A custom 1D distribution object with ``ppf`` and ``pdf`` or
-              ``logpdf`` methods.
-            - A joint object with:
-                * ``transform(u)`` method
-                * optional ``logpdf(x)`` method
-                * ``dim`` or ``dimension`` attribute (otherwise ``sampler.d``).
-        """
-        self.domain = np.array([[0.0, 1.0]])
-
-        >>> true_measure = SciPyWrapper(
-        ...     sampler = DigitalNetB2(3,seed=7),
-        ...     scipy_distribs = [
-        ...         scipy.stats.uniform(loc=1,scale=2),
-        ...         scipy.stats.norm(loc=3,scale=4),
-        ...         scipy.stats.gamma(a=5,loc=6,scale=7)])
-        >>> true_measure.range
-        array([[  1.,   3.],
-               [-inf,  inf],
-               [  6.,  inf]])
-        >>> true_measure(4)
-        array([[ 2.26535046,  6.36077755, 46.10334984],
-               [ 1.37949875,  0.8419074 , 38.66873073],
-               [ 2.79562889, -0.65019733, 17.63758514],
-               [ 1.91172032,  4.67357136, 55.20163754]])
-
[…]
