+++ test/booktests/tb_scipywrapper_demo.py
@@ -1,0 +1,12 @@
+import unittest
+from testbook import testbook
+from __init__ import TB_TIMEOUT, BaseNotebookTest
+
+class NotebookTests(BaseNotebookTest):
+
+    @testbook('..\..\demos\scipywrapper_dependence_custom\scipywrapper_demo.ipynb', execute=True, timeout=TB_TIMEOUT)
+    def test_scipywrapper_demo_notebook(self, tb):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
