+++ mkdocs.yml
@@ -61,14 +61,15 @@ nav:
       - Importance Sampling with True Measures: demos/some_true_measures.ipynb 
       - Sensitivity Analysis for the Iris Dataset: demos/iris.ipynb 
       - Ray Tracing: demos/ray_tracing.ipynb 
+      - Importance Sampling with True Measures:
+        - SciPyWrapper dependence and Custom distributions: demos/scipywrapper_dependence_custom/scipywrapper_demo.ipynb
   - For Developers:
     - Release Manual: RELEASE.md
     - Testing Guidelines: 
       - Doc tests and unit tests on `qmcpy`: tests.md
       - Unit tests on Jupyter Notebooks: booktests.md
   - Printable Docs: print_page/
-  - SciPyWrapper dependence and Custom distributions: demos/scipywrapper_dependence_custom/scipywrapper_demo.ipynb
-  - SciPyWrapper blog: demos/scipywrapper_dependence_custom/blog.md
+  
 
 
 plugins:
