### 2026-01-02 c4f3e46
Speed up test/booktests/tb_asian_option_mlqmc.py
