+++ test/booktests/tb_asian_option_mlqmc.py
@@ -1,13 +1,17 @@
 import unittest, pytest
-from testbook import testbook
-from __init__ import TB_TIMEOUT, BaseNotebookTest
+from __init__ import BaseNotebookTest
 
 @pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
 
-    @testbook('../../demos/asian-option-mlqmc.ipynb', execute=True, timeout=TB_TIMEOUT)
     def test_asian_option_mlqmc_notebook(self, tb):
-        pass
+        notebook_path, _ = self.locate_notebook('../../demos/asian-option-mlqmc.ipynb')
+        replacements = {"for level in range(5):":"for level in range(2)",
+                        "abs_tol=5e-3":"abs_tol=1e-2",
+                        "tolerances = 5*np.logspace(-1, -3, num=5)":"tolerances = 5*np.logspace(-1, -2, num=2)",
+                        "for method in range(4):":"for method in range(2):",
+                        }
+        self.run_notebook(notebook_path, replacements)
 
 if __name__ == '__main__':
     unittest.main()
