### 2026-01-02 473527e
Speedup tb_vectorized_qmc_bayes.py
