### 2026-01-02 d90fea4
Speedup tb_ray_tracing.py
