### 2026-01-02 3bd5f19
Speedup tb_vectorized_qmc
