+++ test/booktests/tb_vectorized_qmc_bayes.py
@@ -1,7 +1,6 @@
 import unittest, pytest
 from __init__ import BaseNotebookTest
 
-@pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
 
     def test_vectorized_qmc_bayes_notebook(self):
