### 2026-01-02 6a31fdc
Speedup test/booktests/tb_qei_demo_for_blog.py
