### 2026-01-02 6c84e46
Update tb_digital_net_b2.py
