### 2026-01-02 e01d723
Apply PR review feedback: fix whitespace and simplify parameter passing (#453)
* Initial plan

* Apply review comments: fix whitespace and simplify parameter

Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>

---------

Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
Co-authored-by: sou-cheng-choi <18743024+sou-cheng-choi@users.noreply.github.com>
