### 2026-01-02 bbc747d
Speed up tb_gbm_demo
