+++ demos/GBM/outputs/gbm_comparison_table.tex
@@ -6,12 +6,10 @@
 Method & Sampler & Mean & Std Dev & Mean Absolute & Std Dev  & Mean Time (s)  & Std Dev (s) & Speedup \\
  &  &  &   &  Error &  Error &   & &  \\
 \midrule
-QuantLib & IIDStdUniform & 105.33130511 & 21.46705631 & 0.20419547 & 0.22961748 & 0.64609085 & 0.00164129 & - \\
-QMCPy & IIDStdUniform & 105.14433550 & 21.38484187 & 0.01722587 & 0.14740305 & 0.28451041 & 0.00064843 & 2.27088649 \\
-QuantLib & Sobol & 105.09290730 & 21.05175738 & 0.03420234 & 0.18568144 & 0.73306963 & 0.00466660 & - \\
-QMCPy & Sobol & 105.12737531 & 21.23370685 & 0.00026567 & 0.00373197 & 0.28089426 & 0.00324197 & 2.30012124 \\
-QMCPy & Lattice & 105.12838774 & 21.24595225 & 0.00127811 & 0.00851343 & 0.26862942 & 0.00121529 & 2.40513810 \\
-QMCPy & Halton & 105.12707683 & 21.23373113 & 0.00003281 & 0.00370770 & 1.95300934 & 0.01922076 & 0.33081810 \\
+QuantLib & IIDStdUniform & 106.10890180 & 21.44799737 & 0.98179217 & 0.21055854 & 0.00192896 & 0.00000000 & - \\
+QMCPy & IIDStdUniform & 103.23414020 & 22.31960754 & 1.89296944 & 1.08216871 & 0.00055561 & 0.00000000 & 3.47177638 \\
+QuantLib & Sobol & 104.74526647 & 18.95833721 & 0.38184316 & 2.27910161 & 0.00221435 & 0.00000000 & - \\
+QMCPy & Sobol & 105.12883681 & 21.27311000 & 0.00172717 & 0.03567117 & 0.00066442 & 0.00000000 & 2.90323590 \\
 \bottomrule
 \end{tabular}
 \end{table}
