### 2026-01-02 e6d4823
Speedup further
