+++ test/booktests/tb_ray_tracing.py
@@ -6,10 +6,10 @@ class NotebookTests(BaseNotebookTest):
     def test_ray_tracing_notebook(self):
         notebook_path, _ = self.locate_notebook('../../demos/ray_tracing.ipynb')
         replacements = {
-            'n = 16': 'n = 8',
-            'd = 16': 'd = 4',
-            'px = 256': 'px = 64',
-            'parallel_x_blocks=1, parallel_y_blocks=1': 'parallel_x_blocks=2, parallel_y_blocks=2'
+            'n = 16': 'n = 2',
+            'd = 16': 'd = 2',
+            'px = 256': 'px = 32',
+            'parallel_x_blocks=1, parallel_y_blocks=1': 'parallel_x_blocks=4, parallel_y_blocks=4'
         }
         self.run_notebook(notebook_path, replacements)
 
