### 2026-01-02 2def4b1
Fix MultivariateStudentTJoint doctest import path
