### 2026-01-02 c95defd
Update tb_gbm_demo.py
