### 2026-01-02 e0de0ca
Update tb_gbm_demo.py
