### 2026-01-02 2bcd8c9
Speedup CI Test Workflow with Injection (In-Memory Replacements) (#452)
* Delete testing.py

* Delete test_geometric_brownian_motion.py

* Update geometric_brownian_motion_copy.py

* changes

* Update geometric_brownian_motion_copy.py

* Create testing.py

to see what results I'm gettinh

* Fix doctests

* Revert changes

* Fix GBM

* Remove testing.py

* add colab button

* Fix Colab button

* Clean up

* Clean up more

* Add QMCPy heat map for covariance

* Add _validate_samples only in long tests. Add _weight()

* Final change

* Add unit tests for (G)BM and more comprehensive gbm_demo note
[…]
