+++ test/booktests/tb_elliptic_pde.py
@@ -1,5 +1,5 @@
 import unittest, pytest
-from __init__ import TB_TIMEOUT, BaseNotebookTest
+from __init__ import BaseNotebookTest
 
 @pytest.mark.slow
 class NotebookTests(BaseNotebookTest):
@@ -8,8 +8,7 @@ class NotebookTests(BaseNotebookTest):
  
     def test_elliptic_pde_notebook(self):
         replacements = {"plot_convergence(execute_convergence_test": "#plot_convergence(execute_convergence_test",}
-        self.change_notebook_cells(self.notebook_path, replacements)        
-        self.run_notebook(self.notebook_path, timeout=TB_TIMEOUT)
+        self.run_notebook(self.notebook_path, replacements)
 
 if __name__ == '__main__':
     unittest.main()
