### 2026-01-03 6911864
Add joint true measures for SciPyWrapper demo + fix booktest path
