+++ qmcpy/true_measure/__init__.py
@@ -10,7 +10,10 @@ from .johnsons_su import JohnsonsSU
 from .scipy_wrapper import SciPyWrapper
 from .matern_gp import MaternGP
 from .multivariate_student_t_joint import MultivariateStudentTJoint
-
+from .multivariate_normal_joint import MultivariateNormalJoint
+from .zero_inflated_exp_uniform_joint import ZeroInflatedExpUniformJoint
+from .uniform_triangle_joint import UniformTriangleJoint
+from .triangular_user_true_measure import TriangularUserTrueMeasure, TriangularUserDistribution, BadTriangularDistribution
 
 TrueMeasure = AbstractTrueMeasure
 _TrueMeasure = AbstractTrueMeasure
