+++ qmcpy/true_measure/multivariate_student_t_joint.py
@@ -1,5 +1,6 @@
 import numpy as np
 import scipy.stats as stats
+from ..discrete_distribution import DigitalNetB2
 
 from ..util import DimensionError, ParameterError
 from .scipy_wrapper import SciPyWrapper
@@ -97,8 +98,6 @@ class MultivariateStudentTJoint(SciPyWrapper):
     Convenience true measure: joint multivariate Student t.
 
     Example:
-    >>> import importlib
-    >>> DigitalNetB2 = importlib.import_module(__package__.rsplit(".", 1)[0] + ".discrete_distribution").DigitalNetB2
     >>> tm = MultivariateStudentTJoint(
     ...     sampler=DigitalNetB2(2, seed=7),
     ...     df=5,
