+++ test/booktests/tb_scipywrapper_demo.py
@@ -4,7 +4,7 @@ from __init__ import TB_TIMEOUT, BaseNotebookTest
 
 class NotebookTests(BaseNotebookTest):
 
-    @testbook('..\..\demos\scipywrapper_dependence_custom\scipywrapper_demo.ipynb', execute=True, timeout=TB_TIMEOUT)
+    @testbook('../../demos/scipywrapper_dependence_custom/scipywrapper_demo.ipynb', execute=True, timeout=TB_TIMEOUT)
     def test_scipywrapper_demo_notebook(self, tb):
         pass
 
