### 2026-01-03 26eb3c2
Delete tb_plot_proj_function.py
