+++ test/booktests/tb_gbm_demo.py
@@ -1,7 +1,6 @@
 import unittest
 from __init__ import BaseNotebookTest
 
-@unittest.skip("Skipping as it is very slow on Ubuntu")
 class NotebookTests(BaseNotebookTest):
     
     def test_gbm_demo_notebook(self):
@@ -17,15 +16,14 @@ class NotebookTests(BaseNotebookTest):
             "qp.Lattice(2**8, seed=42)": "qp.Lattice(4, seed=42)",
             "'n_paths': 2**14": "'n_paths': 4",
             "'n_steps': 252": "'n_steps': 2",
-            "replications = 3": "replications = 1",
-            "%timeit -n 10 -r 3 -o": "%timeit -n 2 -r 1 -o",
+            "replications = 3": "replications = 0",
+            "timeit -n 10 -r 3": "timeit -n 2 -r 1",
             "plt.savefig":"#plt.savefig",
             "n_plot=50": "n_plot=8",
             "'IIDStdUniform', 'Sobol', 'Lattice', 'Halton'": "'Sobol'",
             "S0, mu, sigma, T, n_samples = 100.0, 0.05, 0.20, 1.0, 2**12": "S0, mu, sigma, T, n_samples = 100.0, 0.05, 0.20, 1.0, 4",
             "n=32": "n=4",
         }
-
         self.run_notebook(notebook_path, replacements)
         
         
