### 2026-01-03 34e0ac3
Add skip and stop patterns
