+++ test/booktests/tb_portfolio_allocation_demo.py
@@ -1,6 +1,7 @@
-import unittest
+import unittest, pytest
 from __init__ import BaseNotebookTest
 
+@pytest.mark.slow  # Uses yfinance which requires network access and can be unreliable in CI
 class NotebookTests(BaseNotebookTest):
 
     def test_portfolio_allocation_demo_notebook(self):
