+++ demos/portfolio/portfolio_allocation_to_do.md
@@ -1,20 +1,35 @@
 # QMCPy 2.2 Portfolio allocation to-do
 
-0. What are the differences between the two notebooks in `demos/portfolio`? Keep `portfolio_allocation_demo` and merge content from `archive.ipynb` into it if necessary. Afterwards, delete `archive.ipynb`. --- DONE
-1. Add Sobol and Halton samplers for weights generation in the demo --- DONE
-2. Add more text in Markdown cells before code cells to the demo --- TODO
-3. . Create a single function to generate weights for all sampler types --- DONE. See `gen_weights*()`.
-4. Update visualizations in the demo with all QMC samplers and risk categories --- DONE
-5. Implement a more optimal simplex transformation for low-discrepancy sequences --- TODO
-6. Utilize QMCPy's `replications` parameter for averaged results --- TODO
-7. Make a function for collecting data (?) --- DONE (`evaluate_sampler_sharpe`)
-8. Update start_date and end_date --- DONE
-9.  Save price datasets to `data` folder --- DONE
-10. Save figures as .png in `images` --- DONE
-11. Use same colors for same samplers across all visualizations --- DONE
-12. Use white background for all visualizations --- DONE
-13. Stopping checking in .png or binary files into repository --- DONE
-14. Clean up demo --- TODO
-15. Add references to the demo --- TODO
-16. Update poster in Overleaf --- TODO
-17. Implement out-of-sample backtesting 
+## Demo
+
+1. What are the differences between the two notebooks in `demos/portfolio`? Keep `portfolio_allocation_demo` and merge content from `archive.ipynb` into it if necessary. Afterwards, delete `archive.ipynb`. --DONE
+2. Add Sobol and Halton samplers for weights generation in the demo --DONE
+3. Create a single function to generate weights for all sampler types --DONE. See `gen_weights*()`.
+4. Update visualizations in the demo with all QMC samplers and risk categories --DONE
+5. Make a function for collecting data (?) --DONE (`evaluate_sampler_sharpe`)
+6. Update start_date and end_date --DONE
+7. Save price datasets to `data` folder --DONE
+8. Save figures as .png in `images` --DONE
+9. Use same colors for same samplers across all visualizations --DONE
+10. Use white background for all visualizations --DONE
+11. Implement a more optimal simplex transformation for low-discrepancy sequences --TODO
+12. Clean up demo --TODO
+
+## Documentation
+
+1. Stop checking in .png or binary files into repository --DONE
+2. Create `mkdocs.yml` documentation for the demo and make it like a blog --TODO 
+3. Add references to the demo--TODO
+4. Add more text in Markdown cells before code cells to the demo --TODO
+
+## Tests
+
+1. Utilize QMCPy's `replications` parameter for averaged results --TODO
+2. Implement out-of-sample backtesting --TODO
+3. Make tb_portfolio_allocation_demo.py work in CI tests on Windows ---TODO, Brandon  
+
+
+
+## Poster
+
+1. Update poster in Overleaf --TODO
