### 2026-01-03 1586ba9
Update gbm_todo.md
