+++ demos/GBM/gbm_todo.md
@@ -1,12 +1,30 @@
-# QMCPy 2.2 GBM To-do:
+# QMCPy 2.2 GBM To-do
 
-1. Add 'replications' parameter to qmcpy_util.py --- DONE
-2. Delete qmcpy_util_replications.py --- DONE
-3. Combine averaged MAE computation functions to reduce repeated code --- DONE
-4. Look into Sobol randomization in QuantLib --- DONE
-5. Add replications to the MAE Comparison bar chart (figure_6) --- DONE
-6. Add more text to the demo --- TODO
-7. Add type hints to key functions --- DONE
-8. Update online blog in Overleaf and `qmcpy.org` --- TODO
-9. Prepare for PyData Chicago talk (title, abstract, speaker bio, and slides) --- TODO
-10. Fix tb_gbm_demo.py --- TODO
+## Demo
+
+1. Add `replications` parameter to `qmcpy_util.py` --- DONE
+2. Combine averaged MAE computation functions to reduce repeated code --- DONE
+3. Add replications to the MAE Comparison bar chart (figure_6) --- DONE
+4. Add more text to the demo --- TODO
+
+## Code 
+
+1. Delete `qmcpy_util_replications.py` --- DONE
+2. Add type hints to key functions --- DONE
+
+## Documentation
+
+1. Update online blog in Overleaf and `qmcpy.org` --- TODO
+
+## Tests
+
+1. Fix `tb_gbm_demo.py` and make it fast --- DONE
+
+## Research
+
+1. Look into Sobol randomization in QuantLib --- DONE
+
+## Talk
+
+1. Prepare for PyData Chicago talk (title, abstract, speaker bio, and slides) --- TODO
+
