+++ .github/workflows/alltests.yml
@@ -6,17 +6,20 @@ jobs:
   # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
   #============================================================
   tests:
-    name: All Tests on ${{ matrix.os }}
+    name: All Tests on ${{ matrix.os }} (Python ${{ matrix.python-version }})
     runs-on: ${{ matrix.os }}
     strategy:
         matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
-            os: ["windows-latest", "macos-latest", "ubuntu-latest"] 
+            os: ["windows-latest", "macos-latest", "ubuntu-latest"]
+            python-version: ['3.10', '3.11', '3.12']
     steps:
       - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
       - uses: conda-incubator/setup-miniconda@v3
         with:
           miniconda-version: "latest"
-          python-version: "3.10"
           auto-activate-base: true
           conda-remove-defaults: true
           use-only-tar-bz2: true
