+++ .github/workflows/alltests.yml
@@ -10,7 +10,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
         matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
+<<<<<<< Updated upstream
             os: ["windows-latest", "macos-latest", "ubuntu-latest"] 
+=======
+            os: ["windows-latest", "macos-latest", "ubuntu-latest"]
+            python-version: ['3.5', '3.6', '3.7', '3.8', '3.9', '3.10', '3.11', '3.12', '3,13', '3.14']
+>>>>>>> Stashed changes
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
