+++ .github/workflows/alltests.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
         matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
             os: ["windows-latest", "macos-latest", "ubuntu-latest"]
-            python-version: ['3.10', '3.11', '3.12']
+            python-version: ['3.10']
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
