+++ .github/workflows/unittests.yml
@@ -1,13 +1,9 @@
-name: All Tests
-on:
-  push:
-    branches:
-      - develop
-      - master
+name: Unit Tests
+on: [push]
 
 jobs:
   #============================================================
-  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
+  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu with Python 3.5 to 3.14
   #============================================================
   tests:
     name: All Tests on ${{ matrix.os }}
@@ -15,7 +11,7 @@ jobs:
     strategy:
         matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
             os: ["windows-latest", "macos-latest", "ubuntu-latest"] 
-            python-version: ['3.10']
+            python-version: ['3.5', '3.6', '3.7', '3.8', '3.9', '3.10', '3.11', '3.12', '3,13', '3.14']
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -249,41 +245,3 @@ jobs:
       - name: Run unittests (parallel)
         run: make unittests
       
-      # -----------------------------------------------------------
-      # Free disk + conda cache before booktests (OS-specific)
-      # -----------------------------------------------------------
-      - name: Free space (Linux)
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt-get clean
-          sudo rm -rf /usr/share/dotnet /usr/local/lib/android /opt/ghc
-          conda clean --all --yes
-          pip cache purge || true
-          df -h
-      - name: Free space (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          conda clean --all --yes
-          pip cache purge || true
-          df -h
-      - name: Free space (Windows)
-        if: runner.os == 'Windows'
-        run: |
-          conda clean --all --yes || true
-          pip cache purge || true
-          Get-Volume | Select-Object DriveLetter, Size, SizeRemaining | Format-Table
-      # -----------------------------------------------------------
-      # Run unittests for Jupyter notebooks (OS-specific)
-      # ----------------------------------------------------------- 
-      - name: Run booktests (parallel) on Linux
-        if: runner.os != 'Windows'
-        shell: bash -l {0}
-        run: |
-          make booktests_parallel_no_docker    # Parsl not supported on Windows
-      - name: Run booktests (parallel) on Windows
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          make booktests_parallel_pytest PYTEST_XDIST="-n 4" PYTEST="-m 'not slow'"
-  
-# For success reports, see https://github.com/QMCSoftware/QMCSoftware/actions?query=is%3Asuccess+workflow%3AAll+Tests
