+++ .github/workflows/unittests.yml
@@ -172,55 +172,6 @@ jobs:
         shell: pwsh
         run: |
           pip install --find-links ./.wheels -e '.[test,test_torch,test_gpytorch,test_botorch]'
-      # -----------------------------------------------------------
-      # Install minimal LaTeX required by Jupyter notebooks (OS-specific)
-      # -----------------------------------------------------------
-      - name: Install minimal LaTeX (Linux)
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt update
-          sudo apt install -y texlive-latex-base texlive-latex-extra texlive-latex-recommended latexmk dvipng cm-super
-      - name: Install minimal LaTeX (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          brew install --cask basictex
-          eval "$(/usr/libexec/path_helper)"
-          sudo tlmgr update --self
-          sudo tlmgr install latexmk dvipng collection-fontsrecommended type1cm
-      - name: Cache MiKTeX (Windows)
-        if: runner.os == 'Windows'
-        uses: actions/cache@v4
-        with:
-          path: C:\Program Files\MiKTeX
-          key: ${{ runner.os }}-miktex-${{ hashFiles('.github/workflows/alltests.yml') }}
-          restore-keys: |
-            ${{ runner.os }}-miktex-
-          
-      - name: Install minimal LaTeX (Windows)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          choco install miktex -y
-          choco install strawberryperl -y
-          # Put MiKTeX on PATH for all subsequent steps
-          echo "C:\Program Files\MiKTeX\miktex\bin\x64" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          # Refresh PATH in current session
-          $env:PATH = "C:\Program Files\MiKTeX\miktex\bin\x64;$env:PATH"
-          # Enable auto-install of missing packages (MiKTeX "just enough TeX" approach)
-          initexmf --admin --set-config-value [MPM]AutoInstall=1
-          # Refresh package database + upgrade base
-          mpm --admin --update-db
-          mpm --admin --upgrade
-          # Install a few common "baseline" MiKTeX packages explicitly
-          mpm --admin --install=latexmk
-          mpm --admin --install=dvipng
-          mpm --admin --install=cm-super
-          mpm --admin --install=lmodern
-          mpm --admin --install=type1cm
-          mpm --admin --install=tex-gyre
-          latex --version
-          latexmk -v
-          dvipng --version
       
       # -----------------------------------------------------------
       # Run doctests (OS-specific)
