+++ mkdocs.yml
@@ -62,6 +62,7 @@ nav:
       - Sensitivity Analysis for the Iris Dataset: demos/iris.ipynb 
       - Ray Tracing: demos/ray_tracing.ipynb 
       - Importance Sampling with True Measures:
+        - Some True Measures: demos/some_true_measures.ipynb
         - SciPyWrapper dependence and Custom distributions: demos/scipywrapper_dependence_custom/scipywrapper_demo.ipynb
   - For Developers:
     - Release Manual: RELEASE.md
