### 2026-01-05 edb7fb3
try deleting some unnecessary files
