+++ paper/FJHown25.bib
@@ -1,2127 +1,0 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% https://bibdesk.sourceforge.io/
-
-%% Created for Fred Hickernell at 2025-10-27 16:24:03 -0500 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-@preamble{"\def\Ignore#1{}"}
-@preamble{"\def\notesupp#1{}"}
-@preamble{"\providecommand{\HickernellFJ}{Hickernell\xspace}"}
-
-
-@string{acmtomacs = {ACM Trans.\ Model.\ Comput.\ Simul.}}
-
-@string{acmtoms = {ACM Trans.\ Math.\ Software}}
-
-@string{annstat = {Ann.\ Stat.}}
-
-@string{art = {\Ignore{a}Refereed Journal Articles}}
-
-@string{book = {\Ignore{b}Book}}
-
-@string{chap = {\Ignore{c}Refereed and/or Invited Book Chapters and Conference Papers}}
-
-@string{cma = {Comput.\ Math.\ Appl.}}
-
-@string{ed = {\Ignore{d}Edited Volumes}}
-
-@string{ijnam = {Int.\ J. Numer.\ Anal.\ Model.}}
-
-@string{jasa = {J. Amer.\ Statist.\ Assoc.}}
-
-@string{jcam = {J. Comput.\ Appl.\ Math.}}
-
-@string{jcgs = {J. Comput.\ Graph.\ Statist.}}
-
-@string{jcomp = {J. Complexity}}
-
-@string{jcp = {J. Comput.\ Phys.}}
-
-@string{jrssb = {J. Roy.\ Statist.\ Soc.\ B}}
-
-@string{jspi = {J. Statist.\ Plann.\ Inference}}
-
-@string{mc = {Math.\ Comp.}}
-
-@string{mcm = {Math.\ Comput.\ Modelling}}
-
-@string{mm = {Monatsh.\ Math.}}
-
-@string{nummath = {Numer.\ Math.}}
-
-@string{oth = {\Ignore{o}Other Publications}}
-
-@string{pnas = {Proc.\ Natl.\ Acad.\ Sci.}}
-
-@string{prep = {\Ignore{z}Manuscripts in Preparation}}
-
-@string{sinum = {SIAM J. Numer.\ Anal.}}
-
-@string{sirev = {SIAM Rev.}}
-
-@string{sisc = {SIAM J. Sci.\ Comput.}}
-
-@string{soft = {\Ignore{f}Computer Software}}
-
-@string{ss = {Statist.\ Sinica}}
-
-@string{stam = {Stud.\ Appl.\ Math}}
-
-@string{statprob = {Statist.\ Probab.\ Lett.}}
-
-@string{sub = {\Ignore{u}Manuscripts Under Review}}
-
-
-@inproceedings{SorHicRob25a,
-	author = {A. G. Sorokin and F. J. \HickernellFJ and P. Robbe},
-	booktitle = {Proc. Mach. Learn. Res.},
-	date-added = {2025-05-16 11:35:20 -0500},
-	date-modified = {2025-05-16 11:48:53 -0500},
-	editor = {M. Kanagawa and J. Cockayne and A. Gessner and P. Hennig},
-	pubtype = chap,
-	title = {Fast {G}aussian process regression for high dimensional functions with derivative information},
-	year = {2025}}
-
-@misc{MatHic26a,
-	author = {L. Matiukha and Y. Ding and  F. J. \HickernellFJ},
-	date-added = {2025-05-15 10:14:00 -0500},
-	date-modified = {2025-10-27 16:23:56 -0500},
-	note = {in preparation},
[…]
