+++ paper/README.md
@@ -31,7 +31,7 @@
 - [**`FJHown25.bib`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/paper/FJHown25.bib): Bibliography entries of products by Hickernell
 - [**`choi2025.bib`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/paper/choi2025.bib): Bibliography entries by Sou-Cheng T. Choi and co-authors (2025)
 - [**`other.bib`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/paper/other.bib): Miscellaneous bibliography entries
-- [**`refs_all.bib`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/paper/refs_all.bib): Combined bibliography (auto-generated by `make`)
+- [**`main.bib`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/paper/main.bib): Combined bibliography (auto-generated by `make`)
 
 ### Other Files
 - [**`joss2025.ipynb`**](https://github.com/QMCSoftware/QMCSoftware/blob/joss/demos/talk_paper_demos/JOSS2025/joss2025.ipynb): Jupyter notebook with QMCPy code for reproducing figures in the paper
@@ -70,6 +70,6 @@ Other Commands:
 
 - `make clean`: Removes auxiliary files.
 - `make cleanall`: Deletes all outputs, including PDFs.
-- `make refs_all.bib`: Combines bibliography files.
+- `make main.bib`: Combines bibliography files.
 - `make wordcount`: Counts words in `paper.md` (excludes metadata/references).
 - `make install-mac`:  Install required tools on macOS using Homebrew.
