+++ paper/other.bib
@@ -1,345 +1,0 @@
-@incollection{NEURIPS2019_9015,
-title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
-author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-booktitle = {Advances in Neural Information Processing Systems 32},
-pages = {8024--8035},
-year = {2019},
-publisher = {Curran Associates, Inc.},
-url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
-} 
-
-@Article{harris2020array,
- title         = {Array programming with {NumPy}},
- author        = {Harris, Charles R. and Millman, K. Jarrod and St{\'{e}}fan J. van der Walt and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del R{\'{i}}o, Jaime Fern{\'{a}}ndez and Wiebe, Mark and Peterson, Pearu and G{\'{e}}rard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
- year          = {2020},
- month         = sep,
- journal       = {Nature},
- volume        = {585},
- number        = {7825},
- pages         = {357--362},
- doi           = {10.1038/s41586-020-2649-2},
- publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
-}
-
-@article{giles2015multilevel,
-  title={Multilevel {M}onte {C}arlo methods},
-  author={Giles, Michael B},
-  journal={Acta Numerica},
-  volume={24},
-  pages={259--328},
-  year={2015},
-  publisher={Cambridge University Press}
-}
-
-@article{giles2009multilevel,
-  title={Multilevel quasi-{M}onte {C}arlo path simulation},
-  author={Giles, Michael B and Waterhouse, Benjamin J},
-  journal={Advanced Financial Modelling, Radon Series on Computational and Applied Mathematics},
-  volume={8},
-  pages={165--181},
-  year={2009},
-  publisher={de Gruyter New York}
-}
-@misc{MPS,
-    title={Application of quasi-{M}onte {C}arlo methods to elliptic PDEs with random diffusion coefficients - a survey of analysis and implementation},
-    author={Kuo, Frances Y. and Nuyens, Dirk},
-    year={2016},
-    eprint={1606.06613},
-    archivePrefix={arXiv},
-    primaryClass={math.NA}}
-
-@article{joe2008constructing,
-  title={Constructing {S}obol sequences with better two-dimensional projections},
-  author={Joe, S. and Kuo, Frances Y.},
-  journal={SIAM Journal on Scientific Computing},
-  volume={30},
-  number={5},
-  pages={2635--2654},
-  year={2008},
[…]
