### 2026-01-05 132b1f7
Test Python 3.5 to 3.14 for 3 OS (#454)
* Revert changes

* Fix GBM

* Remove testing.py

* add colab button

* Fix Colab button

* Clean up

* Clean up more

* Add QMCPy heat map for covariance

* Add _validate_samples only in long tests. Add _weight()

* Final change

* Add unit tests for (G)BM and more comprehensive gbm_demo notebook

* Accelerated GBM

* Add unit tests and remove useless methods

* Minor changes to gbm_demo

* Minor changes to demo

* Better latex table

* Restore .gitignore

* Fix latex error

* Add code and seed to QuantLib GBM with Sobol sampler

* Add minor changes and
[…]
