+++ .github/workflows/unittests.yml
@@ -1,17 +1,40 @@
-name: All Tests
-on: [push]
+name: Unit Tests
+on:
+  push:
+    branches:
+      - develop
+      - master
 
 jobs:
-  #============================================================
-  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu
-  #============================================================
+  #======================================================================
+  # 3 Test Jobs on 3 OSes: Windows, macOS, Ubuntu with Python 3.5 to 3.14
+  #======================================================================
   tests:
-    name: All Tests on ${{ matrix.os }}
+    name: Unit Tests on ${{ matrix.os }} (Python ${{ matrix.python-version }})
     runs-on: ${{ matrix.os }}
     strategy:
-        matrix:  # https://docs.github.com/en/actions/reference/runners/github-hosted-runners
-            os: ["windows-latest", "macos-latest", "ubuntu-latest"] 
-            python-version: ['3.10']
+        matrix:
+            include:
+              - os: macos-latest
+                python-version: '3.5'
+              - os: macos-latest
+                python-version: '3.8'
+              - os: macos-latest
+                python-version: '3.11'
+              - os: macos-latest
+                python-version: '3.14'
+              - os: ubuntu-latest
+                python-version: '3.6'
+              - os: ubuntu-latest
+                python-version: '3.9'
+              - os: ubuntu-latest
+                python-version: '3.12'
+              - os: windows-latest
+                python-version: '3.7'
+              - os: windows-latest
+                python-version: '3.10'
+              - os: windows-latest
+                python-version: '3.13'
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -172,114 +195,10 @@ jobs:
         shell: pwsh
         run: |
           pip install --find-links ./.wheels -e '.[test,test_torch,test_gpytorch,test_botorch]'
-      # -----------------------------------------------------------
-      # Install minimal LaTeX required by Jupyter notebooks (OS-specific)
-      # -----------------------------------------------------------
-      - name: Install minimal LaTeX (Linux)
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt update
-          sudo apt install -y texlive-latex-base texlive-latex-extra texlive-latex-recommended latexmk dvipng cm-super
-      - name: Install minimal LaTeX (macOS)
-        if: runner.os == 'macOS'
-        run: |
-          brew install --cask basictex
-          eval "$(/usr/libexec/path_helper)"
-          sudo tlmgr update --self
-          sudo tlmgr install latexmk dvipng collection-fontsrecommended type1cm
-      - name: Cache MiKTeX (Windows)
-        if: runner.os == 'Windows'
-        uses: actions/cache@v4
-        with:
-          path: C:\Program Files\MiKTeX
-          key: ${{ runner.os }}-miktex-${{ hashFiles('.github/workflows/alltests.yml') }}
-          restore-keys: |
-            ${{ runner.os }}-miktex-
-          
-      - name: Install minimal LaTeX (Windows)
-        if: runner.os == 'Windows'
-        shell: pwsh
-        run: |
-          choco install miktex -y
-          choco install strawberryperl -y
-          # Put MiKTeX on PATH for all subsequent steps
-          echo "C:\Program Files\MiKTeX\miktex\bin\x64" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          # Refresh PATH in current session
-          $env:PATH = "C:\Program Files\MiKTeX\miktex\bin\x64;$env:PATH"
-          # Enable auto-install of missing packages (MiKTeX "just enough TeX" approach)
-          initexmf --admin --set-config-value [MPM]AutoInstall=1
-          # Refresh package database + upgrade base
-          mpm --admin --update-db
[…]
