### 2026-01-05 2734f3c
fix broken DOIs
