+++ paper/main.bib
@@ -130,13 +130,13 @@
 	booktitle = {Handbook of Uncertainty Quantification},
 	date-added = {2025-10-07 12:29:47 -0500},
 	date-modified = {2025-10-27 16:45:08 -0500},
-	doi = {10.1007/978-3-319-11259-6\_23-1},
+	doi = {10.1007/978-3-319-11259-6_23-1},
 	editor = {Ghanem, Roger and Higdon, David and Owhadi, Houman},
 	isbn = {978-3-319-11259-6},
 	pages = {1--41},
 	publisher = {Springer International Publishing},
 	title = {Sampling via Measure Transport: An Introduction},
-	url = {https://doi.org/10.1007/978-3-319-11259-6\_23-1},
+	url = {https://doi.org/10.1007/978-3-319-11259-6_23-1},
 	year = {2016},
 	bdsk-url-1 = {https://doi.org/10.1007/978-3-319-11259-6_23-1}}
 
@@ -865,7 +865,7 @@
 	author = {J. Dick and M. Feischl},
 	date-added = {2023-05-22 09:21:22 -0500},
 	date-modified = {2023-05-22 09:34:56 -0500},
-	doi = {https://doi.org/10.1016/j.jco.2021.101587},
+	doi = {10.1016/j.jco.2021.101587},
 	issn = {0885-064X},
 	journal = jcomp,
 	keywords = {Quasi-Monte Carlo, Big data, Statistical learning, Higher-order methods},
@@ -1130,7 +1130,7 @@
 	author = {J. Dick and P. Kritzer and F. Pillichshammer},
 	date-added = {2022-11-01 13:06:34 -0500},
 	date-modified = {2022-11-01 13:09:00 -0500},
-	doi = {https://doi.org/10.1007/978-3-031-09951-9},
+	doi = {10.1007/978-3-031-09951-9},
 	publisher = {Springer Cham},
 	series = {Springer Series in Computational Mathematics},
 	title = {Lattice Rules: {N}umerical Integration, Approximation, and Discrepancy},
@@ -1444,7 +1444,7 @@
 @article{paulin2022,
 	author = {Lo\"is Paulin and Nicolas Bonneel and David Coeurjolly and Jean-Claude Iehl and Alexander Keller and Victor Ostromoukhov},
 	date-modified = {2022-07-14 13:25:49 -0500},
-	doi = {https://doi.org/10.1145/3528223.3530063},
+	doi = {3528223.3530063},
 	journal = {ACM Transactions on Graphics (Proceedings of SIGGRAPH)},
 	month = {August},
 	number = {4},
@@ -1919,10 +1919,10 @@
 	crossref = {LEcTuf20a},
 	date-added = {2020-09-06 15:07:06 -0500},
 	date-modified = {2020-09-06 15:11:27 -0500},
-	doi = {10.1007/978-3-030-43465-6\_2},
+	doi = {10.1007/978-3-030-43465-6_2},
 	pages = {31--67},
 	title = {Multilevel Quasi-{M}onte {C}arlo Uncertainty Quantification for Advection-Diffusion-Reaction},
-	bdsk-url-1 = {https://doi.org/10.1007/978-3-030-43465-6%5C_2}}
+	bdsk-url-1 = {https://doi.org/10.1007/978-3-030-43465-6_2}}
 
 @article{LEc09,
 	author = {P. L'Ecuyer},
@@ -2245,7 +2245,7 @@
 	author = {I. G. Graham and F. Y. Kuo and D. Nuyens and R. Scheichl and I. H. Sloan},
[…]
