### 2026-01-05 a99ef83
Refactor SciPyWrapper demo true_measures and fix booktest path
