### 2026-01-05 bac8c7e
start clean bib
