### 2026-01-05 e9b77e5
tidy bib
