+++ paper/main.bib
@@ -73,853 +73,890 @@
 
 @string{sub = {\Ignore{u}Manuscripts Under Review}}
 
-@inproceedings{Lem04a,
-	author = {C. Lemieux},
-	booktitle = {Proc.\ 2004 Winter Simulation Conference},
-	date-added = {2007-09-26 10:50:17 -0500},
-	date-modified = {2007-11-13 18:07:57 -0600},
-	editor = {R. G. Ingalls and M. D. Rossetti and J. S. Smith and B. A. Peters},
-	pages = {1565--1573},
-	publisher = {IEEE Press},
-	title = {Randomized Quasi-{M}onte {C}arlo: a tool for improving the efficiency of simulations in finance},
-	year = {2004},
-	doi = {10.1109/wsc.2004.1371499},
-	url = {https://doi.org/10.1109/wsc.2004.1371499}}
-
-@article{wangsloan05,
-	author = {Wang, X. and Ian H. Sloan},
-	date-added = {2025-01-27 12:51:03 -0600},
-	date-modified = {2025-01-27 12:51:03 -0600},
-	journal = {SIAM J. Sci. Comput.},
-	pages = {159-183},
-	title = {Why are high-dimensional finance problems often of low effective dimension?},
-	volume = {27 (1)},
-	year = {2005},
-	doi = {10.1137/s1064827503429429},
-	url = {https://doi.org/10.1137/s1064827503429429}}
-
-@article{giles2009multilevel,
-  title={Multilevel quasi-{M}onte {C}arlo path simulation},
-  author={Giles, Michael B and Waterhouse, Benjamin J},
-  journal={Advanced Financial Modelling, Radon Series on Computational and Applied Mathematics},
-  volume={8},
-  pages={165--181},
-  year={2009},
-  publisher={de Gruyter New York},
-  doi = {10.1515/9783110213140.165},
-  url = {https://doi.org/10.1515/9783110213140.165}
+@inproceedings{HicEtal14a,
+  title               = {Guaranteed Conservative Fixed Width Confidence Intervals Via {M}onte {C}arlo Sampling},
+  author              = {F. J. Hickernell and L. Jiang and Y. Liu and Owen, A. B.},
+  pages               = {105--128},
+  doi                 = {10.1007/978-3-642-41095-6_5},
+  url                 = {https://doi.org/10.1007/978-3-642-41095-6\%5F5},
+  crossref            = {KuoEtal14a},
+  date-added          = {2012-06-25 03:41:26 +0000},
+  date-modified       = {2015-09-14 19:41:19 +0000},
+  pubtype             = chap,
[…]
