+++ paper/main.bib
@@ -1,3 +1,5 @@
+% bibtex-tidy --omit=abstract --curly --months --align=20 --sort=year,key --duplicates=key,doi,citation,abstract --merge=overwrite --sort-fields --trailing-commas --no-tidy-comments --remove-empty-fields paper/main.bib
+
 @preamble{"\def\Ignore#1{}"}
 @preamble{"\def\notesupp#1{}"}
 
@@ -68,7 +70,7 @@
   author              = {F. J. Hickernell and L. Jiang and Y. Liu and Owen, A. B.},
   pages               = {105--128},
   doi                 = {10.1007/978-3-642-41095-6_5},
-  url                 = {https://doi.org/10.1007/978-3-642-41095-6\%5F5},
+  url                 = {https://doi.org/10.1007/978-3-642-41095-6_5},
   crossref            = {KuoEtal14a},
   pubtype             = chap,
 }
@@ -78,7 +80,7 @@
   author              = {Hickernell, F. J. and {Jim\'enez Rugama}, {Ll. A.}},
   pages               = {367--383},
   doi                 = {10.1007/978-3-319-33507-0_18},
-  url                 = {https://doi.org/10.1007/978-3-319-33507-0\%5F18},
+  url                 = {https://doi.org/10.1007/978-3-319-33507-0_18},
   crossref            = {CooNuy16a},
   pubtype             = chap,
 }
@@ -89,7 +91,7 @@
   booktitle           = {Advances in Modeling and Simulation: {F}estschrift in Honour of {P}ierre {L'E}cuyer},
   pages               = {301--318},
   doi                 = {10.1007/978-3-031-10193-9_15},
-  url                 = {https://doi.org/10.1007/978-3-031-10193-9\%5F15},
+  url                 = {https://doi.org/10.1007/978-3-031-10193-9_15},
   crossref            = {LEc2022},
   pubtype             = chap,
 }
@@ -99,7 +101,7 @@
   author              = {{{\relax Ll}. A.} {Jim\'enez Rugama} and F. J. Hickernell},
   pages               = {407--422},
   doi                 = {10.1007/978-3-319-33507-0_20},
-  url                 = {https://doi.org/10.1007/978-3-319-33507-0\%5F20},
+  url                 = {https://doi.org/10.1007/978-3-319-33507-0_20},
   crossref            = {CooNuy16a},
   pubtype             = chap,
 }
@@ -110,7 +112,7 @@
   booktitle           = {{M}onte {C}arlo and Quasi-{M}onte {C}arlo Methods in Scientific Computing},
   pages               = {299--317},
   doi                 = {10.1007/978-1-4612-2552-2_19},
-  url                 = {https://doi.org/10.1007/978-1-4612-2552-2\%5F19},
+  url                 = {https://doi.org/10.1007/978-1-4612-2552-2_19},
   crossref            = {NieShi95},
 }
 
@@ -125,7 +127,7 @@
   author              = {Y. Wang and F. J. Hickernell},
   pages               = {158--167},
   doi                 = {10.1007/978-3-642-56046-0_10},
-  url                 = {https://doi.org/10.1007/978-3-642-56046-0\%5F10},
+  url                 = {https://doi.org/10.1007/978-3-642-56046-0_10},
   crossref            = {NieFanHic01},
   mr                  = {1 958 852},
   pubtype             = chap,
@@ -559,7 +561,7 @@
   title               = {{MIT} uncertainty quantification (MUQ) library},
   author              = {Parno, Matthew and Davis, Andrew and Seelinger, Linus and Marzouk, Youssef},
   year                = {2014},
[…]
