+++ paper/main.bib
@@ -141,10 +141,7 @@
   number              = {LA-1342},
   doi                 = {10.2172/4405295},
   url                 = {https://doi.org/10.2172/4405295},
-  date-added          = {2014-02-18 14:46:25 +0000},
-  date-modified       = {2014-02-18 14:49:23 +0000},
   institution         = {Los Alamos Scientific Laboratory},
-  bdsk-file-1         = {YnBsaXN0MDDSAQIDBFxyZWxhdGl2ZVBhdGhZYWxpYXNEYXRhXxBrLi4vLi4vLi4vLi4vLi4vQXJ0aWNsZXNfVGFsa3MvUHJlcHJpbnRzOlJlcHJpbnRzL04tUyByZXByaW50cy9SaWNodG15ZXIvUU1DTG9zQWxhbW9zMTk1MVJlcG9ydFJpY2h0bXllci5wZGZPEQIMAAAAAAIMAAIAAAxNYWNpbnRvc2ggSEQAAAAAAAAAAAAAAAAAAAAAAAAAQkQAAf////8fUU1DTG9zQWxhbW9zMTk1MVJlI0ZGRkZGRkZGLnBkZgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wAAAAAAAAAAAAAAAAAFAAUAAAogY3UAAAAAAAAAAAAAAAAACVJpY2h0bXllcgAAAgB0LzpVc2VyczpmcmVkamhpY2tlcm5lbGw6QXJ0aWNsZXNfVGFsa3M6UHJlcHJpbnRzL1JlcHJpbnRzOk4tUyByZXByaW50czpSaWNodG15ZXI6UU1DTG9zQWxhbW9zMTk1MVJlcG9ydFJpY2h0bXllci5wZGYADgBIACMAUQBNAEMATABvAHMAQQBsAGEAbQBvAHMAMQA5ADUAMQBSAGUAcABvAHIAdABSAGkAYwBoAHQAbQB5AGUAcgAuAHAAZABmAA8AGgAMAE0AYQBjAGkAbgB0AG8AcwBoACAASABEABIAclVzZXJzL2ZyZWRqaGlja2VybmVsbC9BcnRpY2xlc19UYWxrcy9QcmVwcmludHM6UmVwcmludHMvTi1TIHJlcHJpbnRzL1JpY2h0bXllci9RTUNMb3NBbGFtb3MxOTUxUmVwb3J0UmljaHRteWVyLnBkZgATAAEvAAAVAAIAFv//AAAACAANABoAJACSAAAAAAAAAgEAAAAAAAAABQAAAAAAAAAAAAAAAAAAAqI=},
 }
 
 @article{Hal60,
@@ -156,8 +153,6 @@
   pages               = {84--90},
   doi                 = {10.1007/bf01386213},
   url                 = {https://doi.org/10.1007/bf01386213},
-  date-added          = {2007-09-26 10:50:17 -0500},
-  date-modified       = {2007-09-26 10:50:17 -0500},
 }
 
 @article{coveyou1967fourier,
@@ -169,6 +164,8 @@
   volume              = {14},
   number              = {1},
   pages               = {100--119},
+  doi                 = {10.1145/321371.321379},
+  url                 = {https://doi.org/10.1145/321371.321379},
 }
 
 @article{Sob67,
@@ -178,8 +175,8 @@
   journal             = {U.S.S.R. Comput. Math. and Math. Phys.},
   volume              = {7},
   pages               = {86--112},
-  date-added          = {2007-09-26 10:50:17 -0500},
-  date-modified       = {2007-09-26 10:50:17 -0500},
[…]
