+++ paper/main.bib
@@ -116,7 +116,7 @@
   crossref            = {NieShi95},
 }
 
-@misc{QMCPyDocs2,
+@misc{QMCPyDocs,
   title               = {{QMCPy} Documentation},
   author              = {Choi, S.-C. T. and Hickernell, Fred J. and Rathinavel, Jagadeeswaran and McCourt, Michael J. and Sorokin, Aleksei G.},
   url                 = {https://qmcsoftware.github.io/QMCSoftware},
@@ -464,7 +464,7 @@
 }
 
 @webpage{SobolDirection,
-  title               = {Sobol Sequence Generator},
+  title               = {Sobol' Sequence Generator},
   author              = {S. Joe and F. Y. Kuo},
   year                = {2010},
   url                 = {https://web.maths.unsw.edu.au/~fkuo/sobol/},
@@ -646,6 +646,7 @@
   title               = {Fast Automatic {B}ayesian Cubature Using Matching Kernels and Designs},
   author              = {J. Rathinavel},
   year                = {2019},
+  url                 = {http://hdl.handle.net/10560/islandora:1009768},
   school              = {Illinois Institute of Technology},
 }
 
@@ -758,6 +759,7 @@
   author              = {Parno, Matthew and Davis, Andrew and Seelinger, Linus},
   year                = {2021},
   journal             = {Journal of Open Source Software},
+  publisher           = {The Open Journal},
   volume              = {6},
   number              = {68},
   pages               = {3076},
@@ -842,6 +844,19 @@
   doi                 = {10.1109/wsc60868.2023.10408613},
   url                 = {https://doi.org/10.1109/wsc60868.2023.10408613},
   organization        = {IEEE},
+}
+
+@article{Roy2023,
+  title               = {Quasi-{M}onte {C}arlo methods in {P}ython},
+  author              = {Roy, Pamphile T. and Owen, Art B. and Balandat, Maximilian and Haberland, Matt},
+  year                = {2023},
+  journal             = {Journal of Open Source Software},
+  publisher           = {The Open Journal},
+  volume              = {8},
+  number              = {84},
+  pages               = {5309},
+  doi                 = {10.21105/joss.05309},
+  url                 = {https://doi.org/10.21105/joss.05309},
 }
 
 @inproceedings{ChoEtal24a2,
