### 2026-01-05 #9705 @alegresor
Thanks for the review @tupui!

We pushed some updates to the paper which add points of comparison with `spicy.stats.qmc` and `torch.quasirandom`.  Please let us know if you would like to see a more detailed breakdown; we are happy to expand this section but are conscious of already being over the word limit.

We tried to fix the references and add DOIs when they are available. Is there a way for us to re-trigger the editorialbot's reference check?

Our documentation contains links to [QMCPy contributor guidelines](https://qmcsoftware.github.io/QMCSoftware/CONTRIBUTING/) and an actively monitored [QMCPy issues page](https://github.com/QMCSoftware/QMCSoftware/issues). Are there other items you would like to see for the "Community Guidelines" item? We are happy to update.
