### 2026-01-05 #9705 @tupui
Thanks for the updates @alegresor all good for me and checked all the boxes now.

The article is a bit long for JOSS in general (e.g. the MC/QMC part is not needed, same with all the maths basically), but it's there and correct so I personally won't nitpick. One of the goal, essence, of JOSS is to have a fast process for software publication as the hard part is in the software.
