### 2026-01-06 QMCSoftware/QMCSoftware#457 @matt-graham
Non-standard Apache-2.0 LICENSE text
Raising as part of JOSS review openjournals/joss-reviews#9705

The text of the `LICENSE` file

https://github.com/QMCSoftware/QMCSoftware/blob/eb002b049c8ca8602d594effd5e00b3ecf4f978f/LICENSE#L1-L13

does not appear to be the standard Apache-2.0 license text but instead [the notice intended for attaching to individual source files](https://www.apache.org/licenses/LICENSE-2.0#apply). The notice links to the full license text but does not include the text itself. I believe typically the `LICENSE` file included in the root of a repository should be [a full plain text version](https://www.apache.org/licenses/LICENSE-2.0.txt) of the relevant license.
