### 2026-01-06 QMCSoftware/QMCSoftware#458 @matt-graham
JOSS paper comments
Raising as part of JOSS review openjournals/joss-reviews#9705

My main comment would be that the summary is currently quite jargon heavy - it should be a 'description of the high-level functionality and purpose of the software for a diverse, non-specialist audience'. I would say terms like 'low-discrepancy sequences', 'option pricing' and 'kernel methods' are not necessarily likely to be widely understood by non-specialist readers. Additionally I think a one or two sentence high-level description of what (quasi-)Monte Carlo methods entail would be helpful for increasing accessibility here. For example something like

> Monte Carlo methods estimate integrals by computing averages over independent random points sampled from the integration domain. Quasi-Monte Carlo methods replace the independent random samples with low-discrepancy sequences which more uniformly cover the space being integrated over, helping to reduce variance in the estimates computed.

Can I also check if the contributor GitHub user accounts `schoi32` and `sou-cheng-choi` both correspond to author Sou-Cheng T. Choi (checking this as both accounts appear to have made significant contributions and there is a specific point in reviewer checklist to check _Does the full list of paper authors seem appropriate and complete?_).

Minor typos / comments

- [x] Lines 9-10:  _object oriented (OO)_ - acronym short form is not used anywhere so should be removed.
- [x] Line 25:  _independent identically distributed_ is more usually written _independent and identically distributed_.
- [x] In Figure 1 caption it wasn't immediately clear to me that _the digital sequence_ referred to panel labelled _LD Digital Net_ - it would help if label matched reference in text.
- [x] Line 36-39: In bulleted list, as list items are being written as components of a single sentence then they should be capitalized accordingly (that is first letter in each list item should be lowercase unless it is a proper noun or otherwise capitalized) as done in list in lines 45 to 48.
- [x] In central panel of Figure 2, vertical axis needs a unit - presumably time is in seconds here - unless time is being measured in terms of computational operations performed or similarly in which case this should ideally be made more explicit.
- [x] Lines 113-114: _The right panel presents error distributions via violin plots for a single error tolerance_ - it would be nice to have the specific error tolerance used specified here. Is this is what is represented by dotted vertical line on axis? If so it would be useful to explicitly state this.
- [x] Line 118: _via the Jupyer Notebook_ reads a little strangely to me as it sounds as the use of the definite article makes it sounds like there is a singular Jupyter notebook and it wasn't clear to me the reference following this did resolve to a specific Jupyter notebook - I would say just using a plain hyperlink here and possibly rephrasing to _a Jupyter notebook_ might be a bit more intuitive but this is just a personal preference.
- [x] Lines 145-147: _international conference on machine learning_ should be capitalized in reference.
- [x] Lines 295-300: _in neural information processing systems_ should be capitalized in reference.
