### 2026-01-06 QMCSoftware/QMCSoftware#458 @alegresor
Let's use https://github.com/openjournals/joss-reviews/issues/9705 for the review if you don't mind. This will help us keep things self-contained and there I can use the JOSS editorialbot. I'll respond there.
