### 2026-01-07 #9705 @alegresor
Thanks @lockwo!

Our QMCPy DOI for all releases is https://doi.org/10.5281/zenodo.17889172.

We will target a QMCPy-JOSS release for next Friday, Jan 16, and follow up after that.
