### 2026-01-12 #9705 @lockwo
Thanks for the quick update, while the 1,000 words isn't a strict cutoff, being 30% over isn't ideal and it would be great if we could cut it down some more (I know it can be difficult to cut, but the key sections are the statement of need which looks good, related software, and papers enabled by the software, things on the API or fine grained details about the software are often very valuable, but can go in documentation pages rather than the paper). Making the "Components" section more focused should help (e.g. potentially narrowing down 1., maybe lines 80-87, either integrating Fig 2 more into the SC explanation or shrinking the text 112-118 into a caption might be ideas).
