### 2026-01-16 #9705 @lockwo
@alegresor one thing that was suggested was to remove the paper.md in the docs/. Apparently sometimes editorialbot can pick up the wrong paper so they recommend having paper.md as a unique name in the repo (you can move the files to a different branch, or add them after acceptance if they need to be in the docs/ for a particular reason).
