### 2026-01-16 #9705 @lockwo
@alegresor Great, glad it works. Thanks for the DOI, the author list matches which is good, it looks like the archival release title is off tho, the title (and authors, which is done) for the archive should match the paper (ie the DOI title should be "QMCPy: A Python Framework for (Quasi-)Monte Carlo Algorithms", I currently see it as "QMCSoftware/QMCSoftware: QMCPy v2.2", I believe you should be able to just edit the metadata without releasing a new DOI).
