### 2026-01-17 #9705 @alegresor
Just update the DOI zenodo name to match the JOSS title. Are we missing anything else?
