### 2026-01-17 #9705 @lockwo
I believe the links are reference based, I see your updates on the final proof link (please check to verify as well, but I looked at the git diff and saw it when I just downloaded after generating the pdf). If you run the generate pdf, it applies to all links (for example, all previous generate pdfs are overwritten and go to this new pdf)
