### 2026-01-17 #9705 @alegresor
> 👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.09705/joss.09705/10.21105.joss.09705.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#7668](https://github.com/openjournals/joss-papers/pull/7668), then the AEiC can move forward with accepting the submission by compiling again with the command `@editorialbot accept`

@lockwo when click download it looks fine, but the [openjournals/joss-papers#7668](https://github.com/openjournals/joss-papers/pull/7668) PR has not updated the files. Otherwise we are satisfied on our end!
