### 2026-01-17 #9705 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
