### 2026-01-17 #9705 @lockwo
Great, at this point the associate editor in chief for DSAIS can step in, and editorial bot automatically pings them to let them know (for transparency, here is an example of what that could look like https://github.com/openjournals/joss-reviews/issues/9058#issuecomment-3675507770).
