### 2026-01-19 #9705 @alegresor
Thank you @crvernon @lockwo @tupui @matt-graham for creating an enjoyable and efficient JOSS publication process!
